D-173 Stage 2: verbatim decomposition of 3 large files into manifest packages
Kill the remaining large-file complexity (SOLID/SRP) with byte-identical
extraction behind source manifests, following the D-164/D-168 pattern: each
extracted module is one line in a source manifest; gates/tests read the
composed source so structural assertions survive the move.

doctrine.py (1600) → package app/services/skills/doctrine/
  exercise_core / invocation / pedagogy / probability / platform_layer /
  manifest_registry, with _sources.py (DOCTRINE_SOURCE_FILES). __init__
  re-exports all 68 public names verbatim; import path unchanged. All 8
  get_*_summary helpers (never in __all__) stay explicit-importable.

customer_chat.py (1412 → 834) → support pkg customer_chat_support/
  transport (locked_send + keepalive + ws-state), pedagogy (BKT/directive/
  firewall), frames (client-context + terminal frames + card persistence),
  _sources.py (CUSTOMER_CHAT_SOURCE_FILES). Handlers + chat_stream_ws stay
  (D-168 residency); 22 symbols re-exported via noqa F401 so late-binding
  monkeypatch keeps working.

graph/main.py (1300 → 214) → dspy_compat + state + nodes + graph_sources.py
  (GRAPH_SOURCE_FILES). AgentState identity preserved; create_unified_graph
  re-exports node classes.

Dead code removed: app/static/js/superhuman-framework.js (zero refs).

Lockstep: check_skills_doctrine reads composed customer_chat source;
check_legacy_invariants repoints LatexStreamNormalizer/greeting-fastpath to
graph/nodes.py; ci_guardrails db_import waiver extended to the verbatim
pedagogy support module; runtime lock regenerated for new importer modules.
Tests aligned to composed-source reads (D-105: shrink-only, no disable).
