+++ .runtime/truth_table.lock.json
@@ -77,12 +77,12 @@
       "importers_sample": [
         "app/api/routers/admin.py",
         "app/api/routers/customer_chat.py",
+        "app/api/routers/customer_chat_support/pedagogy.py",
         "app/services/chat/orchestrator.py",
         "app/services/skills/bac_exercise_skill.py",
         "scripts/gitpod_live_debug.py",
         "scripts/gitpod_ui_test.py",
-        "scripts/omni_live_test.py",
-        "scripts/test_auto_ui_trigger.py"
+        "scripts/omni_live_test.py"
       ],
       "name": "Orchestrator HTTP client + fallback chain",
       "notes": "Sole entrypoint to chat from customer_chat.py and admin.py.",
@@ -201,9 +201,10 @@
         "app/services/boundaries/customer_chat_boundary_service.py"
       ],
       "id": "boundary_customer",
-      "importer_count": 3,
+      "importer_count": 4,
       "importers_sample": [
         "app/api/routers/customer_chat.py",
+        "app/api/routers/customer_chat_support/pedagogy.py",
         "app/services/boundaries/__init__.py",
         "app/services/chat/dispatcher.py"
       ],
@@ -244,13 +245,13 @@
       "id": "customer_chat_router",
       "importer_count": 7,
       "importers_sample": [
+        "app/api/routers/customer_chat_support/frames.py",
         "microservices/orchestrator_service/src/api/context_utils.py",
         "tests/api/test_chat_event_protocol_error_contract_integration.py",
         "tests/api/test_chat_event_protocol_flag_integration.py",
         "tests/api/test_final_router_gaps.py",
         "tests/services/test_d142_phase2_dialogue_manager.py",
-        "tests/services/test_iss108_explanation_catastrophe.py",
-        "tests/unit/test_context_fragmentation.py"
+        "tests/services/test_iss098_keepalive.py"
       ],
       "name": "Customer chat WS router",
       "notes": "Live WS endpoint /api/chat/ws.",
@@ -322,6 +323,6 @@
       "on_live_chain": true
     }
   ],
-  "generated_at_utc": "2026-07-21T00:11:52.358004+00:00",
+  "generated_at_utc": "2026-07-21T01:11:12.956649+00:00",
   "schema_version": 1
 }
