+++ app/api/routers/customer_chat.py
@@ -16,7 +16,6 @@ import asyncio
 import contextlib
 import uuid
 from datetime import UTC, datetime
-from typing import NamedTuple
 
 from fastapi import APIRouter, Depends, HTTPException, WebSocket, WebSocketDisconnect
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -34,7 +33,6 @@ from app.core.domain.chat import MessageRole
 from app.core.domain.user import User
 from app.deps.auth import CurrentUser, require_permissions
 from app.infrastructure.clients.orchestrator_client import orchestrator_client
-from app.services.analytics.bkt_persistence import BKTAnalyticsService
 from app.services.analytics.tutor_state_service import TutorStateService
 from app.services.auth.token_decoder import decode_token_payload
 from app.services.boundaries.customer_chat_boundary_service import (
@@ -48,70 +46,37 @@ from shared.chat_protocol.event_protocol import normalize_streaming_event
 logger = get_logger(__name__)
 
 
-def _semantic_tutor_enabled() -> bool:
-    """D-142/D-158: علم تحميل+كتابة `tutor_state` — قارئ موحَّد (افتراض True).
-
-    D-158: يُفوَّض إلى `app.core.feature_flags` (مصدر واحد لكل الأعلام، افتراض واحد) بدل
-    التعريف المزدوج المتعارض الذي كان يُعطِّل سلطة الحوار في المُنسّق.
-    """
-    from app.core.feature_flags import semantic_tutor_enabled
-
-    return semantic_tutor_enabled()
-
-
-def _apply_complete_response_firewall(text: str) -> str:
-    """D-086 (V46.0): يُطبِّق OutputFirewall على الإجابة المكتملة قبل الحفظ.
-
-    يُنظِّف HTML/JSX من complete_ai_response بعد اكتمال البث.
-    Fail-open: أي فشل يُعيد النص الأصلي دون كسر المسار.
-    """
-    if not text or not text.strip():
-        return text
-    try:
-        from app.services.skills.output_firewall import apply_channel_b_firewall
-
-        cleaned, was_rejected = apply_channel_b_firewall(text, intent="educational")
-        if was_rejected:
-            logger.warning("customer_chat.complete_response_firewall_rejected chars=%d", len(text))
-            return text  # fail-open
-        return cleaned
-    except Exception as exc:
-        logger.debug("customer_chat.firewall_non_fatal: %s", exc)
-        return text
-
-
-def _strip_display_garbage(text: str) -> str:
-    """D-115: مُطهّر العرض الحيّ المُصفّح — يحذف فواصل ⟦⟧ المشوّهة + تعليمات
-    system prompt المُسرَّبة + أي رموز غريبة قبل عرضها للطالب. شبكة أمان فوق
-    تنظيف الـ orchestrator. fail-open (لا يكسر البثّ).
-    """
-    if not text:
-        return text
-    try:
-        from app.services.skills.content_integrity_skill import _strip_garbage_markers
-
-        return _strip_garbage_markers(text)
-    except Exception:
-        return text
-
-
-def _apply_final_answer_redaction(text: str, support_level: int | None) -> str:
-    """D-114: حجب الإجابة النهائية الواعي بالفواصل على الإجابة المحفوظة.
-
-    support_level==1 ⇒ يُعفي كتلة المثال المحلول (الملفوفة بالفواصل) ويحجب ما
-    عداها؛ غير ذلك ⇒ حجب كامل (شبكة أمان D-113 + إزالة أي علامات شاردة). fail-open.
-    """
-    if not text or not text.strip():
-        return text
-    try:
-        from app.services.skills.answer_redaction_skill import redact_final_answers
-
-        redacted, _ = redact_final_answers(text, support_level)
-        return redacted
-    except Exception as exc:
-        logger.debug("customer_chat.final_redaction_non_fatal: %s", exc)
-        return text
-
+# ── D-173 Stage 2b: الأشرطة المساعدة استُخرجت حرفياً إلى customer_chat_support/ ──
+# الـ re-export يُبقي monkeypatch على `customer_chat.<name>` فعّالاً (المستدعي
+# chat_stream_ws يعيش هنا فيحلّ الأسماء من globals هذه الوحدة — قانون D-168).
+from app.api.routers.customer_chat_support.frames import (
+    _bind_local_conversation_id,  # noqa: F401 — re-export (D-168 late-binding)
+    _bind_stream_metadata,
+    _emit_terminal_frames,
+    _extract_client_context_messages,
+    _merge_history_with_client_context,
+    _try_build_math_ui_component,
+)
+from app.api.routers.customer_chat_support.pedagogy import (  # noqa: F401
+    _apply_complete_response_firewall,
+    _apply_final_answer_redaction,
+    _build_pedagogy_directive,
+    _count_confusion_signals,
+    _derive_correctness_override,
+    _evaluate_bkt_cards,
+    _PedagogySnapshot,
+    _persist_ui_component_cards,
+    _semantic_tutor_enabled,
+    _strip_display_garbage,
+)
+from app.api.routers.customer_chat_support.transport import (  # noqa: F401
+    _TURN_KEEPALIVE_INTERVAL_SECONDS,
+    TEXT_EVENT_TYPES,
+    _is_text_event,
+    _locked_send_json,
+    _run_turn_keepalive,
+    _ws_is_connected,
+)
 
 COMPATIBILITY_FACADE_MODE = True
 # تنبيه معماري: هذا المسار واجهة توافقية فقط ويُمنع فيه أي تنفيذ محلي لمنطق الدردشة.
@@ -122,311 +87,6 @@ router = APIRouter(
     prefix="/api/chat",
     tags=["Customer Chat"],
 )
-
-TEXT_EVENT_TYPES = {"delta", "assistant_delta", "assistant_final"}
-
-
-async def _locked_send_json(
-    websocket: WebSocket,
-    lock: asyncio.Lock,
-    payload: dict[str, object],
-) -> None:
-    """
-    إرسال JSON عبر WebSocket مع قفل تسلسلي لمنع تداخل البايتات.
-
-    ─────────────────────────────────────────────────────────────────
-    ISS-094 round 3 (D-096 — 2026-05-28): WS Send Concurrency Race
-    ─────────────────────────────────────────────────────────────────
-
-    Starlette's `WebSocket.send_json` لا يضمن thread/coroutine safety عند
-    استدعاءات متزامنة على نفس الـ socket. السيناريو الكارثي:
-
-      1. المستخدم يرسل سؤالاً
-      2. `_evaluate_and_emit_bkt` يُطلَق كـ background task
-         - يستدعي `await async_session_factory()` (DB query بطيء عبر Supabase)
-         - ثم `await websocket.send_json({"type":"ui_component", ...})`
-      3. بالتوازي، `stream_and_forward` يبدأ streaming deltas
-         - يستدعي `await websocket.send_json({"type":"assistant_delta", ...})`
-         - عشرات المرات
-      4. **RACE**: BKT's send و stream's send يتزامنان على نفس asgi.send
-      5. النتيجة الكارثية المحتملة:
-         - WebSocket protocol corruption (interleaved frame bytes)
-         - RuntimeError من asyncio (concurrent state machine updates)
-         - silent close بـ code 1006/1011
-      6. Frontend يرى WS close → reconnect → جديد قد يفشل لـ سبب آخر
-      7. بعد 3 محاولات auth retries → auth_error → logout → kick to login
-
-    لماذا SQLite لا يُظهر هذه الكارثة:
-      - BKT يكتمل في <50ms (DB محلي)
-      - stream يبدأ بعد BKT بكثير
-      - لا تزامن فعلي → لا race
-
-    لماذا Supabase يُسبب الكارثة:
-      - BKT DB write يأخذ 300ms-2s عبر network
-      - stream يبدأ فوراً ويستمر >5s
-      - فرصة تزامن عالية على send_json
-
-    D-118 (2026-06-17): بطاقات BKT/المسار لم تَعُد تُبثّ أثناء البثّ — تُقيَّم
-    متزامنةً (`_evaluate_bkt_cards`، بلا send) وتُصدَر بعد الإطار النهائي للمحتوى.
-    القفل يبقى ضرورياً للتزامن بين بثّ الـ deltas و keepalive (D-WS-FLAP-005)
-    وإصدار البطاقات بعد النهاية — كل send يمرّ عبره (D-096).
-    """
-    async with lock:
-        await websocket.send_json(payload)
-
-
-async def _persist_ui_component_cards(
-    *,
-    conversation_id: int,
-    cards: list[dict[str, object]],
-) -> None:
-    """يحفظ بطاقات الـ Generative UI المستقلة كصفوف مساعد content="" (ISS-106).
-
-    غير حرج: أي فشل يُسجَّل ولا يكسر الدور. كل بطاقة صف مستقل يحمل ui_component
-    بالشكل السلكي ({component, props, fallback_text}) فتُصيَّر من التاريخ بعد إعادة الدخول.
-    """
-    valid = [
-        c
-        for c in cards
-        if isinstance(c, dict) and isinstance(c.get("component"), str) and c["component"]
-    ]
-    if not valid:
-        return
-    try:
-        async with async_session_factory() as db:
-            persistence_service = CustomerChatBoundaryService(db)
-            for card in valid:
-                await persistence_service.save_message(
-                    conversation_id=conversation_id,
-                    role=MessageRole.ASSISTANT,
-                    content="",
-                    ui_component=card,
-                )
-    except Exception as exc:
-        logger.warning("ui_component_card_persist_failed: %s", exc, exc_info=True)
-
-
-def _derive_correctness_override(
-    question: str, history_messages: list[dict[str, str]]
-) -> str | None:
-    """D-157 (A1b): إشارة صواب مُتحقَّقة رمزياً لأدوار الاحتمالات — الأوراكل الحتمي
-    (D-155) يحكم بدل التخمين اللفظي. يُرجِع ``"correct"`` عند تأكيد رمزي حاسم لإجابة
-    صحيحة (⇒ يرفع الإتقان الدائم بصدق)، وإلا ``None`` (⇒ الإشارة اللفظية الثلاثية).
-    عالي الدقة: لا يُرجِع ``"incorrect"`` (غياب التأكيد ≠ خطأ — قد لا يُجيب الطالب
-    أصلاً). fail-open مطلق — أي تعذّر يُرجِع None.
-    """
-    try:
-        from app.infrastructure.clients.orchestrator_client import OrchestratorClient
-
-        combo = OrchestratorClient._load_canonical_combinations(question, history_messages)
-        if combo is None:
-            return None
-        pending = OrchestratorClient._pending_focus_from_history(history_messages)
-        if OrchestratorClient._verify_numeric_answer(question, combo, pending) in (
-            "final_ratio",
-            "step_correct",
-            "direction",
-        ):
-            return "correct"
-        if OrchestratorClient._verify_answer_against_combo(question, combo):
-            return "correct"
-    except Exception:  # pragma: no cover - fail-open
-        return None
-    return None
-
-
-async def _evaluate_bkt_cards(
-    *,
-    user_id: int,
-    conversation_id: int | None,
-    question: str,
-    history_messages: list[dict[str, str]],
-    support_level: int | None = None,
-    novel_item: bool = False,
-    correctness_signal: str | None = None,
-) -> None:
-    """D-119: تتبّع معرفي خلف الكواليس — بلا أي بطاقة تظهر للطالب.
-
-    يُجري كتابة BKT التحليلية append-only (D-074 — تغذّي ``support_level``
-    والبيداغوجيا التكيفية) ويشتقّ المسار التعلّمي (D-111) ويُسجِّله
-    (telemetry / لوحة المعلم مستقبلاً) — ثم **يُرجِع ``None``**. قرار المالك
-    (D-119): «تتبّع المعرفة» و«ترسيخ المهارة» خلف الكواليس، لا في سطح الطالب —
-    كانتا تظهران بعد كل دور فتُكرَّران وتُشوّشان. سطح الطالب = تعليم نظيف فقط.
-
-    D-126: يُمرِّر ``support_level`` (إن توفّر) لحساب **القناة الدائمة الصادقة**
-    (durable). في الدردشة الحرة ``support_level=None`` ⇒ durable يُحمَل دون تضخيم
-    (لا نمنح إتقاناً دائماً بلا دليل أداء غير مدعوم — يُغذَّى من وضع التحقق M8).
-
-    معزول كلياً: أي فشل (DB أو غيره) يُسجَّل ولا يكسر مسار المحادثة (D-074).
-    """
-    try:
-        # D-157 (A1b): الأوراكل الرمزي يُحسَب هنا (داخل المهمة الخلفية، خارج المسار
-        # الحيّ) — إشارة صواب مُتحقَّقة رمزياً لأدوار الاحتمالات تفوق التخمين اللفظي.
-        signal = correctness_signal or _derive_correctness_override(question, history_messages)
-        async with async_session_factory() as bkt_db:
-            evaluation = await BKTAnalyticsService(bkt_db).evaluate_and_record(
-                user_id=user_id,
-                session_id=conversation_id,
-                question=question,
-                history=history_messages,
-                support_level=support_level,
-                novel_item=novel_item,
-                correctness_signal=signal,
-            )
-        # D-119/D-126: التتبّع خلف الكواليس — نُسجِّل القناتين + فجوة الوهم بدل بطاقة.
-        logger.info(
-            "bkt_tracking concept=%s assisted=%.2f durable=%.2f illusion_gap=%.2f "
-            "load=%s (behind-the-scenes, no card)",
-            evaluation.concept_id,
-            evaluation.student_mastery_probability,
-            evaluation.durable_mastery,
-            evaluation.illusion_gap,
-            evaluation.cognitive_load_estimate,
-        )
-        # D-111: المسار التعلّمي التكيفي — يُشتقّ فوق إتقان BKT ويُسجَّل (telemetry).
-        # مُستهلَك خلف الكواليس (لا بطاقة طالب — D-119)؛ معزول داخل suppress.
-        with contextlib.suppress(Exception):
-            from app.services.skills.learning_path_skill import (
-                LearningPathInput,
-                get_learning_path_skill,
-            )
-
-            path = get_learning_path_skill().derive(
-                LearningPathInput(
-                    concept_id=evaluation.concept_id,
-                    mastery=evaluation.student_mastery_probability,
-                )
-            )
-            logger.info(
-                "learning_path next=%s difficulty=%s reco=%s (behind-the-scenes, no card)",
-                path.next_concept,
-                path.target_difficulty,
-                path.recommendation_text,
-            )
-    except Exception as exc:
-        # BKT must never break chat — log and continue (no student-facing impact).
-        logger.warning("bkt_tracking_failed: %s", exc, exc_info=True)
-    # D-119: لا بطاقات للطالب — التتبّع خلف الكواليس حصراً (return None).
-
-
-_PEDAGOGY_DIRECTIVE_TIMEOUT_S = 2.0
-
-#: D-114: علامات الحيرة — تكرارها (2+) يفرض المثال المحلول (support_level=1).
-_CONFUSION_MARKERS: tuple[str, ...] = (
-    "لم أفهم",
-    "لم افهم",
-    "مفهمتش",
-    "ما فهمت",
-    "مافهمت",
-    "ما افهم",
-    "لا أفهم",
-    "لا افهم",
-    "صعب",
-    "ماقدرت",
-    "ما قدرت",
-)
-
-
-class _PedagogySnapshot(NamedTuple):
-    """D-114: صورة المتعلم المُجمَّعة — تُمرَّر للسياق وتقود بوّابة المثال المحلول.
-
-    ``support_level`` الافتراض الآمن = 5 (محجوب كلياً) عند أي تعذّر — لا 1 — كي
-    لا يكشف فشلٌ واحد كل الإجابات (fail-closed).
-    """
-
-    directive_text: str
-    support_level: int
-    confusion_count: int
-    concept_id: str
-    mastery: float | None
-
-
-def _count_confusion_signals(history_messages: list[dict[str, str]] | None, question: str) -> int:
-    """يَعُدّ رسائل الطالب التي تُعبّر عن حيرة (السؤال الحالي + سجل المحادثة).
-
-    حتمي، بلا I/O. تكرار الحيرة (≥2) ⇒ المبتدئ عالق ويحتاج مثالاً محلولاً.
-    """
-    texts: list[str] = [str(question or "")]
-    for msg in history_messages or []:
-        if isinstance(msg, dict) and msg.get("role") == "user":
-            texts.append(str(msg.get("content") or ""))
-    count = 0
-    for text in texts:
-        low = text.strip()
-        if any(marker in low for marker in _CONFUSION_MARKERS):
-            count += 1
-    return count
-
-
-async def _build_pedagogy_directive(
-    *,
-    user_id: int,
-    question: str,
-    history_messages: list[dict[str, str]],
-) -> _PedagogySnapshot:
-    """D-104/D-114: يقرأ صورة المتعلم من BKT ويشتق توجيهاً تربوياً + support_level.
-
-    «المنسّق يملك صورة المتعلم» — قراءة الإتقان تقود سلوك المعلم: سقراطي للمتمكن،
-    مثال محلول كامل للمبتدئ المطلق (D-114)، والتقاط المفاهيم الخاطئة قبل أن تتجذر.
-    fail-open مطلق: أي تعذّر (DB/مهلة/تصنيف) ⇒ snapshot آمن (directive=""،
-    support_level=5 محجوب كلياً) — التوجيه لا يكسر دور الطالب أبداً.
-    """
-    try:
-        from app.services.skills.adaptive_pedagogy_skill import (
-            PedagogyInput,
-            get_adaptive_pedagogy_skill,
-        )
-        from app.services.skills.bkt_engine import (
-            classify_concept_with_context,
-            estimate_cognitive_load,
-        )
-
-        # ISS-112: تصنيف واعٍ بالسياق — «اشرح السؤال 2» يلتصق بمفهوم الحوار
-        concept_id = classify_concept_with_context(question, history_messages)
-
-        async def _read_snapshot() -> tuple[float | None, int]:
-            async with async_session_factory() as ped_db:
-                service = BKTAnalyticsService(ped_db)
-                mastery = await service.latest_mastery(user_id, concept_id)
-                count = await service.interaction_count(user_id, concept_id)
-                return mastery, count
-
-        mastery, interaction_count = await asyncio.wait_for(
-            _read_snapshot(), timeout=_PEDAGOGY_DIRECTIVE_TIMEOUT_S
-        )
-        directive = get_adaptive_pedagogy_skill().derive(
-            PedagogyInput(
-                question=question,
-                concept_id=concept_id,
-                mastery=mastery,
-                interaction_count=interaction_count,
-                cognitive_load=estimate_cognitive_load(question),
-            )
-        )
-        # D-114: الحيرة المتكرّرة تفرض المثال المحلول حتى لو الإتقان غير منخفض.
-        confusion = _count_confusion_signals(history_messages, question)
-        support_level = directive.support_level
-        if confusion >= 2:
-            support_level = 1
-        logger.info(
-            "pedagogy_directive level=%s support_level=%s concept=%s mastery=%s confusion=%s",
-            directive.guidance_level,
-            support_level,
-            concept_id,
-            f"{mastery:.2f}" if mastery is not None else "none",
-            confusion,
-        )
-        return _PedagogySnapshot(
-            directive_text=directive.directive_text,
-            support_level=support_level,
-            confusion_count=confusion,
-            concept_id=concept_id,
-            mastery=mastery,
-        )
-    except Exception as exc:
-        logger.debug("pedagogy_directive_failed (fail-open): %s", exc)
-        # fail-closed على support_level: 5 = محجوب كلياً (لا كشف عند الخطأ).
-        return _PedagogySnapshot("", 5, 0, "general", None)
 
 
 def get_chat_actor(
@@ -463,244 +123,6 @@ def get_customer_service(
 ) -> CustomerChatBoundaryService:
     """تبعية للحصول على خدمة حدود محادثة العملاء."""
     return CustomerChatBoundaryService(db)
-
-
-def _is_text_event(event: dict[str, object]) -> bool:
-    """يتحقق من أن الحدث نصي ومسموح بتجميعه داخل مخزن النص النهائي."""
-    return str(event.get("type", "")) in TEXT_EVENT_TYPES
-
-
-def _ws_is_connected(websocket: WebSocket) -> bool:
-    """يتحقق من أن WebSocket لا يزال في حالة CONNECTED قبل أي send_json.
-
-    D-WS-FLAP-001: يمنع RuntimeError عند محاولة الإرسال على socket مغلق،
-    وهو السبب الجذري لنمط الـ flapping (يعمل → يتعطل → يعود).
-    """
-    return websocket.client_state == WebSocketState.CONNECTED
-
-
-# ISS-098 (D-WS-FLAP-005 — 2026-05-29): فاصل الـ keepalive أثناء بثّ الدور.
-_TURN_KEEPALIVE_INTERVAL_SECONDS = 20.0
-
-
-async def _run_turn_keepalive(websocket: WebSocket, lock: asyncio.Lock) -> None:
-    """يُرسل إطار pong خفيفاً كل ~20s أثناء بثّ دور المحادثة.
-
-    الجذر (ISS-098): حلقة الاستقبال محجوبة طوال `await stream_task`، فلا
-    يستطيع الخادم قراءة ping العميل والرد بـ pong. مع زمن Supabase + إجابة
-    طويلة، يتجاوز الدور HEARTBEAT_TIMEOUT (90s) على العميل → close(1001)
-    كاذب → reconnect → الإجابة الجارية تضيع.
-
-    الحل: مهمة keepalive متزامنة تُرسل `pong` دورياً عبر نفس `send_lock`
-    (فلا تتعارض مع إرسال الـ deltas — D-096). العميل يتعامل مع أي رسالة
-    واردة كدليل حياة، وتحديداً `pong` يُلغي مؤقّت الـ heartbeat. تُلغى
-    المهمة فور انتهاء البثّ. لا ترفع استثناءً للخارج أبداً.
-    """
-    try:
-        while True:
-            await asyncio.sleep(_TURN_KEEPALIVE_INTERVAL_SECONDS)
-            if not _ws_is_connected(websocket):
-                return
-            try:
-                await _locked_send_json(
-                    websocket, lock, {"type": "pong", "payload": {"keepalive": True}}
-                )
-            except (WebSocketDisconnect, RuntimeError):
-                return
-            except Exception as exc:  # pragma: no cover - دفاعي
-                logger.debug("customer_chat.keepalive_send_failed: %s", type(exc).__name__)
-                return
-    except asyncio.CancelledError:
-        return
-
-
-def _bind_local_conversation_id(
-    event: dict[str, object], conversation_id: int | None
-) -> dict[str, object]:
-    """يربط معرف المحادثة المحلي بجميع أحداث البث لمنع اختلاط السياق في الواجهة."""
-    if conversation_id is None:
-        return event
-    payload = event.get("payload")
-    if isinstance(payload, dict):
-        payload["conversation_id"] = conversation_id
-    else:
-        event["payload"] = {"conversation_id": conversation_id}
-    return event
-
-
-def _bind_stream_metadata(
-    event: dict[str, object],
-    conversation_id: int | None,
-    request_id: str | None,
-) -> dict[str, object]:
-    """يربط بيانات التتبع القياسية بكل حدث بث لعزل السياق بين الطلبات."""
-    bound_event = _bind_local_conversation_id(event, conversation_id)
-    if not request_id:
-        return bound_event
-    payload = bound_event.get("payload")
-    if isinstance(payload, dict):
-        payload["request_id"] = request_id
-    else:
-        bound_event["payload"] = {"request_id": request_id}
-    return bound_event
-
-
-def _extract_client_context_messages(payload: dict[str, object]) -> list[dict[str, str]]:
-    """استخراج سياق المحادثة المرسل من الواجهة بشكل آمن ومحدود الحجم."""
-    raw_context = payload.get("client_context_messages")
-    if not isinstance(raw_context, list):
-        return []
-
-    sanitized: list[dict[str, str]] = []
-    for item in raw_context:
-        if not isinstance(item, dict):
-            continue
-        role = item.get("role")
-        content = item.get("content")
-        if role not in {"user", "assistant"}:
-            continue
-        if not isinstance(content, str):
-            continue
-        text = content.strip()
-        if not text:
-            continue
-        sanitized.append({"role": role, "content": text})
-        if len(sanitized) >= 50:
-            break
-    return sanitized
-
-
-def _merge_history_with_client_context(
-    persisted_history: list[dict[str, str]],
-    client_context: list[dict[str, str]],
-) -> list[dict[str, str]]:
-    """
-    دمج تاريخ قاعدة البيانات مع سياق العميل مع منع تسريب رسائل محادثات أخرى.
-
-    يعتمد على كشف التداخل بين ذيل persisted_history وبداية client_context
-    لضمان أن الرسائل المضافة تنتمي فعلاً لنفس المحادثة.
-    """
-    if not client_context:
-        return persisted_history
-    if not persisted_history:
-        return []
-
-    persisted_tail = persisted_history[-3:] if len(persisted_history) >= 3 else persisted_history
-    overlap_index: int | None = None
-    max_start = len(client_context) - len(persisted_tail)
-    for start in range(max_start, -1, -1):
-        window = client_context[start : start + len(persisted_tail)]
-        if window == persisted_tail:
-            overlap_index = start + len(persisted_tail)
-            break
-
-    if overlap_index is None:
-        return persisted_history
-
-    safe_client_tail = client_context[overlap_index:][-12:]
-    merged_history = list(persisted_history)
-    for message in safe_client_tail:
-        if message not in merged_history:
-            merged_history.append(message)
-    return merged_history[-80:]
-
-
-def _try_build_math_ui_component(response_text: str) -> dict | None:
-    """
-    مُعطَّل عمداً (ISS-108 / D-097 — قرار المستخدم: «إيقاف التقطيع + بطاقات محقَّقة فقط»).
-
-    كان يُمرِّر **نص الإجابة الحر** إلى ``math_pipeline._build_ui_component`` الذي
-    يُقطّع النثر بـ regex فضفاضة فيُنتج «خطوات» عشوائية بلا معنى (مؤكَّد حياً على
-    بيانات الإنتاج msg 3411: «بالعربي»، «أتمنى أن تكون الفكرة الآن واضحة») ويُصنّف
-    شرح الاحتمالات كـ«معادلات». بطاقة ``math_explanation_card`` صالحة فقط لمخرَج
-    محتوم منظَّم — لا لنثر LLM حر. البطاقات البصرية المحقَّقة (probability_tree /
-    combinations_visualizer / full_exercise_story) تأتي من المسار المنفصل
-    ``orchestrator_client._build_calculated_ui`` وتبقى عاملة.
-
-    القاعدة الدائمة (D-097): ممنوع تقطيع نثر LLM حر إلى بطاقة خطوات. أي بطاقة
-    Generative UI يجب أن تُبنى من بيانات محتومة مُتحقَّقة، لا من نص مُولَّد.
-    """
-    return None
-
-
-async def _emit_terminal_frames(
-    *,
-    websocket: WebSocket,
-    send_lock: asyncio.Lock,
-    pending_terminal_event: dict[str, object] | None,
-    assistant_message_persisted: bool,
-    complete_ai_response: str,
-    stream_error: HTTPException | Exception | None,
-    local_conversation_id: int | None,
-    stream_request_id: str,
-) -> None:
-    """
-    يضمن إرسال إطار نهائي واحد فقط لكل دور (assistant_final أو error)
-    وإطار 'persisted' فقط بعد حفظ فعلي. لا يُسمح بالفشل الصامت
-    الذي يبقي واجهة المستخدم في حالة تحميل أبدية (ISS-016/ISS-017).
-
-    D-096: يستخدم send_lock لمنع التزامن مع BKT task و stream_and_forward.
-    """
-    if assistant_message_persisted:
-        # بناء ui_component من النص المكتمل — لا يكسر المسار عند الفشل
-        ui_component = _try_build_math_ui_component(complete_ai_response)
-
-        if pending_terminal_event is not None:
-            # حقن ui_component في الـ payload الموجود
-            if ui_component and isinstance(pending_terminal_event.get("payload"), dict):
-                pending_terminal_event["payload"]["ui_component"] = ui_component
-            await _locked_send_json(websocket, send_lock, pending_terminal_event)
-        else:
-            payload: dict = {"content": ""}
-            if ui_component:
-                payload["ui_component"] = ui_component
-            await _locked_send_json(
-                websocket,
-                send_lock,
-                _bind_stream_metadata(
-                    normalize_streaming_event({"type": "assistant_final", "payload": payload}),
-                    local_conversation_id,
-                    stream_request_id,
-                ),
-            )
-        await _locked_send_json(
-            websocket,
-            send_lock,
-            _bind_stream_metadata(
-                normalize_streaming_event({"type": "persisted"}),
-                local_conversation_id,
-                stream_request_id,
-            ),
-        )
-        return
-
-    if isinstance(stream_error, HTTPException):
-        details = str(stream_error.detail)
-        status_code = stream_error.status_code
-    elif complete_ai_response or pending_terminal_event is not None:
-        details = "Failed to confirm assistant persistence before completion."
-        status_code = 500
-    elif stream_error is not None:
-        details = "Stream interrupted before assistant response completed."
-        status_code = 500
-    else:
-        details = "Assistant produced no response."
-        status_code = 500
-
-    await _locked_send_json(
-        websocket,
-        send_lock,
-        _bind_stream_metadata(
-            normalize_streaming_event(
-                {
-                    "type": "error",
-                    "payload": {"details": details, "status_code": status_code},
-                }
-            ),
-            local_conversation_id,
-            stream_request_id,
-        ),
-    )
 
 
 @router.websocket("/ws")
