+++ app/api/routers/customer_chat_support/__init__.py
@@ -1,0 +1,1 @@
+"""حزمة دعم customer_chat (D-173 Stage 2b) — انظر `_sources.py`."""
