--- app/services/skills/doctrine.py
@@ -1,1600 +1,0 @@
-"""
-Skills Doctrine — Single Source of Truth لقواعد كل Skill في المشروع.
-
-CLAUDE.md §0.5: «كل قدرة AI في النظام يجب أن تكون Skill — وحدة مستقلة قابلة
-للقياس والاختبار والاستبدال». هذا الملف هو المرجع المركزي لكل قواعد التشغيل
-التي يجب أن تحترمها الـ Skills.
-
-**لماذا doctrine module مستقل؟**
-
-قبل (Prompt Spaghetti):
-    - قواعد الشرح متناثرة بين system prompts و LLM instructions
-    - تغيير قاعدة = تعديل عشرات الـ prompts يدوياً
-    - لا يوجد invariant CI gate يضمن أن الـ doctrine لم تنحرف
-
-بعد (Skills Doctrine Module):
-    - كل قاعدة تعيش كثابت Python (`tuple[str, ...]` أو `dict[str, str]`)
-    - versioning صريح: تغيير قاعدة = bump version + update CI gate
-    - اختبارات unit تثبت أن كل caller يحترم الـ doctrine الحية
-    - CI gate يحرس على عدم drift بين الـ doctrine في الـ Skill والـ system prompt
-
-**كيف يستخدم Skill قاعدة؟**
-
-```python
-from app.services.skills.doctrine import (
-    RETRIEVAL_DOCTRINE,
-    EXPLANATION_DOCTRINE,
-    MODEL_ANSWER_RELIANCE_RULES,
-)
-
-# في Skill:
-class BACExerciseSkill:
-    def _build_explanation_prompt(self, full_content: str) -> str:
-        return (
-            "أنت أستاذ بكالوريا جزائر. اشرح وفقاً للقواعد التالية:\\n"
-            + "\\n".join(f"- {r}" for r in EXPLANATION_DOCTRINE)
-            + f"\\n\\nالتمرين + الإجابة النموذجية:\\n{full_content}"
-        )
-```
-
-**كيف يضمن CI أن الـ doctrine متّسقة؟**
-
-`scripts/fitness/check_skills_doctrine.py` يفحص:
-1. كل Skill في `app/services/skills/` يستورد على الأقل قاعدة doctrine واحدة.
-2. أي تعديل على قاعدة doctrine يجب أن يرافقه bump في الـ version.
-3. الـ system prompts في `local_graph.py` تحتوي تمثيلاً قابلاً للتحقق من
-   `EXPLANATION_DOCTRINE` (drift detection).
-
-التاريخ:
-    2026-05-18 (ISS-080 / D-068): إنشاء الـ doctrine module — استخراج
-    EXPLANATION_DOCTRINE من `bac_exercise_skill.py`.
-
-    2026-05-18 (ISS-CI-GREEN-001 / D-069): إضافة RETRIEVAL_DOCTRINE،
-    MODEL_ANSWER_RELIANCE_RULES، CONTENT_INVOCATION_RULES، DETAILED_EXPLANATION_RULES
-    كقواعد رسمية مستقلة، تطبيقاً لطلب المستخدم بتطوير منظومة الـ skills.
-
-    2026-05-19 (D-070): تطوير منظومة Skills — إضافة:
-    - CONTENT_INVOCATION_DOCTRINE: بروتوكول استدعاء المحتوى التعليمي خطوة بخطوة.
-    - MODEL_ANSWER_EXPLANATION_DOCTRINE: كيفية شرح الإجابة النموذجية للطالب.
-    - STEP_BY_STEP_EXPLANATION_RULES: قواعد الشرح خطوة بخطوة.
-    - SKILL_INVOCATION_PROTOCOL: بروتوكول استدعاء الـ Skills.
-"""
-
-from __future__ import annotations
-
-from typing import Final
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Doctrine Version Pinning
-# ─────────────────────────────────────────────────────────────────────────────
-#
-# عند تعديل أي قاعدة من الـ doctrines أدناه:
-#   1. حدّث الـ version (`MAJOR.MINOR.PATCH`).
-#   2. شغّل `python scripts/fitness/check_skills_doctrine.py` ليُجدد الـ
-#      doctrine hash في الـ lock file.
-#   3. أضف entry في `.memory/decisions.md` يشرح *لماذا* تغيّرت.
-
-#: نسخة doctrine استدعاء/استرجاع المحتوى — تُستخدم لـ drift detection في CI.
-RETRIEVAL_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-#: نسخة doctrine الشرح — تطابق `EXPLANATION_DOCTRINE_VERSION` في
-#: `bac_exercise_skill.py` (يبقى متزامناً عبر تعريف موحَّد هنا).
-#: v3.0.0 (D-113 / ISS-115 — وَهْم الإتقان): تحوُّل ثوري من «شرح يكشف الحل»
-#: إلى «توليد مُجبَر سقراطي» — ممنوع منعاً باتاً كشف أي نتيجة نهائية للطالب.
-#: v3.1.0 (D-114): استثناء المثال المحلول المكشوف — عُكِس في D-115 (كارثة حيّة).
-#: v3.2.0 (D-115 / ISS-116): لا كشف؛ المبتدئ الغارق يتلقّى بروتوكولاً سقراطياً
-#: أبسط (فكرة واحدة + خطوة واحدة) لا مثالاً مكشوفاً. القاعدة الذهبية مطلقة.
-EXPLANATION_DOCTRINE_VERSION: Final[str] = "3.2.0"
-
-#: نسخة doctrine الاعتماد على الإجابة النموذجية.
-#: ⚠️ D-113: هذه القواعد تُستخدم *حصراً* في وضع التحقق (verification mode) —
-#: حيث يُقدّم الطالب حلّه ونتحقق منه. ممنوع استخدامها في الشرح العادي
-#: (الشرح العادي سقراطي بحت — لا يكشف الإجابة النموذجية).
-MODEL_ANSWER_RELIANCE_VERSION: Final[str] = "1.0.0"
-
-#: نسخة doctrine حجب الإجابة النهائية (D-113 — شبكة الأمان الحتمية).
-#: v1.1.0 (D-114): إعفاء واعٍ بالفواصل — كتل المثال المحلول الملفوفة بالفواصل
-#: الحارسة معفاة من الحجب عند support_level==1؛ غياب الفواصل = حجب كامل (fail-closed).
-ANSWER_REDACTION_DOCTRINE_VERSION: Final[str] = "1.1.0"
-
-#: نسخة doctrine البروتوكول السقراطي المنضبط (D-115 يَعكِس D-114 / ISS-116).
-WORKED_EXAMPLE_DOCTRINE_VERSION: Final[str] = "2.0.0"
-
-#: D-114: الفواصل الحارسة لكتلة المثال المحلول — عربي خالص (لا لاتيني/أقواس
-#: ASCII كي لا يكسرها مُطهّر اللغة أو مُفكِّك Markdown). الحجب الواعي يُعفي ما
-#: *بداخلها* عند support_level==1 ويحجب ما خارجها (راجع redact_final_answers).
-#: نفس القيمتين منسوختان في microservices response_sanitizer (حدّ معماري).
-WORKED_EXAMPLE_OPEN: Final[str] = "⟦مثال_محلول⟧"
-WORKED_EXAMPLE_CLOSE: Final[str] = "⟦/مثال_محلول⟧"
-
-#: نسخة doctrine الشرح التفصيلي.
-DETAILED_EXPLANATION_VERSION: Final[str] = "1.0.0"
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Content Invocation / Retrieval Doctrine
-# ─────────────────────────────────────────────────────────────────────────────
-# هذه القواعد تحكم *كيف يجب أن يُستدعى المحتوى التعليمي* من قاعدة المعرفة.
-# المُستفيدون: `BACExerciseSkill._retrieve`, `_stream_local_retrieval_response`,
-# `exercise_retrieval.detect_exercise_retrieval`.
-# ─────────────────────────────────────────────────────────────────────────────
-
-RETRIEVAL_DOCTRINE: Final[tuple[str, ...]] = (
-    "Indexed-first retrieval: قَبل البحث الواسع، حاول مطابقة الفهرس المحدد "
-    "(year + session + subject + exercise_number) في `knowledge_index`.",
-    "Atomic file load: عند توفر `matched_entry`، حمِّل ملفاً واحداً بالضبط. "
-    "لا scan على `knowledge_base/` يقرأ ملفات أخرى.",
-    "Display strip: المُخرَج للمستخدم يجب أن يخلو من YAML frontmatter، "
-    "أي قسم بعنوان «عناصر الإجابة» / «الإجابة النموذجية» / «الحل» / "
-    "«Solutions» / «وسوم البحث»، أو أي علامات chunk الداخلية مثل `[ex:`، "
-    "`[sol:`، `[grading:`.",
-    "Streaming chunk integrity: التقسيم يجب أن يحافظ على سلامة LaTeX. "
-    "أي `$...$` أو `$$...$$` أو `\\(...\\)` يُكشف ذرّياً.",
-    "Pre-empt orchestrator: عندما `matched_entry is not None`، يجب أن يبث الـ "
-    "monolith المحتوى المُفهرَس النظيف *قبل* محاولة orchestrator-service.",
-    "Conversation context: عند سؤال شرح/استفسار بدون مرجع صريح، يجب فحص "
-    "`history_messages` أولاً قبل اللجوء إلى retriever عام.",
-    'No leakage: لا يُرسل JSON envelope (مثل `{"المصدر":"...","التمرين":"..."}`) '
-    "للمستخدم. الـ caller المسؤول هو `_extract_human_readable_response`.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Explanation Doctrine — Socratic No-Answer (v3.0.0)
-# ─────────────────────────────────────────────────────────────────────────────
-# D-113 / ISS-115 (وَهْم الإتقان — The Illusion of Mastery):
-#   الكارثة: كشف الحلّ الكامل (P(A)=14/165، C(11,3)=165، E(X)=1.73) يُولّد عند
-#   الطالب «وهم الطلاقة» — يشعر أنه فهم بينما درّب التعرّف لا التوليد، فينهار
-#   يوم الامتحان. المعالجة (Bjork / Roediger / Kapur / Renkl): توليد مُجبَر،
-#   سُلّم تلميحات متصاعد، محاولة قبل الشرح، و«لم أفهم» = تشخيص لا إعادة اشتقاق.
-#
-#   القاعدة الذهبية فوق كل القواعد: لا تكشف أبداً نتيجةً أو خطوةً يستطيع الطالب
-#   توليدها بنفسه. ابحث عن حافّة ما يستطيعه واعمل خطوةً واحدة وراءها (ZPD).
-#
-# ⚠️ الإجابة النموذجية تُستخدم *داخلياً* كمرجع لتوجيهك الصحيح — لكنها لا تصل
-#    الطالب أبداً في الشرح العادي (تُكشف فقط في وضع التحقق المنفصل).
-#
-# ملاحظة CI: أول 5 قواعد تُبنى منها `build_exercise_explanation_prompt`، ويجب
-#   أن تحوي مجتمعةً المراسي: «الإجابة النموذجية» و«LaTeX» و«حرفياً».
-# ─────────────────────────────────────────────────────────────────────────────
-
-EXPLANATION_DOCTRINE: Final[tuple[str, ...]] = (
-    # — القاعدة الذهبية (Golden Rule) —
-    "القاعدة الذهبية: لا تكشف أبداً نتيجةً نهائية أو خطوةً يستطيع الطالب توليدها "
-    "بنفسه. قُد بالأسئلة والتلميحات لا بالحلول.",
-    # — المنع المطلق (anchors: «الإجابة النموذجية» + «حرفياً») —
-    "ممنوع منعاً باتاً كشف الإجابة النموذجية أو أي نتيجة نهائية (أرقام/كسور/صيغ "
-    "مثل P(A)=14/165 أو C(11,3)=165 أو $$\\boxed{...}$$) — لا حرفياً ولا بإعادة "
-    "صياغة. الطالب يصل للإجابة بنفسه؛ أنت تقود فقط.",
-    # — محاولة قبل الشرح (Productive Failure) —
-    "محاولة قبل الشرح: لا تشرح مسألةً لم يحاولها الطالب. ادعُه للمحاولة أولاً، "
-    "ثم وجِّه محاولته. حتى المحاولة الخاطئة تستحق التوجيه — غير المُحاوَلة لا.",
-    # — سُلّم التلميحات المتصاعد (Hint Ladder) —
-    "سُلّم التلميحات المتصاعد عند التعثّر: (1) أشِر إلى المبدأ، (2) اسأل سؤالاً "
-    "موجِّهاً، (3) اكشف الهدف الجزئي التالي، (4) اكشف الخطوة التالية فقط — لا "
-    "تقفز إلى الإجابة أبداً. معظم التعثّر ينحلّ عند 1-2.",
-    # — بروتوكول «لم أفهم» (anchor: LaTeX) —
-    "عند «لم أفهم» أو «اشرح لي»: شخّص موضع الكسر بسؤال أولاً («أين بالضبط: في "
-    "اختيار الطريقة أم في عدّ الحالات؟»)، ثم تلميح أدنى واحد عند تلك النقطة — "
-    "لا تُعِد الاشتقاق. اكتب أي رمز رياضي بـ LaTeX داخل $...$ أو $$...$$.",
-    # — الشرح الذاتي (Self-Explanation) —
-    "اطلب «لماذا؟» قبل الانتقال: بعد كل تلميح، اطلب من الطالب أن يبرّر الخطوة "
-    "أو يتوقّع التالية. لا قراءة سلبية — بناء فعّال.",
-    # — التفنيد بعد الخطأ (Refutation, not preemption) —
-    "دع الخطأ يقع ثم فنّده: لا تستبق الأخطاء المتوقّعة. إذا أخطأ الطالب، أظهِر "
-    "الخطأ ثم فنّد النموذج الخاطئ صراحةً («ظننتَ الترتيب مهمّاً، لكن... لأن...»).",
-    # — نطاق الشرح —
-    "إذا طلب الطالب جزءاً محدداً (I/II/III/أ/ب/ج)، اقتصر عليه ولا تتعدّاه.",
-    # — اللغة والشكل (D-069) —
-    "اللغة عربية فصحى نقية. لا كلمات روسية/صينية/إسبانية مُسرَّبة "
-    "(يُطبَّق `sanitize_response` نهاية كل بث للضمان).",
-    "تجنّب جدار النص: أقسام قصيرة، أسئلة موجزة، بلا جداول Markdown، تشبيه واحد يكفي.",
-    # — مكافحة تسرّب الموضوع (D-097) —
-    "التزم بمعطيات هذا التمرين حصراً. ممنوع أي موضوع خارجي (طب/أشعة/مثال بعيد). "
-    "إن غابت المعلومة عن التمرين فقل ذلك — لا تخترع سياقاً.",
-    # — منع تسرّب الواجهات (D-106) —
-    "ممنوع إخراج HTML/JSX أو أكواد واجهات (<div>، className...) في نص الطالب — "
-    "الواجهات تُبنى حصراً عبر قناة ui_component المُهيكلة.",
-    # — المبتدئ الغارق: انضباط لا كشف (D-115 يَعكِس D-114) —
-    "المبتدئ الغارق (support_level==1 أو حيرة متكرّرة) لا يتلقّى حلاً مكشوفاً، بل "
-    "بروتوكولاً سقراطياً أبسط: ثبّت فكرة واحدة («ما المعطى؟ ما المطلوب؟ هل نعدّ أم "
-    "نحسب احتمالاً شرطياً؟») ثم خطوة واحدة صغيرة جداً — لا جدران، لا مثال موازٍ، "
-    "لا قفز موضوع، لا لغة غير عربية. القاعدة الذهبية تبقى مطلقة (راجع WORKED_EXAMPLE_DOCTRINE).",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Answer Redaction Doctrine — شبكة الأمان الحتمية (D-113)
-# ─────────────────────────────────────────────────────────────────────────────
-# الدفاع الأول هو الـ doctrine السقراطي + إرسال أسئلة-فقط للـ LLM. لكن النماذج
-# المجانية قد تخترع/تخمّن نتيجة نهائية. هذه القواعد تحكم الحارس الحتمي الأخير
-# (`AnswerRedactionSkill`) الذي يحذف أي نتيجة نهائية تسرّبت قبل وصولها للطالب.
-# المُستهلكون: `sanitize_final_text` (orchestrator + local) + المسار المحلي.
-# ─────────────────────────────────────────────────────────────────────────────
-
-ANSWER_REDACTION_DOCTRINE: Final[tuple[str, ...]] = (
-    "احذف دائماً `$$\\boxed{...}$$` (و`\\boxed{...}`) — هو علامة النتيجة النهائية "
-    "في سياق البكالوريا، استبدله بدعوة سقراطية «احسبها بنفسك».",
-    "احذف أسطر النتيجة النهائية الصريحة: «الإجابة النهائية:»، «النتيجة:»، «إذن "
-    "... =» المُذيَّلة برقم/كسر كخلاصة (لا كخطوة وسيطة).",
-    "نطاق ضيّق لتجنّب false-positive: لا تلمس الصيغ التعليمية الوسيطة "
-    "(`نستخدم C(n,k)`، تعريف القانون) — فقط النتائج الختامية الصريحة.",
-    "احذف جداول توزيع الاحتمال الكاملة `P(X=...)` التي تكشف القانون جاهزاً.",
-    "fail-open مطلق: أي فشل في الحارس ⇒ النص كما هو (لا يكسر دور الطالب).",
-    "الحارس حتمي — بلا LLM، بلا عشوائية — قابل للاختبار بـ pytest.",
-    # — الإعفاء الواعي بالفواصل (D-114) —
-    "إعفاء support_level==1 الواعي بالفواصل: عند support_level==1 ووجود الفواصل "
-    "الحارسة (⟦مثال_محلول⟧ … ⟦/مثال_محلول⟧)، يُحجب كل ما *خارج* الفواصل ويمرّ ما "
-    "*بداخلها* (مثال محلول على مسألة مماثلة آمن الكشف). تُزال علامات الفواصل من "
-    "المُخرَج النهائي. غياب الفواصل أو support_level != 1 ⇒ حجب كامل (fail-closed) — "
-    "فلا تتسرّب إجابة تمرين الطالب المُقيَّم أبداً.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Disciplined Socratic Protocol Doctrine (D-115 — يَعكِس D-114)
-# ─────────────────────────────────────────────────────────────────────────────
-# D-114 (المثال المحلول المكشوف) كان كارثة حيّة: النموذج المجاني سرّب فواصل ⟦⟧
-# مشوّهة + system prompt إنجليزي + هلوسة فرنسية/ألمانية + قفز مواضيع + جدران نص.
-# التصحيح (مالك المشروع): المشكلة ليست «لا تكشف» بل التنفيذ الكارثي. القاعدة
-# الذهبية: «الرد الجيد لا يقول الجواب بل يقول ما الخطوة التالية الصحيحة».
-# (الاسم WORKED_EXAMPLE_DOCTRINE يبقى للتوافق؛ المحتوى صار بروتوكول سقراطي منضبط.)
-# المُستهلكون: search.SynthesizerNode (التوليد السقراطي)، content_integrity +
-# response_sanitizer (المُطهّر المُصفّح)، answer_redaction (الحجب fail-closed).
-# ─────────────────────────────────────────────────────────────────────────────
-
-WORKED_EXAMPLE_DOCTRINE: Final[tuple[str, ...]] = (
-    "لا مثال محلول مكشوف (D-115 يَعكِس D-114): الطالب الغارق يتلقّى بروتوكولاً "
-    "سقراطياً منضبطاً قصيراً — صياغة المطلوب + نوع المسألة + سؤال تشخيصي واحد + "
-    "أصغر خطوة تالية. لا جدران نص، لا أمثلة موازية، لا كشف للجواب.",
-    "support_level (1..5) يحكم عمق التلميح فقط: 1=ثبّت فكرة واحدة للغارق بأبسط "
-    "لغة ← 5=ادفع المتمكّن ليستنتج بنفسه. لا يفتح كشف الجواب أبداً في التدفّق العادي.",
-    "السلطة الوحيدة للتوليد التعليمي هي الـ orchestrator/LangGraph (SynthesizerNode) "
-    "— لا توليد تعليمي في المونوليث؛ النظام يتعطّل صراحةً بدون الخدمات المصغرة (D-112).",
-    "حجب نهائي fail-closed: أي نتيجة نهائية (أرقام/كسور/$$\\boxed{}$$) تُحجب قبل "
-    "وصول الطالب. الكشف لوضع مراجعة منفصل فقط (M8 — مؤجَّل، خارج هذا التدفّق).",
-    "مُطهّر مُصفّح إلزامي: أي فواصل ⟦⟧ مشوّهة/مقسّمة، أو تعليمات system prompt "
-    "مُسرَّبة (WARM-UP...)، أو لغة غير عربية معزولة — تُحذف على deltas + النهائي "
-    "في المونوليث والـ orchestrator (شبكة أمان حتمية، fail-open).",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Model Answer Reliance Rules
-# ─────────────────────────────────────────────────────────────────────────────
-# قواعد رسمية لكيفية الاعتماد على الإجابة النموذجية أثناء الشرح المفصل.
-# هذه قواعد *أحكم* من EXPLANATION_DOCTRINE — لا تُكسر بدون ADR.
-# ─────────────────────────────────────────────────────────────────────────────
-
-MODEL_ANSWER_RELIANCE_RULES: Final[tuple[str, ...]] = (
-    # — مبدأ الاحتجاج (citation principle) —
-    "الإجابة النموذجية هي *مرجع الحُجَّة*. كل نتيجة عددية يجب أن تُحتَجَّ بها.",
-    # — الأرقام (numeric invariants) —
-    "كل قيمة عددية في الشرح يجب أن تطابق نظيرتها في الإجابة النموذجية. "
-    "مثال: لو الإجابة تقول `g'(x) = (x+2)e^(-x)`، الشرح لا يكتب "
-    "`g'(x) = (x+2)·e^(-x)` بصيغة مختلفة فحسب — بل نفس الصيغة بالضبط.",
-    # — الصيغ الرمزية —
-    "الصيغ النهائية (limits, derivatives, integrals) المكتوبة في الإجابة "
-    "تُحفَظ كما هي. الشرح يضيف *لماذا* وليس *كيف نُعيد كتابتها*.",
-    # — التحقق (verification) —
-    "بعد كل خطوة كبيرة، تَحقَّق أن النتيجة المستنبطة تطابق الإجابة النموذجية. "
-    "إن لم تطابق، فالشرح غير صحيح ويجب تعديله، لا تعديل الإجابة النموذجية.",
-    # — المنطقة الحرة (interpretive freedom) —
-    "الشرح حر في:\n"
-    "  - اختيار ترتيب الأفكار التعليمية.\n"
-    "  - إضافة تذكيرات بقواعد سابقة (مثلاً تذكير بقاعدة لوبيتال قبل تطبيقها).\n"
-    "  - إضافة تفسيرات هندسية/فيزيائية للنتيجة.\n"
-    "  - استخدام أمثلة عددية مساعدة (مع تأشيرها كـ «مثال توضيحي»).",
-    # — المنطقة الممنوعة (forbidden zone) —
-    "الشرح ممنوع من:\n"
-    "  - تعديل النتائج العددية في الإجابة النموذجية.\n"
-    "  - إعادة صياغة الصيغ بشكل مختلف عن الإجابة النموذجية.\n"
-    "  - نسخ الإجابة النموذجية حرفياً (يجب وجود قيمة تعليمية مضافة).\n"
-    "  - تجاهل أجزاء من الإجابة النموذجية إذا كانت مطلوبة في السؤال.",
-    # — الحالة الخاصة: الإجابة النموذجية مفقودة —
-    "إذا لم تتوفر الإجابة النموذجية في `full_content`، الـ Skill يُرجِع "
-    "`SkillFailure(reason='no_explanation_context')` بدلاً من توليد شرح "
-    "بدون مرجع حُجَّة.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Detailed Explanation Rules — الشرح المفصل للطالب
-# ─────────────────────────────────────────────────────────────────────────────
-# هذه القواعد تُكمِّل EXPLANATION_DOCTRINE بضوابط الـ *حجم* و *التدريج*
-# لكل سؤال نوع. يُستخدمها `_classify_question_budget` في `local_graph.py`
-# لتخصيص `max_tokens` المناسب.
-# ─────────────────────────────────────────────────────────────────────────────
-
-DETAILED_EXPLANATION_RULES: Final[tuple[str, ...]] = (
-    # — السؤال المفهومي (concept) —
-    "سؤال «ماذا نقصد بـ X» / «ما هو معنى X» = شرح قصير (≤ 350 token).",
-    "السؤال المفهومي يبدأ بـ تعريف صريح، ثم مثال واحد، ثم تطبيق على التمرين.",
-    # — السؤال التبريري (justification) —
-    "سؤال «لماذا X» / «علِّل X» = شرح متوسط (≤ 450 token).",
-    "السؤال التبريري يبدأ بـ القاعدة، ثم تطبيقها على بيانات التمرين، "
-    "ثم استنتاج النتيجة المُحتجة بها.",
-    # — السؤال المنهجي (method) —
-    "سؤال «كيف نُثبت X» / «كيف نحسب X» = شرح موسَّع (≤ 600 token).",
-    "السؤال المنهجي يبدأ بـ المنهجية العامة، ثم تطبيقها على بيانات التمرين، "
-    "ثم النتيجة في `$$\\boxed{...}$$`.",
-    # — السؤال الافتراضي (default) —
-    "سؤال «اشرح ...» بدون تخصيص = شرح متوازن (≤ 700 token).",
-    # — السؤال الشامل (full) —
-    "سؤال «اشرح التمرين كاملاً» = شرح طويل (≤ 900 token).",
-    "الشرح الشامل يَطوي كل الأجزاء (I/II/III) بفقرات منفصلة، كل واحدة مُختصرة.",
-    # — قواعد تنسيقية موحَّدة —
-    "كل خطوة تبدأ بـ رقم (1) (2) (3) لتسهيل المتابعة.",
-    "العناوين الفرعية بـ `### الجزء I` / `### الخطوة 1` لتفصيل البصري.",
-    "الفرضيات بـ `**فرضية:**`، النتائج بـ `**نتيجة:**`.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Content Invocation Doctrine (D-070)
-# ─────────────────────────────────────────────────────────────────────────────
-# بروتوكول استدعاء المحتوى التعليمي — يحدد الخطوات الإلزامية التي يجب أن
-# يتبعها أي Skill أو handler عند استدعاء محتوى من `knowledge_base/`.
-# المُستفيدون: `BACExerciseSkill`, `local_graph`, `orchestrator_client`.
-# ─────────────────────────────────────────────────────────────────────────────
-
-CONTENT_INVOCATION_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-CONTENT_INVOCATION_DOCTRINE: Final[tuple[str, ...]] = (
-    # — الخطوة 1: تحديد النية (Intent Classification) —
-    "الخطوة 1 — تحديد النية: قبل أي استدعاء، صنِّف نية الطالب إلى: "
-    "(retrieval) طلب التمرين | (explanation) طلب شرح | (concept) سؤال مفهومي. "
-    "النية تحدد الـ Skill المُستدعى.",
-    # — الخطوة 2: تحديد المرجع (Reference Resolution) —
-    "الخطوة 2 — تحديد المرجع: ابحث عن مرجع صريح (سنة + دورة + رقم تمرين) "
-    "في السؤال الحالي أولاً، ثم في `history_messages` آخر 3 رسائل. "
-    "إن لم يوجد مرجع صريح → `reference=None` → fallback للـ LLM العام.",
-    # — الخطوة 3: الفهرسة (Index Lookup) —
-    "الخطوة 3 — الفهرسة: مع `reference != None`، ابحث في `knowledge_index` "
-    "بـ (year, session, subject_number, exercise_number). "
-    "الفهرس يُرجع `matched_entry: ExerciseEntry | None`.",
-    # — الخطوة 4: تحميل الملف (File Load) —
-    "الخطوة 4 — تحميل الملف: مع `matched_entry != None`، حمِّل ملفاً واحداً "
-    "بالضبط عبر `load_exercise_content(matched_entry.file_path)`. "
-    "لا scan على المجلد. لا قراءة ملفات إضافية.",
-    # — الخطوة 5: التنظيف (Display Strip) —
-    "الخطوة 5 — التنظيف: مرِّر المحتوى عبر `format_exercise_for_display()` "
-    "لإزالة: YAML frontmatter، أقسام الإجابة النموذجية، علامات chunk الداخلية. "
-    "المُخرَج للمستخدم يجب أن يكون نصاً تعليمياً نظيفاً فقط.",
-    # — الخطوة 6: البث (Streaming) —
-    "الخطوة 6 — البث: ابثّ المحتوى النظيف عبر `assistant_delta` envelopes. "
-    "حافظ على سلامة LaTeX: لا تقطع `$...$` أو `$$...$$` بين chunks.",
-    # — الخطوة 7: السياق (Context Injection) —
-    "الخطوة 7 — السياق: إذا كانت النية (explanation)، أضف `full_content` "
-    "(يشمل الإجابة النموذجية) إلى سياق الـ LLM. "
-    "لا تُرسل `full_content` للمستخدم مباشرة — فقط للـ LLM.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Model Answer Explanation Doctrine (D-070)
-# ─────────────────────────────────────────────────────────────────────────────
-# كيفية شرح الإجابة النموذجية للطالب — يُكمِّل EXPLANATION_DOCTRINE بتفاصيل
-# المنهجية التعليمية لكل نوع من أنواع الأسئلة في بكالوريا الجزائر.
-# ─────────────────────────────────────────────────────────────────────────────
-
-MODEL_ANSWER_EXPLANATION_VERSION: Final[str] = "1.0.0"
-
-MODEL_ANSWER_EXPLANATION_DOCTRINE: Final[tuple[str, ...]] = (
-    # — مرحلة الفهم (Understanding Phase) —
-    "مرحلة الفهم: ابدأ بـ «ماذا يطلب السؤال؟» — اقرأ المطلوب بصوت عالٍ "
-    "للطالب (جملة واحدة). هذا يُثبِّت الهدف قبل الشرح.",
-    # — مرحلة الأدوات (Tools Phase) —
-    "مرحلة الأدوات: اذكر القاعدة/النظرية/الأداة الرياضية المُستخدمة "
-    "(مثال: «نستخدم قاعدة لوبيتال لأن النهاية من الشكل 0/0»). "
-    "لا تطبِّق الأداة قبل أن تُعرِّفها.",
-    # — مرحلة التطبيق (Application Phase) —
-    "مرحلة التطبيق: طبِّق الأداة على بيانات التمرين خطوة بخطوة. "
-    "كل خطوة في سطر منفصل. كل رمز رياضي في LaTeX. "
-    "اربط بين الخطوات بـ «إذن» / «ومنه» / «بالتعويض».",
-    # — مرحلة التحقق (Verification Phase) —
-    "مرحلة التحقق: بعد الوصول للنتيجة، تحقق منها بطريقة مختلفة "
-    "(مثال: تعويض قيمة في المعادلة الأصلية، أو فحص الإشارة، "
-    "أو التحقق من الوحدات في الفيزياء).",
-    # — مرحلة التفسير (Interpretation Phase) —
-    "مرحلة التفسير: اختم بـ «ماذا تعني هذه النتيجة؟» — تفسير هندسي "
-    "(مثال: «هذا يعني أن الدالة تتزايد على المجال...») أو فيزيائي "
-    '(مثال: «هذا يعني أن الجسم يتسارع بمعدل...").',
-    # — قواعد خاصة بالرياضيات —
-    "في الرياضيات: كل نتيجة وسيطة تُكتب في `$...$`. "
-    "النتيجة النهائية في `$$\\boxed{...}$$`. "
-    "الجداول (جدول الإشارات، جدول التغيرات) تُرسم بـ Markdown.",
-    # — قواعد خاصة بالفيزياء —
-    "في الفيزياء: كل معادلة تُكتب أولاً بالرموز، ثم بالتعويض العددي، "
-    "ثم بالنتيجة مع الوحدة. مثال: $v = d/t = 100/5 = 20 \\text{ m/s}$.",
-    # — قواعد خاصة بالعلوم الطبيعية —
-    "في العلوم الطبيعية: الشرح يتبع منطق السبب والنتيجة. "
-    "كل استنتاج يُبنى على ملاحظة أو تجربة مذكورة في التمرين.",
-    # — التكيف مع مستوى الطالب —
-    "إذا أعاد الطالب السؤال أو قال «لم أفهم»: أعِد الشرح بمثال عددي "
-    "أبسط أولاً، ثم انتقل للتمرين الأصلي. لا تكرر نفس الشرح حرفياً.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Step-by-Step Explanation Rules (D-070)
-# ─────────────────────────────────────────────────────────────────────────────
-# قواعد الشرح خطوة بخطوة — تُطبَّق على كل شرح مفصل بغض النظر عن النوع.
-# هذه القواعد تُكمِّل DETAILED_EXPLANATION_RULES بضوابط التسلسل المنطقي.
-# ─────────────────────────────────────────────────────────────────────────────
-
-STEP_BY_STEP_EXPLANATION_VERSION: Final[str] = "1.0.0"
-
-STEP_BY_STEP_EXPLANATION_RULES: Final[tuple[str, ...]] = (
-    # — التسلسل المنطقي —
-    "كل خطوة تنبثق من السابقة: لا قفز منطقي. إذا كانت الخطوة n تعتمد على "
-    'نتيجة الخطوة n-2، اذكر ذلك صراحةً («من الخطوة 2 وجدنا أن...").',
-    # — الترقيم الإلزامي —
-    "الترقيم إلزامي: **الخطوة 1:** / **الخطوة 2:** / ... "
-    "لا شرح بدون ترقيم واضح. الطالب يجب أن يعرف أين هو في الشرح.",
-    # — الحجم المتناسب —
-    "كل خطوة تحتل 2-4 أسطر. خطوة أطول من 6 أسطر = يجب تقسيمها. "
-    "خطوة أقصر من سطر = يجب دمجها مع المجاورة.",
-    # — الانتقالات (Transitions) —
-    "استخدم كلمات الانتقال: «إذن» (استنتاج)، «ومنه» (تحويل رياضي)، "
-    "«بالتعويض» (تطبيق عددي)، «نلاحظ أن» (ملاحظة)، "
-    "«نستنتج أن» (استنتاج نهائي).",
-    # — التحقق الذاتي —
-    "في نهاية كل خطوة كبيرة: جملة تحقق قصيرة بين قوسين (مثال: «(تحقق: بتعويض x=0 نجد 0 = 0 ✓)»).",
-    # — التكيف مع نوع السؤال —
-    "سؤال الإثبات (prove): ابدأ من المعطيات وانتهِ بالمطلوب. "
-    "لا تبدأ من المطلوب وترجع للمعطيات (circular reasoning).",
-    "سؤال الحساب (calculate): ابدأ بالصيغة العامة، ثم التعويض، ثم النتيجة.",
-    "سؤال الدراسة (study): ابدأ بتحديد المجال، ثم الاشتقاق، ثم الجدول، ثم الرسم (إن طُلب).",
-    # — الخاتمة الإلزامية —
-    "كل شرح ينتهي بـ «**الخلاصة:**» — جملة واحدة تلخص النتيجة الرئيسية. "
-    "هذا يُثبِّت المعلومة في ذاكرة الطالب.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Skill Invocation Protocol (D-070)
-# ─────────────────────────────────────────────────────────────────────────────
-# بروتوكول استدعاء الـ Skills — يحدد كيف يجب أن يستدعي الـ orchestrator
-# أو الـ monolith أي Skill، وما هي الضمانات المطلوبة.
-# ─────────────────────────────────────────────────────────────────────────────
-
-SKILL_INVOCATION_PROTOCOL_VERSION: Final[str] = "1.0.0"
-
-SKILL_INVOCATION_PROTOCOL: Final[tuple[str, ...]] = (
-    # — قبل الاستدعاء (Pre-invocation) —
-    "قبل استدعاء أي Skill: تحقق من توفر المدخلات الإلزامية. "
-    "Skill مفقود مدخل إلزامي → `SkillFailure(reason='missing_input')` فوراً. "
-    "لا تستدعِ Skill بمدخلات ناقصة.",
-    # — الاستدعاء (Invocation) —
-    "الاستدعاء عبر HTTP فقط (لا import مباشر بين microservices). "
-    "كل طلب يحمل `X-Correlation-ID` للتتبع الموزع. "
-    "timeout إلزامي: ≤ 30s للـ Skills التي تستدعي LLM، ≤ 5s للـ Skills الأخرى.",
-    # — معالجة النتيجة (Result Handling) —
-    "نتيجة الـ Skill إما `SkillSuccess` أو `SkillFailure`. "
-    "لا تفترض النجاح. تحقق من النوع قبل الاستخدام.",
-    # — الـ Fallback —
-    "كل Skill يجب أن يملك fallback mode: إذا فشل الـ Skill الأساسي، "
-    "يُرجع `SkillFailure` مع `fallback_available=True` ليُعلم الـ caller "
-    "بإمكانية الاستدعاء البديل.",
-    # — القياس (Metrics) —
-    "كل استدعاء Skill يُسجَّل في Prometheus: "
-    "`cogniforge_{skill}_invocations_total{mode, status}` + "
-    "`cogniforge_{skill}_duration_seconds{mode}`. "
-    "لا Skill بدون metrics.",
-    # — العزل (Isolation) —
-    "Skill لا يستدعي Skill آخر مباشرة. كل تنسيق يمر عبر orchestrator. "
-    "هذا يمنع circular dependencies ويُبسِّط الـ debugging.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# BKT Cognitive Doctrine (D-074 / Protocol V6.0) — الطبقة المعرفية الأساسية
-# ─────────────────────────────────────────────────────────────────────────────
-BKT_COGNITIVE_DOCTRINE_VERSION: Final[str] = "3.0.0"
-
-#: قوانين غير قابلة للكسر تحكم طبقة التتبّع المعرفي البايزي (BKT) — الأساس
-#: المعرفي لكل skill تربوي مستقبلي في منصة E-TAALEEM.
-BKT_COGNITIVE_DOCTRINE: Final[tuple[str, ...]] = (
-    "BKT هي الطبقة المعرفية الأساسية لكل skill تربوي مستقبلي — أي قدرة تكيّفية "
-    "(تدرّج الصعوبة، التلميحات، مسار التعلّم) تُبنى فوق student_mastery_probability.",
-    "Zero Cognitive Overload: الجمهور طلاب بكالوريا — الرموز المجرّدة (A, B|A, Ā) "
-    "ممنوعة نهائياً في عُقَد الواجهة التوليدية.",
-    "Abstraction Ban: كل مكوّن UI توليدي يستخدم نموذج الاستخراج الهجين "
-    "(deterministic-first ثم LLM عند الحاجة) لإخراج تسميات ملموسة بشرية "
-    "(كرة حمراء، سحب ناجح، قطعة معيبة) — وحتى الـ fallback ملموس، لا حرف مجرّد.",
-    "BKT سجل append-only فقط — كل تفاعل صف جديد؛ لا تحديث في المكان. الإتقان "
-    "السابق يُقرأ من آخر صف لنفس (user_id, concept_id). يحفظ السلسلة الزمنية كاملة.",
-    "المخطط الإلزامي: concept_id + cognitive_load_estimate (low/medium/high) + "
-    "student_mastery_probability ∈ [0,1] + interaction_timestamp.",
-    "student_mastery_probability يُحدَّث بمعادلة BKT بايزية حتمية "
-    "(P_L0/P_T/P_S/P_G) — نفس المدخلات تُنتج نفس المخرجات (قابل للاختبار).",
-    "BKT لا يكسر المحادثة أبداً — كل استدعاء معزول في try/except بجلسة DB مستقلة.",
-    # D-126/D-157 (الإتقان الصادق ثنائي القناة — roadmap M6):
-    "الإتقان الصادق ثنائي القناة: نفصل الأداء المدعوم (assisted — مُضخَّم بالمساعدة) "
-    "عن الإتقان الدائم (durable). durable يُحدَّث بمنحنى نسيان متّصل (Half-Life "
-    "Regression، D-157): يتراكم بالدليل غير المدعوم المؤجَّل ويضمحلّ بالزمن — حتمي، صفر LLM.",
-    "الثابت المضاد للوهم (D-157): كسب durable ∝ generation_weight(support) ⇒ ≈0 عند "
-    "الدعم الثقيل (support=1)، فلا يُضخَّم الإتقان الدائم بالمساعدة أبداً — يبقى صادقاً.",
-    "إشارة الصواب ثلاثية الحالة (D-157): CORRECT/INCORRECT/UNKNOWN. UNKNOWN (لا دليل) "
-    "⇒ لا يُحدَّث الإتقان (يُصلِح التحيّز الهابط الذي كان يفترض الخطأ افتراضياً).",
-    "الإتقان المعروض للطالب = durable الصادق، لا المدعوم المُضخَّم. المساعدة الثقيلة "
-    "تُضخّم P(صحيح بلا معرفة) (scaffold_leak) فالإجابة المدعومة غير تشخيصية وحدها.",
-    "فجوة الوهم = assisted − durable = مقياس النجاح الوحيد (§0.6) — تُبَثّ إلى "
-    "Prometheus (M9، D-157). ممنوع التحسين على «الرضا اللحظي» أو مدّة الجلسة أو عدد "
-    "الرسائل — نُحسّن على تقليص الفجوة.",
-)
-
-# ── D-127/D-128: المعمارية الإدراكية العصبية-الرمزية — الطبقات 1 (الفهم) و 4 (السقراطي) و 5 ──
-# D-131: event_meaning يشمل خصائص الجداء (B/C/D) عبر الطبقة الدلالية العامة.
-CONCEPT_DIAGNOSIS_DOCTRINE_VERSION: Final[str] = "1.2.0"
-
-#: قوانين غير قابلة للكسر تحكم تشخيص المفهوم (الطبقة 1) + النموذج العقلي (الطبقة 5) +
-#: السرد السقراطي المُولَّد (الطبقة 4، D-128).
-CONCEPT_DIAGNOSIS_DOCTRINE: Final[tuple[str, ...]] = (
-    "الطبقة 1 (LLM) تُصنّف فقط — مخرج (concept + misconception) من enum ثابت، بلا "
-    "رياضيات/أرقام/نثر/إجابة. مجرّد تشخيص لِـ «ما الذي يحاول الطالب قوله؟».",
-    "النموذج العقلي (الطبقة 5): نخزّن *سبب* الحيرة (misconception: sample_space_confusion / "
-    "fraction_meaning_confusion / order_vs_selection_confusion) لا «أخطأ». المفهوم الخاطئ "
-    "يقود اختيار التدخّل السقراطي التالي.",
-    "كل صياغات المفهوم الواحد («ما هو البسط»/«لماذا 14»/«ما الهدف من 14»/«واش راه 14») ⇒ "
-    "مفهوم واحد (numerator) ⇒ تدخّل واحد مُوجَّه. لا قاعدة جديدة لكل صياغة.",
-    "التشخيص هجين: خريطة حتمية أولاً (سريعة/آمنة)، LLM محروس فقط عند unknown — مخرج مُتحقَّق "
-    "ضد الـ enum + timeout + fallback حتمي. أي مخرج خارج الـ enum ⇒ unknown (آمن).",
-    "الحقيقة الرياضية (الأرقام) تبقى في المحرك الرمزي حصراً (الطبقة 3) — صفر LLM. 14 تبقى 14.",
-    "لا تكرار حرفي: نفس المفهوم مرّتين ⇒ تصعيد سقراطي (شرح → سؤال → تشخيص/إعادة توجيه)، "
-    "لا نصّ متطابق. الأستاذ الحقيقي لا يُعيد الشرح للمرّة الثالثة — بل يسأل.",
-    # D-128 (الـ LLM كجهاز عصبي — السرد السقراطي المُولَّد):
-    "الـ LLM = الجهاز العصبي: يُنتج **الفهم/التربية** (السؤال السقراطي عند التصعيد)، لا "
-    "**الحقيقة** (الأرقام تُحقن من المحرك الرمزي ولا يُولّدها الـ LLM). LLM = فهم؛ symbolic = حقيقة.",
-    "السرد السقراطي المُولَّد محروس إلزامياً: عربي فقط + لا كشف جواب نهائي (14/165) + لا garbage "
-    "+ timeout + fallback حتمي. أي فشل/شكّ ⇒ القالب الحتمي (لا تدهور للطالب). يُلغي التكرار.",
-    "مفهوم event_meaning (معنى الحادثة A/B/C/D) = تعريف الحدث الملموس من نص التمرين (حقيقة)، "
-    "لا الحل الكامل. كل صياغات معنى الحادثة ⇒ event_meaning.",
-    # D-131 (الطبقة الدلالية العامة):
-    "event_meaning يشمل خصائص الجداء (B فردي / C زوجي / D معدوم) عبر الطبقة الدلالية العامة "
-    "(SemanticPropertySkill) — لا special-casing لكل حادثة. «ماذا نقصد بجداء معدوم» ⇒ تعريف "
-    "حتمي (الرقم 0)، لا «نفس اللون» الخاطئ ولا حساب.",
-)
-
-
-# ── D-131: الطبقة الدلالية + Misconception Graph — الطبقة 2 (substrate التوأم المعرفي) ──
-# D-132: جاهزية الأسئلة الجديدة عبر LLM Listener-Definer المحروس + لا default مُجمَّد.
-# D-159: الحواف (MISCONCEPTION_EDGES) + تشخيص الجذر عبر الاجتياز + مرآة knowledge_nodes/edges.
-SEMANTIC_PROPERTY_DOCTRINE_VERSION: Final[str] = "1.4.0"
-
-#: قوانين غير قابلة للكسر تحكم الطبقة الدلالية + شبكة المفاهيم الخاطئة (D-131/D-132/D-136/D-137).
-SEMANTIC_PROPERTY_DOCTRINE: Final[tuple[str, ...]] = (
-    "طبقة دلالية عامة data-driven لا special-casing: أي خاصية/اعتقاد خاطئ جديد = مدخل في "
-    "PROPERTY_REGISTRY / MISCONCEPTION_GRAPH، لا فرع if/elif جديد. مفسّر واحد لكل العبارات الرياضية.",
-    "المعمارية العصبية-الرمزية: الـ LLM = Listener فقط (لغة حرّة → property/misconception canonical "
-    "عبر مخرج enum مُقيَّد محروس + fallback حتمي) — ليس حَكماً خفياً ولا مصدراً للحقيقة. الأرقام من "
-    "المحرك الرمزي حصراً (الطبقة 3). دمج لا اختيار.",
-    "Misconception Graph «شخّص ثم تدخّل»: نفس المفهوم بـ misconceptions مختلفة ⇒ تدخّلات مختلفة. "
-    "عند الغموض يُصدَر probe تشخيصي قبل التدخّل. ممنوع تدخّل واحد موحَّد لكل من يسأل عن نفس المفهوم.",
-    "تصنيف الاعتقاد الخاطئ مبدئي صغير: rule_property | symbol_meaning | example_linking — لا اسم "
-    "جديد لكل خطأ (يمنع تضخّم عقد متشابهة). كل عقدة لها bkt_concept صريح (لا عقدة بلا أثر مقاس).",
-    "التعريف حقيقة حتمية (الطبقة 3) يُسلَّم للسياسة لا نهايةُ المسار: التعريف ثم السياسة التربوية "
-    "(D-129: تعريف→سؤال→اعتراف→تقدّم) — فتُجاب «لماذا/كيف» لا «ما معنى» فقط.",
-    "القياس قبل التوسّع: أي توسّع (مفاهيم منهج أخرى / twin دائم) مؤجَّل حتى تُثبت المقاييس السلوكية "
-    "(cogniforge_tutor_*) تحسّناً حقيقياً: تقليل التكرار، تعريف فوري، اختلاف التدخّلات، تقدّم بعد كل تفاعل.",
-    # D-132 (جاهزية الأسئلة الجديدة):
-    "جاهزية الأسئلة الجديدة دائماً: المفاهيم المعروفة تعريفها حتمي من السجلّ؛ المفهوم الجديد يُعرَّف "
-    "عبر LLM Listener-Definer محروس (تعريف مفاهيمي عام قصير، عربي، لا حساب، لا كشف نتيجة نهائية). "
-    "ممنوع default مُجمَّد لمفهوم واحد (unknown ≠ event_meaning، interpret None ≠ same_color) — المفهوم "
-    "غير المعروف يُعرَّف أو يُترك للسياسة، لا يُقحَم في الحادثة A. القياس عبر source=llm.",
-    # D-136 (مثال واعٍ بالمفهوم — قتل المثال الأعمى المتكرّر):
-    "المثال واعٍ بالمفهوم النشط لا أعمى: «اعطني مثال» ⇒ مثال **المفهوم الذي يجري الحوار عنه** "
-    "(detect_active_concept من السياق: PropertySpec.example الحتمي لكل مفهوم)، لا مثال الحادثة A "
-    "الافتراضي. ممنوع مثال/تعريف مكرّر حرفياً — المعروض already ⇒ زاوية مختلفة (LLM محروس) أو تقدّم. "
-    "ومحرّك حالة الفهم (D-135) لا يختطف أسئلة خارج مسار نفس-اللون (decide ⇒ None) فتُسلَّم للمعالج الواعي.",
-    # D-137 (فصل التوجيه — قتل تصارع النظامين):
-    "فصل صارم بين طبقتي المفهوم والحساب: أسئلة **المفهوم** (تعريف/معنى/مثال لـ A/B/C/D/X/E(X)/الشرطي/"
-    "الحالات الملائمة) ⇒ semantic_property حصراً؛ أسئلة **خطوة الحساب** (كيف/لماذا رقمٌ في حلّ P(A)) ⇒ "
-    "understanding_state (D-135) حصراً. لا تداخل، لا اختطاف.",
-    "«ما هو X» تعريف موثوق: أي سؤال نيّته definition (StudentState) أو «ما هو/ما هي/عرّف/ماذا نقصد» ⇒ "
-    "مسار التعريف **قبل** D-135. إشارات D-135 يجب أن تكون خاصة بخطوة الحساب لا كلمات مفهوم عامة — "
-    "ممنوع «الاحتمال»/«نفس اللون» المجرّدة في gap_signals (تختطف). الحيرة تُعيد إشراك المفهوم النشط "
-    "(detect_active_concept) لا تسقط للحادثة A الافتراضية؛ و_current_focus_kc لا يقفل على مثال/تعريف مفهوم.",
-    # D-159 (WP-D — الـ graph الحقيقي بالحواف):
-    "الشبكة graph حقيقي قابل للاجتياز لا dict مسطّحاً: الحواف (MISCONCEPTION_EDGES — prerequisite_of/"
-    "confused_with) تربط المفاهيم، و`diagnose_root` يجتازها ليجد **أعمق شرطٍ مسبق ضعيف** بحسب الحالة "
-    "الدائمة (tutor_state.kc_progress) ⇒ التدخّل يستهدف الجذر لا العرَض (أساس فجوة الوهم). مفهوم بلا "
-    "سجل دائم ليس «ضعيفاً» — لا تدخّل أعمى. أي حافة جديدة = مدخل data لا فرع كود.",
-    "جدولا knowledge_nodes/knowledge_edges مرآة حيّة للـ graph (scripts/sync_knowledge_graph.py عبر "
-    "جسر HTTPS + بوّابة تكافؤ code↔DB) — المسار الحي يبقى in-code حتمياً (لا I/O لقاعدة البيانات في "
-    "مسار الدور). المزامنة idempotent (UUIDv5 + ON CONFLICT DO UPDATE).",
-)
-
-
-# ── D-129: محرّك السياسة التربوية — الطبقة 4 (أقل تدخّل مفيد الآن) ──────────────────
-# D-130 (تحسينا المالك): كشف الإجابة بالفعل الكلامي لا بالطول + ميزانية تكيّفية.
-PEDAGOGICAL_POLICY_DOCTRINE_VERSION: Final[str] = "1.1.0"
-
-#: قوانين غير قابلة للكسر تحكم اختيار التدخّل التربوي (الطبقة 4، D-129/D-130).
-PEDAGOGICAL_POLICY_DOCTRINE: Final[tuple[str, ...]] = (
-    "العقد التربوي: تعريف → سؤال سقراطي محدود → اعتراف بالإجابة + تقدّم → حلّ رمزي عند الحاجة. "
-    "كل خطوة = أصغر تدخّل مفيد الآن، لا «رد» منفصل بل إدارة لرحلة الإتقان.",
-    "السقراطية محدودة بميزانية: بعد استنفادها بلا تقدّم ⇒ حلّ رمزي متدرّج (إنقاذ تربوي)، "
-    "لا سؤال إضافي. ممنوع الاستجواب اللانهائي.",
-    "الاعتراف بإجابة الطالب إلزامي: إجابة بعد سؤال ⇒ اعتراف «أحسنت» + تقدّم خطوة، لا سؤال "
-    "أصمّ آخر. الإجابة لا تُقابل بالصمت بل بالتقدير.",
-    "القرار حتمي تماماً — لا LLM في اختيار التدخّل. الـ LLM للسرد السقراطي/التقييم فقط (D-128/D-130، "
-    "محروس)؛ الحلّ الرمزي من المحرك الرمزي (D-113-safe).",
-    "القاعدة الذهبية: لا سؤال إضافي بلا معنى · لا تكرار بلا تقدّم · لا إجابة بلا سياق · "
-    "لا حلّ رمزي قبل أن تُستنفد محاولة الفهم.",
-    # D-130 تحسين A (Concern 1): كشف الإجابة بالفعل الكلامي لا بالطول.
-    "كشف «هل أجاب الطالب» بالفعل الكلامي (Speech-Act) من التاريخ لا بطول الرسالة: إجابة = "
-    "«سؤال معلّق + ليست أمراً صريحاً ولا تبديل موضوع». «نفس اللون» القصيرة والإجابة العبقرية "
-    "الطويلة كلتاهما إجابة. ممنوع ربط الإجابة بعدد الكلمات.",
-    # D-130 تحسين B (Concern 2): ميزانية تكيّفية حسب المفهوم وحالة الطالب.
-    "ميزانية الأسئلة تكيّفية لا ثابتة: socratic_budget(concept, support_level, confusion_count) "
-    "مرتبطة بالمفهوم (مفاهيمي أعمق، بصيرة سريعة أقصر) وبحالة الطالب (BKT/الإتقان: المتعثّر يُسرَّع "
-    "للسقالة، المتمكّن يحتمل عمقاً)، clamp [1,3]. ممنوع سقف تدخّل واحد ثابت لكل المفاهيم.",
-)
-
-
-# ── D-130: مُقيّم الإجابات السقراطي — الطبقة 1 العصبي-الرمزي (الإصغاء النشط) ──────────
-SOCRATIC_EVALUATOR_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-#: قوانين غير قابلة للكسر تحكم الإصغاء النشط وتقييم الرد الحرّ (الطبقة 1، D-130).
-SOCRATIC_EVALUATOR_DOCTRINE: Final[tuple[str, ...]] = (
-    "إصغاء نشط إلزامي: إذا كانت أحدث رسالة مساعد سؤالاً سقراطياً طرحناه، فرسالة الطالب التالية "
-    "= إجابة تُقيَّم قبل أي إعادة استرجاع. ممنوع منعاً باتاً إعادة طباعة التمرين ردّاً على إجابة.",
-    "اللُّبّ العصبي-الرمزي: الـ LLM يُقيّم *فهم* الطالب (يقارن رده الحرّ بالهدف التربوي المستتر) — "
-    "الطبقة 1؛ المحرك الرمزي يكتب *الخطوة* الناتجة (الأرقام 4/10/14/165) — الطبقة 3. LLM = فهم؛ symbolic = حقيقة.",
-    "التشجيع (encouragement) لا يحوي نتيجة نهائية أبداً (14/165/56/165)؛ محروس بـ "
-    "redact_final_answers + _strip_garbage_markers + is_probably_non_arabic + timeout.",
-    "التسليم الرمزي متدرّج: عند الفهم يُكشف خطوة المفهوم الحالي (مثلاً البسط: C(4,3)+C(5,3)=14) + "
-    "سؤال المتابعة، لا الحل كاملاً — يُبقي السُّلّم السقراطي حياً.",
-    "fail-open مطلق: أي تعذّر في المُقيّم ⇒ اعتراف لطيف + تلميح (لا عقاب، لا إعادة طباعة). الطالب "
-    "الذي شغّل عقله لا يُصفَع برمجياً أبداً.",
-    "حارس تبديل الموضوع (D-101) يبقى: إجابة سقراطية لا تُلتقَط إن كانت أمر تمرين جديد صريح أو موضوعاً آخر.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Adaptive Pedagogy Doctrine (D-104) — إغلاق حلقة BKT: القراءة تقود السلوك
-# ─────────────────────────────────────────────────────────────────────────────
-ADAPTIVE_PEDAGOGY_DOCTRINE_VERSION: Final[str] = "1.2.0"
-
-#: «المعيار الأعلى ليس الانبهار اللحظي، بل الاستقلال المعرفي» — E-TAALEEM.
-#: قوانين طبقة البيداغوجيا التكيفية: قراءة الإتقان تُحدِّد عمق التدريس.
-ADAPTIVE_PEDAGOGY_DOCTRINE: Final[tuple[str, ...]] = (
-    "عمق التدريس يتبع الإتقان: socratic عند mastery >= 0.7 (قُد بأسئلة، لا تكشف "
-    "الحل مباشرة)، guided في المنتصف (تلميح ثم خطوات بفجوات)، scaffolded تحت 0.35 "
-    "أو عند غياب البيانات (شرح كامل بسقالات).",
-    "يُبنى حصراً فوق student_mastery_probability من BKT (D-074) — ممنوع إعادة "
-    "اختراع تتبّع إتقان موازٍ.",
-    "fail-open مطلق: أي تعذّر (DB/تصنيف/مهلة) ينتج توجيهاً فارغاً — التوجيه "
-    "التربوي لا يكسر دور الطالب أبداً.",
-    "السياسة حتمية بعتبات ثابتة — نفس (mastery, load) تُنتج نفس المستوى دائماً؛ "
-    "تغيير العتبات يتطلب ترقية إصدار هذا الـ doctrine.",
-    "directive_text أقل من 400 حرف، عربي بسيط، بلا LaTeX وبلا رموز box-drawing "
-    "(قواعد D-067) — يُسبق في السؤال كتوجيه لا كـ system prompt.",
-    "الحِمل المعرفي المرتفع يُخفِّض درجة واحدة — «دع العقل يعمل، لكن لا تحرقه بالحمل الزائد».",
-    "التقاط المفهوم الخاطئ قبل أن يتجذر: markers حتمية لكل concept_id — التصحيح "
-    "«على مستوى الفكر لا على مستوى الحكم»، بلطف ودون لوم.",
-    "الغاية العليا الاستقلال المعرفي: المتمكن يُقاد لحلّه بنفسه — أن يصبح الطالب "
-    "أكبر من حاجته إلى الجواب.",
-    # D-113 (v1.1.0): سُلّم الدعم الخماسي — السقالة تتلاشى مع الإتقان.
-    "سُلّم الدعم 5 درجات (support_level 1..5): 1=مثال محلول كامل ← 2=إكمال "
-    "← 3=تلاشٍ عكسي ← 4=تلميحات ← 5=بلا دعم (unaided). السقالة تتلاشى مع نموّ "
-    "الإتقان (worked-example fading) — السقالة التي لا تُزال تصير قفصاً.",
-    # D-114 (v1.2.0): support_level يعبر السلسلة كاملةً ويحكم الكشف.
-    "support_level (int 1..5) يُمرَّر صراحةً عبر السلسلة كاملةً (customer_chat → "
-    "context → orchestrator_client → routes → AgentState → SynthesizerNode) "
-    "ويحكم عمق الكشف وبوّابة المثال المحلول (D-114). الافتراض الآمن عند أي "
-    "تعذّر = 5 (محجوب كلياً) — لا 1 — كي لا يكشف فشلٌ واحد كل الإجابات.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Probability Calculation Doctrine (D-075 — Protocol V14.0 / Generative UI)
-# ─────────────────────────────────────────────────────────────────────────────
-
-PROBABILITY_CALCULATION_DOCTRINE_VERSION: Final[str] = "1.5.0"
-
-#: قوانين غير قابلة للكسر تحكم محرّك حساب الاحتمالات الحتمي الذي يغذّي شجرة
-#: الاحتمالات التوليدية. الخلفية لا تُخرج HTML ولا قيماً وهمية — تحسب كسوراً
-#: حقيقية من تركيبة المسألة العربية (Protocol V14.0 §3).
-PROBABILITY_CALCULATION_DOCTRINE: Final[tuple[str, ...]] = (
-    "حساب حقيقي لا قيم وهمية: P(الحدث) = العدد / المجموع كـ كسر مختزَل دقيق "
-    "(مثل 4/11) — ممنوع منعاً باتاً إخراج 0.5 افتراضية حين تتوفّر التركيبة.",
-    "حتمي تماماً: نفس نص المسألة يُنتج نفس الكسور دائماً — لا LLM، لا عشوائية، "
-    "قابل للاختبار بـ pytest عادي.",
-    "استخراج التركيبة من العربية: أعداد رقمية وكلمات (كرتان=2، أربع=4، خمس=5) "
-    "مقرونة بالكيانات الملموسة (حمراء، بيضاء، خضراء) — لا رموز مجرّدة.",
-    "السحب بدون إرجاع يُحدِّث المقام والبسط شرطياً ((العدد-1)/(المجموع-1))؛ "
-    "السحب مع الإرجاع يُبقي الاحتمال ثابتاً — يُكتشف الوضع من نص المسألة.",
-    "كل عقدة شجرة تحمل الكسر الدقيق (p_num/p_den) إضافةً للقيمة العشرية، "
-    "فتُصيّر الواجهة 4/11 تماماً دون إعادة بناء تقريبية من العشري.",
-    "Abstraction Ban: تسميات العُقَد ملموسة بشرية مستخرَجة من السياق — وحتى "
-    "fallback المجموع المجهول لا يُخرج حرفاً مجرّداً.",
-    "العقد فصل صارم: المحرّك يُرجع Pydantic منظَّماً فقط؛ التصيير مسؤولية "
-    "الواجهة (RSC) — الخلفية لا تعرف شيئاً عن SVG أو HTML.",
-    "الموجِّه التربوي (V19.0): «دفعة واحدة» (سحب آني) = تأليفي C(n,k) → مكوّن "
-    "combinations_visualizer؛ «على التوالي» (تتابعي) → probability_tree. ممنوع "
-    "فرض شجرة تتابعية على مسألة آنية (كارثة تربوية).",
-    "جذر الشجرة بلا احتمال (لا 1/1 وهمي)، ولا كسر منحلّ (1/0) يصل للطالب أبداً.",
-    "استخراج الكيانات الملموسة (ISS-083): يقبل اسم الرقم الصريح المستقل فقط "
-    "(«بطاقة رقم 1»، «تحمل الرقم 2») ويرفض صفة «مرقمة/مرقمتان/مرقم» (تعني "
-    "«مُعلَّمة بـ» وتصف ترقيم العناصر لا نوعها). مطابقة «رقم» كسلسلة فرعية داخل "
-    "«مرقمة» كانت تُنتج كياناً زائفاً («كرة رقم 0») يُضخّم المجموع (17 بدل 11) "
-    "ويُخرج كسوراً خاطئة وشجرة تتابعية مضلِّلة لمسألة سحب آني — كارثة تربوية.",
-    "القصة التربوية الشاملة (V31.5 — Full Exercise OS): حين يعبّر الطالب عن "
-    "حيرة كاملة في مسألة سحب آني، تُولَّد سلسلة خطوات بصرية مستقلّة (Carousel) "
-    "تغطّي التمرين كاملاً — معطيات → فضاء عيّنة → حدث مركّب → متغيّر عشوائي "
-    "(توزيع فوق-هندسي بـ math.comb) — بدل مكوّن تأليفات واحد. كل خطوة تفصل "
-    "visual_directives عن numerical_state عن pedagogical_message فصلاً صارماً، "
-    "والمجموعة المستحيلة (count<k) تحمل is_possible=False ولا يتسرّب منها C_n^k=0 "
-    "ولا كسر صفري للطالب. الخطوات معمّمة من التركيبة لا مفصّلة لمسألة بعينها.",
-    "نية السؤال الحالي إلزامية (ISS-110): سياق الاحتمالات القادم من history "
-    "المحادثة وحده لا يكفي لتفعيل الأداة البصرية — السؤال الحالي نفسه يجب أن "
-    "يحمل نية احتمالية (كلمة سياق صريحة، أو إشارة حيرة، أو متابعة خطوة مثل "
-    "E(X)/جداء/نفس اللون). طلب موضوع آخر صريح («اعطني تمرين الدوال العددية») "
-    "يحجب الأداة تماماً، والاسترجاع المُفهرَس يسبق دائماً الواجهة المحسوبة في "
-    "chat_with_agent — مكوّن احتمالات مبني من كيس تمرين سابق رداً على طلب "
-    "موضوع جديد = كارثة توجيه حية.",
-    "كلمات فضاء العيّنة (نرد/عملة) تُؤخذ من سؤال الطالب ورسائله فقط — نص "
-    "المساعد في الـ history ليس دليلاً (ISS-114 · مرآة D-102): مثال توضيحي "
-    "بالنرد في إجابة سابقة كان يطغى على تركيبة الكيس الحقيقية. والتركيبة "
-    "المستخرَجة من نص التمرين تسبق فضاء العيّنة العام في خط الاستراتيجيات.",
-)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Impossible Case UX Doctrine (D-082)
-# ─────────────────────────────────────────────────────────────────────────────
-# قواعد تجربة الحالة المستحيلة — تُطبَّق عند requested > bag_count.
-# المُستفيدون: content-retrieval-skill /impossible-case + frontend renderer.
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: نسخة doctrine الحالة المستحيلة.
-IMPOSSIBLE_CASE_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-IMPOSSIBLE_CASE_DOCTRINE: Final[tuple[str, ...]] = (
-    "لا تبدأ بالرياضيات: المرحلة 1 و2 و3 خالية تماماً من المعادلات والصيغ.",
-    "مشهد واحد في كل مرحلة: لا تعرض أكثر من فكرة واحدة في أي شاشة.",
-    "لا شجرة احتمالات في الحالة المستحيلة: الشجرة تظهر فقط للحالات الممكنة.",
-    "السؤال الحدسي (المرحلة 2) يجب أن يكون مباشراً وبسيطاً — لا معادلة، لا صفر.",
-    "الحركة البصرية (المرحلة 3) قصيرة وهادئة — shake ≤ 600ms، لا تكرار.",
-    "الرسالة التربوية (المرحلة 4) واحدة فقط — واضحة ولطيفة، تذكر الأعداد الفعلية.",
-    "deep_dive اختياري: يُعرض فقط عند ضغط المستخدم على 'اشرح لي أكثر'.",
-    "backend يُرسل contract بصري فقط (حالة + بيانات + رسالة) — لا HTML، لا SVG.",
-    "frontend مسؤول عن تحويل الـ contract إلى قصة بصرية — لا خلط منطق بحركة.",
-    "الشرح الرياضي العميق (deep_dive) يحتوي: formula + explanation + why_zero.",
-)
-
-
-def get_impossible_case_summary() -> str:
-    """يُرجِع ملخصاً موجزاً لـ doctrine الحالة المستحيلة."""
-    return (
-        f"[v{IMPOSSIBLE_CASE_DOCTRINE_VERSION}] "
-        f"{len(IMPOSSIBLE_CASE_DOCTRINE)} قاعدة — "
-        "4 مراحل بصرية، لا رياضيات قبل الفهم الحدسي، deep_dive اختياري."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Realtime Protocol Doctrine — D-WS-FLAP-002 / ISS-WS-FLAP-002 (2026-05-26)
-# ─────────────────────────────────────────────────────────────────────────────
-# هذه القواعد تحكم سلوك الـ WebSocket message loop عبر كل WS endpoints في النظام.
-# الكارثة المُشخَّصة: الواجهة ترسل `{type:"ping"}` كل 25 ثانية كـ app-level
-# heartbeat، لكن الـ WS routers (customer_chat, admin, conversation_service)
-# كانت تعالج كل رسالة واردة كـ «سؤال»، تُرجع خطأ «Question is required»،
-# ولا تُرسل `{type:"pong"}` → timeout بعد 10s → close(1001) → reconnect → flap.
-#
-# المُستفيدون:
-# - `app.services.skills.ws_heartbeat_skill.handle_control_message`
-# - `app.api.routers.customer_chat.chat_stream_ws` (WS loop)
-# - `app.api.routers.admin.admin_chat_stream_ws` (WS loop)
-# - `microservices.conversation_service.main` (WS endpoints)
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: نسخة doctrine بروتوكول الـ realtime.
-REALTIME_PROTOCOL_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-REALTIME_PROTOCOL_DOCTRINE: Final[tuple[str, ...]] = (
-    "Control messages first: أي WS endpoint يستقبل رسالة بحقل `type` يجب أن "
-    "يفحص أولاً ما إذا كانت رسالة تحكم (ping/heartbeat/noop) قبل معالجتها كسؤال.",
-    'Ping ⇒ Pong: الرسالة `{"type":"ping"}` يجب أن تُرَدّ عليها بـ '
-    '`{"type":"pong","ts":<iso-utc>}` فوراً بدون أي معالجة إضافية.',
-    'Heartbeat ⇒ Ack: الرسالة `{"type":"heartbeat"}` يجب أن تُرَدّ عليها بـ '
-    '`{"type":"heartbeat_ack","ts":<iso-utc>}`.',
-    'Noop ⇒ Silent: الرسالة `{"type":"noop"}` تُبتلَع بصمت بدون رد (لا flooding).',
-    "Heartbeat هو Skill، ليس inline code: يجب استخدام "
-    "`app.services.skills.ws_heartbeat_skill.handle_control_message` — "
-    "ممنوع نسخ منطق ping/pong في كل router (Single Source of Truth).",
-    "Fail-open: إذا فشل إرسال pong (مثل client disconnected بين receive و send)، "
-    "يُسجَّل DEBUG ولا يُكسَر loop — receive_json التالي سيكشف الإغلاق.",
-    'Correlation: لو أرسل العميل `{"type":"ping","id":"..."}`، يجب أن '
-    "يُضمَّن نفس الـ `id` في الـ pong لتمكين tracing.",
-    "Backend-initiated heartbeat ليس بديلاً عن Application-level ping: uvicorn "
-    "`--ws-ping-interval` يفحص الـ TCP layer فقط — لا يكشف إذا كان الـ handler "
-    "عالق. الـ app-level heartbeat هو نظام liveness حقيقي للطبقة العليا.",
-)
-
-
-def get_realtime_protocol_summary() -> str:
-    """يُرجِع ملخصاً موجزاً لـ doctrine بروتوكول الـ realtime (للـ logs/system prompts)."""
-    return (
-        f"[v{REALTIME_PROTOCOL_DOCTRINE_VERSION}] "
-        f"{len(REALTIME_PROTOCOL_DOCTRINE)} قاعدة — "
-        "ping/pong + heartbeat/ack موحَّد عبر Skill، fail-open، correlation IDs."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Skills Platform Doctrine (D-100 — Registry + Composition + Observability)
-# ─────────────────────────────────────────────────────────────────────────────
-#: نسخة doctrine منصّة الـ Skills الموحَّدة.
-SKILLS_PLATFORM_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-SKILLS_PLATFORM_DOCTRINE: Final[tuple[str, ...]] = (
-    "Single registry: كل Skill يُسجَّل في `registry.get_skill_registry()` ببيانات "
-    "وصفية موحَّدة (name, contract, primary_method, consumed_by, status).",
-    "No ZOMBIE: كل Skill مُسجَّل يجب أن يملك مُستهلِكاً حيّاً في `consumed_by` — "
-    "الـ CI gate يحرس على ذلك (مرآة D-073).",
-    "Additive only: المنصّة لا تُعيد توصيل المسار الحيّ — تُوفِّر صياغة رسمية "
-    "(compose_text_refinement) + سطح رصد (/api/v1/skills) دون لمس الدردشة الحيّة.",
-    "Graceful degradation: كل خطوة في `compose_text_refinement` معزولة — فشلها "
-    "يُتجاهَل ويُحتفَظ بالنص (لا يكسر المسار أبداً).",
-    "Flagged dormants: الـ Skills المُفعَّلة من قدرات DORMANT (retrieval_rerank, "
-    "mcp_tool) مُعطَّلة افتراضياً وتُرجِع None عند التعطيل (صفر تغيير سلوكي).",
-    "Metric-emitter contract (§6.21): كل مقياس يُعرَض في المنصّة له مُصدِر حقيقي "
-    "في الكود — لا zombie metrics.",
-)
-
-
-def get_skills_platform_summary() -> str:
-    """يُرجِع ملخصاً موجزاً لـ doctrine منصّة الـ Skills (للـ logs/system prompts)."""
-    return (
-        f"[v{SKILLS_PLATFORM_DOCTRINE_VERSION}] "
-        f"{len(SKILLS_PLATFORM_DOCTRINE)} قاعدة — registry موحَّد، no-ZOMBIE، "
-        "additive، graceful degradation، flagged dormants، metric-emitter contract."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Content Integrity Doctrine (D-106 · ISS-114)
-# ─────────────────────────────────────────────────────────────────────────────
-# حارس نزاهة المحتوى المواجه للطالب — يلتقط الغارباج اللاتيني وتسريب HTML على
-# كامل بثّ الإجابة (لا أول نافذة فقط). المُستهلِك: orchestrator_client (مخرج
-# الـ orchestrator HTTP) + local_graph (المسار المحلي).
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: نسخة doctrine نزاهة المحتوى.
-CONTENT_INTEGRITY_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-CONTENT_INTEGRITY_DOCTRINE: Final[tuple[str, ...]] = (
-    "كل بثّ مواجه للطالب يمرّ عبر StreamIntegrityFilter على **كامل التيار** — "
-    "لا أول 200 حرف فقط (ثغرة arabic_stream_guard التاريخية).",
-    "الثغرة الأكبر مُغلقة: بثّ orchestrator-service (MODE_B) عبر HTTP لم يكن "
-    "يمرّ بأي حارس — الآن يُلفّ بالمرشّح عند مخرج المونوليث (نقطة الاختناق "
-    "الوحيدة المواجهة للطالب)؛ response_sanitizer للخدمة يبقى دفاعاً ثانوياً.",
-    "وضع اللغة العربي (نسبة ≥ 0.5 في أول نافذة) يُفعِّل فلترة اللاتيني؛ الإجابة "
-    "الفرنسية/Darija المشروعة تمرّ بلا حذف (تنظيف HTML فقط) — لا إفراط حذف.",
-    "spans الرياضيات (LaTeX: $...$, $$...$$, \\(...\\), \\[...\\]) تُبثّ حرفياً "
-    "أبداً — صفر مساس (قوانين D-051/D-062).",
-    "الـ allowlist التقنية محافظة (مفردات رياضية + BAC فرنسية، مقارنة بعد "
-    "NFD-strip): توسيعها = ترقية doctrine. كل ما عداها من لاتيني >2 حرف يُحذف.",
-    "fail-open مطلق: أي استثناء في المرشّح ⇒ تعطيل دائم وإرجاع النص الخام — "
-    "حارس النزاهة لا يكسر دور الطالب أبداً.",
-    "HTML/JSX لا يصل نص الطالب أبداً — الواجهات حصراً عبر قناة ui_component "
-    "المُهيكلة (Channel A — D-086)، لا عبر نص يكتبه الـ LLM.",
-)
-
-
-def get_content_integrity_summary() -> str:
-    """يُرجِع ملخص doctrine نزاهة المحتوى (للـ logs / system prompts)."""
-    return (
-        f"[v{CONTENT_INTEGRITY_DOCTRINE_VERSION}] "
-        f"{len(CONTENT_INTEGRITY_DOCTRINE)} قاعدة — فلترة كامل البثّ، حماية "
-        "اللاتيني التقني، LaTeX حرفي، fail-open، HTML ممنوع في نص الطالب."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Learning Path Doctrine (D-111 · المسار التعلّمي التكيفي)
-# ─────────────────────────────────────────────────────────────────────────────
-# القفزة التربوية فوق BKT (D-074): يقترح الخطوة التالية وصعوبة متكيفة من الإتقان.
-# «المعيار الأعلى ليس الانبهار اللحظي، بل الاستقلال المعرفي» — فلسفة E-TAALEEM.
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: نسخة doctrine المسار التعلّمي.
-LEARNING_PATH_DOCTRINE_VERSION: Final[str] = "1.1.0"
-
-LEARNING_PATH_DOCTRINE: Final[tuple[str, ...]] = (
-    "يُبنى حصراً فوق student_mastery_probability (D-074) — لا يُعيد اختراع تتبّع "
-    "الإتقان؛ يستهلك BKT_COGNITIVE_DOCTRINE و ADAPTIVE_PEDAGOGY_DOCTRINE.",
-    "حتمي تماماً: نفس (المفهوم، الإتقان) ⇒ نفس التوصية — لا LLM، لا I/O داخل "
-    "الـ Skill (الإتقان يُمرَّر من المُستدعي) — قابل لاختبار pytest.",
-    "إتقان ≥ 0.7 ⇒ تقدّم للمفهوم التالي في تسلسل منهج BAC + صعوبة hard (الطالب جاهز للتحدّي).",
-    "0.35 ≤ إتقان < 0.7 ⇒ ابقَ على المفهوم + صعوبة medium (ترسيخ قبل التقدّم).",
-    "إتقان < 0.35 أو مجهول ⇒ ابقَ على المفهوم + صعوبة easy (ثبّت الأساس أولاً).",
-    "التسلسل خريطة منهج ثابتة (لا overfitting لمسألة بعينها)؛ المفهوم الأخير في "
-    "السلسلة يبقى عليه عند الإتقان (لا يخترع مفهوماً غير موجود).",
-    "fail-open مطلق: أي تعذّر ⇒ توصية محايدة آمنة — لا يكسر دور الطالب.",
-    "D-119: التوصية تُشتقّ وتُسجَّل خلف الكواليس (telemetry / لوحة المعلم) — "
-    "لا تُعرَض كبطاقة للطالب (سطح الطالب = تعليم نظيف فقط، لا تتبّع مُكرَّر).",
-)
-
-
-def get_learning_path_summary() -> str:
-    """يُرجِع ملخص doctrine المسار التعلّمي (للـ logs)."""
-    return (
-        f"[v{LEARNING_PATH_DOCTRINE_VERSION}] "
-        f"{len(LEARNING_PATH_DOCTRINE)} قاعدة — فوق BKT، حتمي، 3 عتبات إتقان، "
-        "تسلسل منهج ثابت، fail-open."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Mandatory Orchestration Doctrine (D-112 · العمود الفقري الإلزامي)
-# ─────────────────────────────────────────────────────────────────────────────
-# قرار المستخدم (2026-06-13): الخدمات المصغرة + الرسم الـ13-node هي القلب
-# الإلزامي الوحيد. عند تعذّرها ⇒ خطأ صريح، لا fallback صامت إلى local_graph.
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: نسخة doctrine العمود الفقري الإلزامي.
-MANDATORY_ORCHESTRATION_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-MANDATORY_ORCHESTRATION_DOCTRINE: Final[tuple[str, ...]] = (
-    "الخدمات المصغرة + الرسم الـ13-node هي القلب الإلزامي الوحيد للتوليد — "
-    "لا نتظاهر أن الرسم المحلي ذا العقدتين بديل عنها (runtime truth over synthetic certainty).",
-    "عند تعذّر الـ orchestrator (تعذّر اتصال/empty_stream) ⇒ إطار error صريح بالرمز "
-    "ORCHESTRATOR_REQUIRED ورسالة عربية واضحة — صفر سقوط صامت إلى local_graph.",
-    "علم REQUIRE_ORCHESTRATOR=1 افتراضي مُفعَّل؛ =0 يُعيد الـ fallback القديم "
-    "(rollback فوري بلا deploy — نمط D-025).",
-    "المقايضة الصريحة: فقدان المرونة عمداً — سقوط الـ orchestrator في الإنتاج = "
-    "انقطاع كامل (لا إجابة ضعيفة). القرار يخصّ المالك ومُوثَّق.",
-    "الموصول حيّاً: orchestrator_client.chat_with_agent بعد فشل كل المرشّحين.",
-)
-
-
-def get_mandatory_orchestration_summary() -> str:
-    """يُرجِع ملخص doctrine العمود الفقري الإلزامي (للـ logs)."""
-    return (
-        f"[v{MANDATORY_ORCHESTRATION_DOCTRINE_VERSION}] "
-        f"{len(MANDATORY_ORCHESTRATION_DOCTRINE)} قاعدة — microservices-only، "
-        "hard-fail ORCHESTRATOR_REQUIRED، علم rollback، مقايضة مرونة صريحة."
-    )
-
-
-# ── D-133: حالة الطالب كإشارة قرار — النيّة + الإحباط (إعادة توجيه المالك) ─────────────
-STUDENT_STATE_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-#: قوانين غير قابلة للكسر تحكم قراءة حالة الطالب (النيّة + الإحباط، D-133).
-STUDENT_STATE_DOCTRINE: Final[tuple[str, ...]] = (
-    "إشارة قرار لا قاموس labels: كل intent يجب أن يُنتج **نوع رد مغايراً** فعلاً (تعريف/مثال/"
-    "خطوات/علاقة/تعريف+مثال+سؤال/تلميح). label بلا تغيير سياسة = فشل. القيمة في البيداغوجيا المغايرة.",
-    "متعدّد الإشارات لا تصنيف جامد: الخرج primary_intent + secondary_signals — «لم أفهم X» = "
-    "confusion + (definition) — تحفظ الحزمة النفسية/اللغوية كاملة، لا تقسيمها لنيّة واحدة صارمة.",
-    "النيّة تُحدّد نوع الرد لا الكلمة: «لم أفهم X» = confusion (تعريف + مثال ملموس + سؤال موجِّه "
-    "واحد)، لا تعريف-فقط ولا سقراطية متكررة؛ «أعطني مثالاً» = مثال قبل النظرية؛ «كيف نحسب» = خطوات.",
-    "الإحباط حالة شعورية تُوقف التكرار: «لم أفهم»×كثير = frustration عالٍ ⇒ تقليل الأسئلة + كشف "
-    "خطوة رمزية أقرب (socratic_budget→0)، ممنوع سقراطية متكررة على طالب محبَط.",
-    "الإحباط لحظي يتلاشى لا هوية ثابتة (نقد المالك 3): يُحسب من نافذة حديثة، يتعافى فور رسالة "
-    "غير-حيرة، يُغيّر السياسة لهذا الدور فقط، **لا يُخزَّن** كصفة طالب. لا I/O، لا حالة دائمة.",
-    "LLM = مُفسِّر ثانوي لا حَكَم (نقد المالك 2): الكواشف الحتمية + BKT (knowledge tracing) هي "
-    "السلطة؛ الـ LLM يُستدعى **فقط** عند primary=unknown (صياغة جديدة) ولا يطغى على إشارة حتمية.",
-    "حتمي-أولاً + fail-open: markers أولاً ثم LLM للصياغات الجديدة؛ أي تعذّر ⇒ حالة آمنة "
-    "(definition/none) — قراءة الحالة لا تكسر دور الطالب أبداً.",
-)
-
-
-def get_student_state_summary() -> str:
-    """يُرجِع ملخص doctrine حالة الطالب (للـ logs)."""
-    return (
-        f"[v{STUDENT_STATE_DOCTRINE_VERSION}] {len(STUDENT_STATE_DOCTRINE)} قاعدة — "
-        "إشارة قرار، متعدّد-إشارات، LLM مُفسِّر ثانوي، إحباط لحظي يتلاشى."
-    )
-
-
-# ── D-142 (Phase 2): مدير الحوار الموحَّد + الحالة الدائمة ─────────────────────────────
-DIALOGUE_MANAGER_DOCTRINE_VERSION: Final[str] = "1.1.0"
-
-#: قوانين غير قابلة للكسر تحكم سلطة قرار الدور التعليمي (D-142 Phase 2).
-DIALOGUE_MANAGER_DOCTRINE: Final[tuple[str, ...]] = (
-    "قرار الدور = f(دليل الفهم، قدرة الطالب، صعوبة الخطوة التالية) كل دور — لا «مصفوفة levels» "
-    "جامدة (نقد المالك). الإشارات الثلاث تُحسَب حيّاً، لا «هل وصلنا level 3».",
-    "التقدّم حتمي عند دليل فهم موثوق (verified رمزياً أو understood): اعتراف + خطوة تالية، "
-    "ممنوع إعادة السؤال نفسه.",
-    "لا تكرار حرفي: النصّ المُرشَّح المُكرّر لـ last_step_emitted (مرساة دائمة تنجو من نافذة "
-    "التاريخ) ⇒ تصعيد للحلّ الرمزي، لا إعادة.",
-    "لا قفز فوق قدرة الطالب: قدرة < صعوبة الخطوة التالية ⇒ خطوة وسطى أصغر، لا تمثيل أعلى من "
-    "مستوى المتعلّم.",
-    "الميزانية السقراطية **لكل مفهوم** (من الحالة الدائمة) لا عامّة تتسرّب بين المفاهيم؛ استنفادها "
-    "⇒ حلّ رمزي لا سؤال إضافي.",
-    "حتمي 100% (صفر LLM) قابل للاختبار بـ pytest؛ الأرقام/الصحّة من المحرك الرمزي؛ القدرة من BKT "
-    "(D-126). fail-open: أي خطأ ⇒ السلوك القائم.",
-    "بوّابة القياس (§2D): التوسّع مرهون بهبوط repetition_rate→0 + advance_after_correct≈100% + "
-    "over-jump→0 + فجوة الوهم (D-126) ⤵ — لا توسيع للحالة قبل إثبات اللِّفت.",
-    "D-158: قرار الدور يقرأ tutor_state.kc_progress الدائم (representations_delivered / _pending) "
-    "لا مسح نصّ السجل؛ خطوة موجودة في representations_delivered لا تُعاد أبداً (منع تكرار عبر الكتل). "
-    "طبقة _cognitive_turn خلف COGNITIVE_TURN_ENABLED هي المستهلِك الجديد فوق tutor_state المُخزَّن.",
-)
-
-
-def get_dialogue_manager_summary() -> str:
-    """يُرجِع ملخص doctrine مدير الحوار (للـ logs)."""
-    return (
-        f"[v{DIALOGUE_MANAGER_DOCTRINE_VERSION}] {len(DIALOGUE_MANAGER_DOCTRINE)} قاعدة — "
-        "evidence×ability×difficulty، تقدّم حتمي، لا تكرار، لا قفز، ميزانية لكل مفهوم."
-    )
-
-
-# ── D-135: محرّك حالة الفهم — Learning State («ماذا فهم الطالب فعلاً؟») ────────────────
-UNDERSTANDING_STATE_DOCTRINE_VERSION: Final[str] = "1.2.0"
-
-#: قوانين غير قابلة للكسر تحكم تتبّع حالة الفهم (Learning State، D-135).
-UNDERSTANDING_STATE_DOCTRINE: Final[tuple[str, ...]] = (
-    "Learning State هو المحور: القرار يُبنى على «ماذا فُهم وما بقي غير مفهوم» (حالة المكوّنات "
-    "المعرفية)، لا على آخر رسالة. السؤال يتحوّل من «ما المفهوم؟» إلى «ماذا فهم الطالب فعلاً؟».",
-    "KC-aware لا number-match (نقد المالك 1): الفجوة تُشخَّص بالمعنى (gap_signals)؛ المكوّنات المعرفية "
-    "مُشتقّة من المحرك الرمزي (combo)؛ الرقم إشارة واحدة لا مستخرَج. «لماذا قسمنا على 3!» (بلا رقم) ⇒ kc_factorial.",
-    "التكرار دلالي لا حرفي (نقد المالك 2): الحالة على مستوى المكوّن (شُرح/فُهم)، لا توقيع نصّي. مكوّن "
-    "مفهوم لا يُعاد شرحه أبداً؛ «14/165» ≡ «14 من 165» ≡ «عدد الحالات 14» = نفس المكوّن.",
-    "الفهم ببرهان لا صحة (نقد المالك 3): understood يتطلّب understanding_evidence (إظهار مرتبط بالمكوّن)، "
-    "لا مجرّد correct_answer. «فهمت» المجرّدة = إقرار ضعيف لا برهان — لا تُرقّي المكوّن إلى مفهوم.",
-    "استباقي لا تفاعلي (نقد المالك 4): مكوّن شُرح بلا فهم ⇒ تمثيل أعلى تلقائياً (شرح→مثال→تشبيه→سؤال "
-    "تطبيقي) قبل شكوى الطالب «انت تكرر» — Adaptive Learning.",
-    "الاعتراف والتقدّم أساسي: برهان فهم ⇒ «أحسنت ✅» + التقدّم لأوّل مكوّن غير مفهوم (الجبهة التعليمية)، "
-    "تغيير الاستراتيجية لا مزيد كلام.",
-    "الأرقام من المحرك الرمزي حصراً (ProbabilityCalculatorSkill/combo)؛ صفر LLM في الحقيقة؛ كل التمثيلات "
-    "حتمية وتَنجو من حجب D-113. الـ LLM (سؤال موجِّه/تقييم) محروس فقط (D-128/D-130).",
-    "D-159 (WP-C — تقاعد مسح النصّ): الحالة الدائمة (tutor_state.kc_progress) هي **سلطة** حالة المكوّنات "
-    "(understood/explained) — مسح نصّ المحادثة fallback للأدوار السابقة على التخزين فقط. الترقية أحادية "
-    "الاتجاه (الدائم يرفع ولا يُنزِل ما أثبته النصّ)، وقرار المحرّك (understood_kc_id + المكوّن المشروح) "
-    "يُكتب فوراً في kc_progress عبر دلتا record_turn — الذاكرة تتراكم لا تُعاد هندستها كل دور.",
-    "D-165 (ISS-129 — غاية التمرين = مكوّناته المعرفية): سؤال الغاية meta («ماذا نستفيد/نتعلم من هذا "
-    "التمرين؟») يُجاب حتمياً من عناوين الـ KCs (exercise_purpose_summary — data-driven من combo، عناوين "
-    "فقط بلا أي نسبة نهائية) ولا يُختطف بالـ probe التشخيصي أبداً؛ وأي سؤال غير-إجرائي لم تُجبه الطبقات "
-    "الحتمية يهرب لطبقات الإجابة (D-112) **قبل** الـ probe — لا سؤال يُقابَل بسؤال يتجاهله.",
-)
-
-
-def get_understanding_state_summary() -> str:
-    """يُرجِع ملخص doctrine حالة الفهم (للـ logs)."""
-    return (
-        f"[v{UNDERSTANDING_STATE_DOCTRINE_VERSION}] {len(UNDERSTANDING_STATE_DOCTRINE)} قاعدة — "
-        "Learning State، KC-aware، تكرار دلالي، فهم ببرهان، تمثيل استباقي."
-    )
-
-
-# ── D-138: المصفوفة التصعيدية التكيّفية — Escalation + Understanding-Signal + Misconception ──
-# ── D-147: نفاد السُّلّم ⇒ سؤال تطبيق مرتبط بالتمرين بدل الـ punt العام ──
-# ── D-159 (WP-B): FSM دائم — الرُّتب المُسلَّمة من tutor_state.kc_progress لا مسح النصّ وحده ──
-PEDAGOGICAL_ESCALATION_DOCTRINE_VERSION: Final[str] = "1.2.0"
-
-#: قوانين غير قابلة للكسر تحكم المصفوفة التصعيدية التكيّفية (D-138).
-PEDAGOGICAL_ESCALATION_DOCTRINE: Final[tuple[str, ...]] = (
-    "السقالة تُختار بسؤال واحد قبل كل تدخّل (حكم المالك): «هل تناسب قدرة الطالب الحالية "
-    "(support_level/BKT) وصعوبة الخطوة القادمة؟» — لا تُمرَّر كل الحالات عبر سلّم جامد L1/L2/L3. "
-    "نفس النيّة + طالب مختلف ⇒ رُتبة مختلفة (مرونة لا جمود).",
-    "التصعيد يعتمد على أثر الفهم (Understanding Signal) لا عدّ «لم أفهم»: دليل فهم المفهوم "
-    "(evidence_markers — آلية لا تقليد عبارة) ⇒ action=mastered (توقّف + اعتراف). القيمة في تتبّع "
-    "الإتقان لا تصنيف إجابة واحدة.",
-    "Misconception Check قبل التصعيد (تشخيص دقيق للاستدلال): عقدة مُشخَّصة (diagnose_misconception) ⇒ "
-    "action=target_misconception (intervention مُوجَّه) بدل تسلّق السلّم أعمى. كل عقدة لها bkt_concept صريح.",
-    "Ability + Step-Difficulty Calibration: support_level (1..5 من BKT) يُعايِر الرُّتبة — قدرة منخفضة "
-    "(1-2) ⇒ تصعيد أسرع لسقالة أعلى (محاكاة مصغّرة)؛ قدرة عالية (4-5) ⇒ رُتبة أخف (تعريف + سؤال موجِّه، "
-    "student agency) لا إفراط في السقالة.",
-    "ممنوع تكرار رُتبة استراتيجية: ذاكرة التصعيد (stateless من history) تمنع إعادة نفس المستوى؛ الحيرة ⇒ "
-    "المستوى التالي غير المُسلَّم المُعايَر بالقدرة؛ استنفاد L3 ⇒ exhausted (تسليم لطيف، لا تكرار، لا L4 الآن).",
-    "concept-scoped حصراً: تعمل على المفهوم النشط (detect_active_concept)، لا على مكوّنات تمرين البكالوريا "
-    "(لا انحراف لـ event-A/نفس اللون). سُلّم الاستراتيجية: L1_DEFINITION → L2_ANALOGY → L3_MICRO_SIMULATION "
-    "(L4_VISUAL مؤجَّل — لا واجهة توليدية في القلب الآن).",
-    "policy أولاً، content-server ثانياً: المصفوفة تقرّر، MicroSimulationSkill يخدم L3. نقيس الأثر السلوكي "
-    "فقط (تراجع التكرار + تدرّج الرُّتب + ارتفاع الإتقان BKT) — لا تحسين على «الرضا اللحظي».",
-    "D-147: نفاد السُّلّم (exhausted) ⇒ سؤال «خطوة تطبيق» مرتبط ببنية هذا التمرين (APPLY_STEPS، concept-scoped) "
-    "يقود الطالب للخطوة الأولى الملموسة — لا punt عامّ («أيّ خطوة تريد؟»). يبقى سؤالاً (توليد لا كشف) ⇒ يَنجو "
-    "من حجب D-113؛ المفهوم غير المُغطّى يرجع للتسليم العامّ القديم (لا انحدار).",
-    "D-159 (WP-B — FSM دائم): ذاكرة التصعيد سلطتها الحالة الدائمة (delivered_levels من "
-    "tutor_state.kc_progress.escalation_levels — تَنجو من نافذة الـ50 رسالة وإعادة تشغيل العملية)، "
-    "ومسح نصّ المساعد شبكة أمان للأدوار السابقة على التخزين — **الاتحاد** يمنع تكرار الرُّتبة من أي "
-    "مصدر. كل رُتبة تُبثّ (action=teach) تُكتب فوراً في kc_progress عبر دلتا record_turn.",
-)
-
-
-def get_pedagogical_escalation_summary() -> str:
-    """يُرجِع ملخص doctrine المصفوفة التصعيدية (للـ logs)."""
-    return (
-        f"[v{PEDAGOGICAL_ESCALATION_DOCTRINE_VERSION}] {len(PEDAGOGICAL_ESCALATION_DOCTRINE)} قاعدة — "
-        "Escalation + Understanding-Signal + Misconception-Check + ability calibration."
-    )
-
-
-# ── D-138: محرّك المحاكيات المصغّرة الحتمي — خادم محتوى L3 ────────────────────────────
-MICRO_SIMULATION_DOCTRINE_VERSION: Final[str] = "1.0.0"
-
-#: قوانين غير قابلة للكسر تحكم محرّك المحاكيات المصغّرة (D-138).
-MICRO_SIMULATION_DOCTRINE: Final[tuple[str, ...]] = (
-    "حتمي 100% صفر-LLM (نقد MathDial): الأمثلة ثابتة في كود، لا يُولّدها نموذج (النماذج تكشف الحل "
-    "مبكراً أو تُعطي feedback غير دقيق). محرّك المحتوى لا يخترع الأرقام.",
-    "قصير جداً isomorphic لا شرح جديد كامل: ≤320 حرف (≤3 جُمل)، مُصمَّم للتشابه البنيوي مع المفهوم — "
-    "أرقام صغيرة آمنة خاصة به (لا أرقام التمرين 11/165) فيَنجو من حجب D-113 ولا يكشف جواب التمرين.",
-    "خادم L3 للمصفوفة لا مسار مستقلّ: يُستدعى فقط حين تقرّر المصفوفة التصعيدية أن الوقت مناسب "
-    "لمحاكاة مصغّرة (قدرة منخفضة / طلب مثال عددي / استنفاد الرُّتب الأخفّ). مفهوم بلا محاكاة ⇒ None.",
-)
-
-
-def get_micro_simulation_summary() -> str:
-    """يُرجِع ملخص doctrine المحاكيات المصغّرة (للـ logs)."""
-    return (
-        f"[v{MICRO_SIMULATION_DOCTRINE_VERSION}] {len(MICRO_SIMULATION_DOCTRINE)} قاعدة — "
-        "deterministic، isomorphic، ≤320 حرف، خادم L3."
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Skill Doctrine Manifest (للـ CI gate)
-# ─────────────────────────────────────────────────────────────────────────────
-
-#: All-in-one manifest يُصدِّره الـ Skill module للـ CI gate لفحص الـ drift.
-SKILL_DOCTRINE_MANIFEST: Final[dict[str, dict[str, object]]] = {
-    "retrieval": {
-        "version": RETRIEVAL_DOCTRINE_VERSION,
-        "rules_count": len(RETRIEVAL_DOCTRINE),
-        "consumed_by": (
-            "BACExerciseSkill._retrieve",
-            "orchestrator_client._stream_local_retrieval_response",
-            "exercise_retrieval.detect_exercise_retrieval",
-        ),
-    },
-    "explanation": {
-        "version": EXPLANATION_DOCTRINE_VERSION,
-        "rules_count": len(EXPLANATION_DOCTRINE),
-        "consumed_by": (
-            "BACExerciseSkill._explain",
-            "local_graph._EXERCISE_EXPLANATION_SYSTEM_PROMPT",
-            "BACSkillExplanationOutput.methodology_handle",
-        ),
-    },
-    "model_answer_reliance": {
-        "version": MODEL_ANSWER_RELIANCE_VERSION,
-        "rules_count": len(MODEL_ANSWER_RELIANCE_RULES),
-        "consumed_by": (
-            "BACExerciseSkill._explain",
-            "local_graph.run_local_graph_with_exercise_context",
-        ),
-    },
-    "detailed_explanation": {
-        "version": DETAILED_EXPLANATION_VERSION,
-        "rules_count": len(DETAILED_EXPLANATION_RULES),
-        "consumed_by": ("local_graph._classify_question_budget",),
-    },
-    "content_invocation": {
-        "version": CONTENT_INVOCATION_DOCTRINE_VERSION,
-        "rules_count": len(CONTENT_INVOCATION_DOCTRINE),
-        "consumed_by": (
-            "BACExerciseSkill._retrieve",
-            "local_graph.run_local_graph_with_exercise_context",
-            "orchestrator_client._stream_local_retrieval_response",
-        ),
-    },
-    "model_answer_explanation": {
-        "version": MODEL_ANSWER_EXPLANATION_VERSION,
-        "rules_count": len(MODEL_ANSWER_EXPLANATION_DOCTRINE),
-        "consumed_by": (
-            "BACExerciseSkill._explain",
-            "local_graph._EXERCISE_EXPLANATION_SYSTEM_PROMPT",
-        ),
-    },
-    "step_by_step_explanation": {
-        "version": STEP_BY_STEP_EXPLANATION_VERSION,
-        "rules_count": len(STEP_BY_STEP_EXPLANATION_RULES),
-        "consumed_by": (
-            "local_graph._classify_question_budget",
-            "BACExerciseSkill._explain",
-        ),
-    },
-    "skill_invocation_protocol": {
-        "version": SKILL_INVOCATION_PROTOCOL_VERSION,
-        "rules_count": len(SKILL_INVOCATION_PROTOCOL),
-        "consumed_by": (
-            "orchestrator_client.chat_with_agent",
-            "local_graph.run_local_graph",
-        ),
-    },
-    "answer_quality": {
-        "version": "1.0.0",
-        "rules_count": 6,
-        "consumed_by": (
-            "AnswerQualitySkill.evaluate",
-            # D-073 (2026-05-19): wired into the live monolith path.
-            "local_graph._apply_answer_quality_skill",
-        ),
-    },
-    "answer_redaction": {
-        "version": ANSWER_REDACTION_DOCTRINE_VERSION,
-        "rules_count": len(ANSWER_REDACTION_DOCTRINE),
-        "consumed_by": (
-            # D-113 (ISS-115): الحارس الحتمي الأخير ضد كشف الإجابة النهائية.
-            "AnswerRedactionSkill.redact",
-            "content_integrity.sanitize_final_text",
-            "local_graph._apply_answer_redaction",
-        ),
-    },
-    "bkt_cognitive": {
-        "version": BKT_COGNITIVE_DOCTRINE_VERSION,
-        "rules_count": len(BKT_COGNITIVE_DOCTRINE),
-        "consumed_by": (
-            # D-074 (Protocol V6.0): الطبقة المعرفية الأساسية — موصولة حيّاً.
-            "BKTEngine.evaluate",
-            "bkt_persistence.BKTAnalyticsService.evaluate_and_record",
-            "customer_chat._evaluate_bkt_cards",
-            "orchestrator_client._build_probability_tree_props",
-        ),
-    },
-    "concept_diagnosis": {
-        "version": CONCEPT_DIAGNOSIS_DOCTRINE_VERSION,
-        "rules_count": len(CONCEPT_DIAGNOSIS_DOCTRINE),
-        "consumed_by": (
-            # D-127: الطبقتان 1+5 — موصولة حيّاً في مخرج الطوارئ الاحتمالي.
-            "ConceptDiagnosisSkill.diagnose",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "semantic_property": {
-        "version": SEMANTIC_PROPERTY_DOCTRINE_VERSION,
-        "rules_count": len(SEMANTIC_PROPERTY_DOCTRINE),
-        "consumed_by": (
-            # D-131: الطبقة 2 — الطبقة الدلالية + Misconception Graph.
-            "SemanticPropertySkill.interpret",
-            "orchestrator_client._build_cognitive_response",
-        ),
-    },
-    "pedagogical_policy": {
-        "version": PEDAGOGICAL_POLICY_DOCTRINE_VERSION,
-        "rules_count": len(PEDAGOGICAL_POLICY_DOCTRINE),
-        "consumed_by": (
-            # D-129: الطبقة 4 — موصولة حيّاً في مخرج الطوارئ الاحتمالي.
-            "PedagogicalPolicySkill.decide",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "socratic_evaluator": {
-        "version": SOCRATIC_EVALUATOR_DOCTRINE_VERSION,
-        "rules_count": len(SOCRATIC_EVALUATOR_DOCTRINE),
-        "consumed_by": (
-            # D-130: الطبقة 1 (الإصغاء النشط) — موصولة حيّاً قبل الاسترجاع المُفهرَس.
-            "SocraticEvaluatorSkill.evaluate",
-            "orchestrator_client._stream_socratic_evaluation",
-        ),
-    },
-    "adaptive_pedagogy": {
-        "version": ADAPTIVE_PEDAGOGY_DOCTRINE_VERSION,
-        "rules_count": len(ADAPTIVE_PEDAGOGY_DOCTRINE),
-        "consumed_by": (
-            "AdaptivePedagogySkill.derive",
-            "customer_chat._build_pedagogy_directive",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "probability_calculation": {
-        "version": PROBABILITY_CALCULATION_DOCTRINE_VERSION,
-        "rules_count": len(PROBABILITY_CALCULATION_DOCTRINE),
-        "consumed_by": (
-            # D-075 (Protocol V14.0): محرّك الاحتمالات الحتمي — موصول حيّاً.
-            "ProbabilityCalculatorSkill.analyze",
-            "orchestrator_client._build_probability_tree_props",
-            # D-078 (Protocol V19.0): الموجِّه التربوي (آني/تتابعي) — موصول حيّاً.
-            "orchestrator_client._build_calculated_ui",
-        ),
-    },
-    "impossible_case": {
-        "version": IMPOSSIBLE_CASE_DOCTRINE_VERSION,
-        "rules_count": len(IMPOSSIBLE_CASE_DOCTRINE),
-        "consumed_by": (
-            # D-082: content-retrieval-skill /impossible-case endpoint.
-            "content_retrieval_skill.src.impossible_case.analyze_impossible_case",
-            # frontend: ImpossibleCaseRenderer (4 مراحل بصرية).
-            "frontend.components.ImpossibleCaseRenderer",
-        ),
-    },
-    "realtime_protocol": {
-        "version": REALTIME_PROTOCOL_DOCTRINE_VERSION,
-        "rules_count": len(REALTIME_PROTOCOL_DOCTRINE),
-        "consumed_by": (
-            # D-WS-FLAP-002 (2026-05-26): معالج موحَّد لـ ping/pong عبر كل WS endpoints.
-            "ws_heartbeat_skill.handle_control_message",
-            "customer_chat.chat_stream_ws",
-            "admin.admin_chat_stream_ws",
-            "conversation_service.main.chat_ws",
-            "conversation_service.main.admin_chat_ws",
-        ),
-    },
-    "skills_platform": {
-        "version": SKILLS_PLATFORM_DOCTRINE_VERSION,
-        "rules_count": len(SKILLS_PLATFORM_DOCTRINE),
-        "consumed_by": (
-            # D-100: المنصّة الموحَّدة — registry + composition + observability.
-            "api.routers.skills.list_skills",
-            "api.routers.skills.refine_endpoint",
-            "registry.get_skill_registry",
-        ),
-    },
-    "content_integrity": {
-        "version": CONTENT_INTEGRITY_DOCTRINE_VERSION,
-        "rules_count": len(CONTENT_INTEGRITY_DOCTRINE),
-        "consumed_by": (
-            # D-106 (ISS-114): حارس نزاهة المحتوى — موصول حيّاً عند مخرج البثّ.
-            "ContentIntegritySkill.check",
-            "orchestrator_client.chat_with_agent",
-            "local_graph.run_local_graph_stream",
-            "local_graph.run_local_graph_with_exercise_context",
-        ),
-    },
-    "learning_path": {
-        "version": LEARNING_PATH_DOCTRINE_VERSION,
-        "rules_count": len(LEARNING_PATH_DOCTRINE),
-        "consumed_by": (
-            # D-111: المسار التعلّمي التكيفي — موصول حيّاً بعد تقييم BKT.
-            "LearningPathSkill.derive",
-            "customer_chat._evaluate_bkt_cards",
-        ),
-    },
-    "mandatory_orchestration": {
-        "version": MANDATORY_ORCHESTRATION_DOCTRINE_VERSION,
-        "rules_count": len(MANDATORY_ORCHESTRATION_DOCTRINE),
-        "consumed_by": (
-            # D-112: العمود الفقري الإلزامي — hard-fail عند تعذّر الـ orchestrator.
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "worked_example": {
-        "version": WORKED_EXAMPLE_DOCTRINE_VERSION,
-        "rules_count": len(WORKED_EXAMPLE_DOCTRINE),
-        "consumed_by": (
-            # D-115 (ISS-116): البروتوكول السقراطي المنضبط — التوليد في الـ orchestrator.
-            "search.SynthesizerNode",
-            "content_integrity.strip_garbage_markers",
-            "answer_redaction.redact_final_answers",
-        ),
-    },
-    "student_state": {
-        "version": STUDENT_STATE_DOCTRINE_VERSION,
-        "rules_count": len(STUDENT_STATE_DOCTRINE),
-        "consumed_by": (
-            # D-133: قراءة حالة الطالب (نيّة + إحباط) — إشارة قرار موصولة حيّاً.
-            "StudentStateSkill.read",
-            "orchestrator_client.chat_with_agent",
-            "pedagogical_policy.PolicyInput",
-        ),
-    },
-    "dialogue_manager": {
-        "version": DIALOGUE_MANAGER_DOCTRINE_VERSION,
-        "rules_count": len(DIALOGUE_MANAGER_DOCTRINE),
-        "consumed_by": (
-            # D-142 Phase 2: سلطة قرار الدور — موصولة حيّاً في المسار السقراطي (خلف العلم).
-            "DialogueManagerSkill.decide",
-            "orchestrator_client._stream_socratic_evaluation",
-            # D-158: طبقة القرار الموحَّدة فوق tutor_state.kc_progress الدائم (خلف COGNITIVE_TURN_ENABLED).
-            "orchestrator_client._cognitive_turn",
-        ),
-    },
-    "understanding_state": {
-        "version": UNDERSTANDING_STATE_DOCTRINE_VERSION,
-        "rules_count": len(UNDERSTANDING_STATE_DOCTRINE),
-        "consumed_by": (
-            # D-135: محرّك حالة الفهم (Learning State) — موصول حيّاً في مخرج الطوارئ.
-            "UnderstandingStateSkill.decide",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "pedagogical_escalation": {
-        "version": PEDAGOGICAL_ESCALATION_DOCTRINE_VERSION,
-        "rules_count": len(PEDAGOGICAL_ESCALATION_DOCTRINE),
-        "consumed_by": (
-            # D-138: المصفوفة التصعيدية التكيّفية — موصولة حيّاً في بلوك تعليم المفهوم.
-            "PedagogicalEscalationSkill.decide",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-    "micro_simulation": {
-        "version": MICRO_SIMULATION_DOCTRINE_VERSION,
-        "rules_count": len(MICRO_SIMULATION_DOCTRINE),
-        "consumed_by": (
-            # D-138: خادم محتوى L3 الحتمي — موصول حيّاً عبر المصفوفة التصعيدية.
-            "MicroSimulationSkill.get_micro_simulation",
-            "PedagogicalEscalationSkill.decide",
-            "orchestrator_client.chat_with_agent",
-        ),
-    },
-}
-
-
-def get_retrieval_doctrine_summary() -> str:
-    """يُرجِع doctrine الاستدعاء كنص قصير (للاستخدام في prompts / logs)."""
-    return " | ".join(RETRIEVAL_DOCTRINE)
-
-
-def get_explanation_doctrine_summary() -> str:
-    """يُرجِع doctrine الشرح كنص قصير (للاستخدام في system prompts)."""
-    return " | ".join(EXPLANATION_DOCTRINE)
-
-
-def get_model_answer_reliance_summary() -> str:
-    """يُرجِع doctrine الاحتجاج بالإجابة النموذجية (للـ prompts)."""
-    return " | ".join(MODEL_ANSWER_RELIANCE_RULES)
-
-
-def get_detailed_explanation_summary() -> str:
-    """يُرجِع doctrine الشرح المفصل (للـ prompts)."""
-    return " | ".join(DETAILED_EXPLANATION_RULES)
-
-
-def get_content_invocation_summary() -> str:
-    """يُرجِع بروتوكول استدعاء المحتوى كنص قصير (للـ prompts / logs)."""
-    return " | ".join(CONTENT_INVOCATION_DOCTRINE)
-
-
-def get_model_answer_explanation_summary() -> str:
-    """يُرجِع doctrine شرح الإجابة النموذجية (للـ system prompts)."""
-    return " | ".join(MODEL_ANSWER_EXPLANATION_DOCTRINE)
-
-
-def get_step_by_step_summary() -> str:
-    """يُرجِع قواعد الشرح خطوة بخطوة (للـ prompts)."""
-    return " | ".join(STEP_BY_STEP_EXPLANATION_RULES)
-
-
-def get_skill_invocation_protocol_summary() -> str:
-    """يُرجِع بروتوكول استدعاء الـ Skills (للـ orchestrator / logs)."""
-    return " | ".join(SKILL_INVOCATION_PROTOCOL)
-
-
-def get_bkt_cognitive_summary() -> str:
-    """يُرجِع doctrine الطبقة المعرفية BKT كنص قصير (للـ logs / prompts)."""
-    return " | ".join(BKT_COGNITIVE_DOCTRINE)
-
-
-def get_adaptive_pedagogy_summary() -> str:
-    """ملخص doctrine البيداغوجيا التكيفية (D-104) — سطر واحد للاستهلاك البرمجي."""
-    return " | ".join(ADAPTIVE_PEDAGOGY_DOCTRINE)
-
-
-def get_probability_calculation_summary() -> str:
-    """يُرجِع doctrine حساب الاحتمالات كنص قصير (للـ logs / prompts)."""
-    return " | ".join(PROBABILITY_CALCULATION_DOCTRINE)
-
-
-def get_answer_redaction_summary() -> str:
-    """يُرجِع doctrine حجب الإجابة النهائية كنص قصير (للـ logs)."""
-    return " | ".join(ANSWER_REDACTION_DOCTRINE)
-
-
-def get_worked_example_summary() -> str:
-    """يُرجِع doctrine المثال المحلول المُتدرّج كنص قصير (للـ prompts / logs)."""
-    return " | ".join(WORKED_EXAMPLE_DOCTRINE)
-
-
-def build_exercise_explanation_prompt() -> str:
-    """يبني system prompt شرح التمرين مباشرةً من الـ doctrine.
-
-    D-071 (2026-05-19): هذه الدالة هي **single source of truth** للـ prompt
-    المُرسَل للـ LLM عند شرح تمارين البكالوريا. أي تغيير في الـ doctrine
-    ينعكس تلقائياً على الـ prompt — لا drift ممكن.
-
-    القيود الإلزامية (D-067):
-    - الـ prompt يجب أن يبقى < 1000 حرف (النماذج المجانية تدخل reasoning-mode
-      مع prompts > 1500 حرف).
-    - لا box-drawing chars (U+2500–U+257F) — تُربك tokenizers النماذج المجانية.
-    - لا تكرار — كل قاعدة تُذكر مرة واحدة فقط.
-
-    المُستهلكون:
-    - `app.services.chat.local_graph._EXERCISE_EXPLANATION_SYSTEM_PROMPT`
-    - `microservices.conversation_service.src.conversation_graph`
-    """
-    # الجزء الثابت — هوية الأستاذ السقراطي (D-113)
-    header = (
-        "أنت أستاذ بكالوريا جزائري تُعلّم بالطريقة السقراطية.\n"
-        "مهمتك أن تجعل الطالب يفكّر ويصل للحل بنفسه — لا أن تحلّ مكانه.\n\n"
-    )
-
-    # نسخة مُوجَزة من المبادئ السقراطية الخمسة (تبقى مشتقّة من EXPLANATION_DOCTRINE،
-    # لكن مكثّفة لتبقى < 1000 حرف مع المراسي — D-067). كل سطر يطابق قاعدة في
-    # EXPLANATION_DOCTRINE[:5] (القاعدة الذهبية، المنع المطلق، المحاولة، السُلّم، «لم أفهم»).
-    # المراسي الإلزامية للـ CI: «الإجابة النموذجية» + «حرفياً» + «LaTeX».
-    core_rules = (
-        "1. القاعدة الذهبية: لا تكشف أبداً نتيجةً أو خطوةً يستطيع الطالب توليدها — قُد بالأسئلة."
-        "\n2. ممنوع منعاً باتاً كشف الإجابة النموذجية أو أي نتيجة نهائية (P(A)=14/165، "
-        "C(11,3)=165، $$\\boxed{...}$$) — لا حرفياً ولا بإعادة صياغة."
-        "\n3. محاولة قبل الشرح: لا تشرح مسألةً لم يحاولها الطالب — ادعُه للمحاولة ثم وجِّهه."
-        "\n4. سُلّم التلميحات: مبدأ ← سؤال موجِّه ← هدف جزئي ← خطوة تالية فقط. لا تقفز للإجابة."
-        "\n5. عند «لم أفهم»: شخّص موضع الكسر بسؤال، ثم تلميح أدنى واحد — لا تُعِد الاشتقاق."
-        "\n6. استخدم LaTeX لكل رمز رياضي ($...$) — لكن لا تكتب النتيجة النهائية."
-        "\n7. التزم بمعطيات هذا التمرين حصراً (لا موضوع خارجي). لا HTML في نصك."
-    )
-
-    prompt = header + "قواعد إلزامية:\n" + core_rules
-
-    # ضمان عدم تجاوز 1000 حرف (D-067 invariant)
-    if len(prompt) > 1000:
-        prompt = prompt[:997] + "..."
-
-    return prompt
-
-
-#: الـ prompt المُبنى من الـ doctrine — يُستخدم في local_graph و conversation_graph.
-#: D-071: هذا الثابت هو المرجع الوحيد — لا تعريف prompt محلي في أي ملف آخر.
-EXERCISE_EXPLANATION_SYSTEM_PROMPT: Final[str] = build_exercise_explanation_prompt()
-
-
-def list_all_doctrines() -> dict[str, dict[str, object]]:
-    """يُرجِع manifest كامل لكل الـ doctrines + versions + consumers (للـ CI)."""
-    return SKILL_DOCTRINE_MANIFEST
-
-
-__all__ = [
-    "BKT_COGNITIVE_DOCTRINE",
-    "BKT_COGNITIVE_DOCTRINE_VERSION",
-    "CONCEPT_DIAGNOSIS_DOCTRINE",
-    "CONCEPT_DIAGNOSIS_DOCTRINE_VERSION",
-    "CONTENT_INTEGRITY_DOCTRINE",
-    "CONTENT_INTEGRITY_DOCTRINE_VERSION",
-    "CONTENT_INVOCATION_DOCTRINE",
-    "CONTENT_INVOCATION_DOCTRINE_VERSION",
-    "DETAILED_EXPLANATION_RULES",
-    "DETAILED_EXPLANATION_VERSION",
-    "EXERCISE_EXPLANATION_SYSTEM_PROMPT",
-    "EXPLANATION_DOCTRINE",
-    "EXPLANATION_DOCTRINE_VERSION",
-    "IMPOSSIBLE_CASE_DOCTRINE",
-    "IMPOSSIBLE_CASE_DOCTRINE_VERSION",
-    "LEARNING_PATH_DOCTRINE",
-    "LEARNING_PATH_DOCTRINE_VERSION",
-    "MANDATORY_ORCHESTRATION_DOCTRINE",
-    "MANDATORY_ORCHESTRATION_DOCTRINE_VERSION",
-    "MICRO_SIMULATION_DOCTRINE",
-    "MICRO_SIMULATION_DOCTRINE_VERSION",
-    "MODEL_ANSWER_EXPLANATION_DOCTRINE",
-    "MODEL_ANSWER_EXPLANATION_VERSION",
-    "MODEL_ANSWER_RELIANCE_RULES",
-    "MODEL_ANSWER_RELIANCE_VERSION",
-    "PEDAGOGICAL_ESCALATION_DOCTRINE",
-    "PEDAGOGICAL_ESCALATION_DOCTRINE_VERSION",
-    "PEDAGOGICAL_POLICY_DOCTRINE",
-    "PEDAGOGICAL_POLICY_DOCTRINE_VERSION",
-    "PROBABILITY_CALCULATION_DOCTRINE",
-    "PROBABILITY_CALCULATION_DOCTRINE_VERSION",
-    "REALTIME_PROTOCOL_DOCTRINE",
-    "REALTIME_PROTOCOL_DOCTRINE_VERSION",
-    "RETRIEVAL_DOCTRINE",
-    "RETRIEVAL_DOCTRINE_VERSION",
-    "SEMANTIC_PROPERTY_DOCTRINE",
-    "SEMANTIC_PROPERTY_DOCTRINE_VERSION",
-    "SKILLS_PLATFORM_DOCTRINE",
-    "SKILLS_PLATFORM_DOCTRINE_VERSION",
-    "SKILL_DOCTRINE_MANIFEST",
-    "SKILL_INVOCATION_PROTOCOL",
-    "SKILL_INVOCATION_PROTOCOL_VERSION",
-    "SOCRATIC_EVALUATOR_DOCTRINE",
-    "SOCRATIC_EVALUATOR_DOCTRINE_VERSION",
-    "STEP_BY_STEP_EXPLANATION_RULES",
-    "STEP_BY_STEP_EXPLANATION_VERSION",
-    "WORKED_EXAMPLE_CLOSE",
-    "WORKED_EXAMPLE_DOCTRINE",
-    "WORKED_EXAMPLE_DOCTRINE_VERSION",
-    "WORKED_EXAMPLE_OPEN",
-    "build_exercise_explanation_prompt",
-    "get_bkt_cognitive_summary",
-    "get_content_invocation_summary",
-    "get_detailed_explanation_summary",
-    "get_explanation_doctrine_summary",
-    "get_impossible_case_summary",
-    "get_micro_simulation_summary",
-    "get_model_answer_explanation_summary",
-    "get_model_answer_reliance_summary",
-    "get_pedagogical_escalation_summary",
-    "get_probability_calculation_summary",
-    "get_realtime_protocol_summary",
-    "get_retrieval_doctrine_summary",
-    "get_skill_invocation_protocol_summary",
-    "get_skills_platform_summary",
-    "get_step_by_step_summary",
-    "get_worked_example_summary",
-    "list_all_doctrines",
-]
