+++ app/services/skills/doctrine/exercise_core.py
@@ -1,0 +1,282 @@
+"""Exercise Core — sliced verbatim from the former doctrine.py (D-173 Stage 2a).
+
+جزء من حزمة `app.services.skills.doctrine` — المصدر الوحيد للحقيقة لقواعد كل
+Skill. لا تُعرَّف doctrines خارج هذه الحزمة (بوّابة skills-doctrine تحرس ذلك).
+"""
+
+from __future__ import annotations
+
+from typing import Final
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Doctrine Version Pinning
+# ─────────────────────────────────────────────────────────────────────────────
+#
+# عند تعديل أي قاعدة من الـ doctrines أدناه:
+#   1. حدّث الـ version (`MAJOR.MINOR.PATCH`).
+#   2. شغّل `python scripts/fitness/check_skills_doctrine.py` ليُجدد الـ
+#      doctrine hash في الـ lock file.
+#   3. أضف entry في `.memory/decisions.md` يشرح *لماذا* تغيّرت.
+
+#: نسخة doctrine استدعاء/استرجاع المحتوى — تُستخدم لـ drift detection في CI.
+RETRIEVAL_DOCTRINE_VERSION: Final[str] = "1.0.0"
+
+#: نسخة doctrine الشرح — تطابق `EXPLANATION_DOCTRINE_VERSION` في
+#: `bac_exercise_skill.py` (يبقى متزامناً عبر تعريف موحَّد هنا).
+#: v3.0.0 (D-113 / ISS-115 — وَهْم الإتقان): تحوُّل ثوري من «شرح يكشف الحل»
+#: إلى «توليد مُجبَر سقراطي» — ممنوع منعاً باتاً كشف أي نتيجة نهائية للطالب.
+#: v3.1.0 (D-114): استثناء المثال المحلول المكشوف — عُكِس في D-115 (كارثة حيّة).
+#: v3.2.0 (D-115 / ISS-116): لا كشف؛ المبتدئ الغارق يتلقّى بروتوكولاً سقراطياً
+#: أبسط (فكرة واحدة + خطوة واحدة) لا مثالاً مكشوفاً. القاعدة الذهبية مطلقة.
+EXPLANATION_DOCTRINE_VERSION: Final[str] = "3.2.0"
+
+#: نسخة doctrine الاعتماد على الإجابة النموذجية.
+#: ⚠️ D-113: هذه القواعد تُستخدم *حصراً* في وضع التحقق (verification mode) —
+#: حيث يُقدّم الطالب حلّه ونتحقق منه. ممنوع استخدامها في الشرح العادي
+#: (الشرح العادي سقراطي بحت — لا يكشف الإجابة النموذجية).
+MODEL_ANSWER_RELIANCE_VERSION: Final[str] = "1.0.0"
+
+#: نسخة doctrine حجب الإجابة النهائية (D-113 — شبكة الأمان الحتمية).
+#: v1.1.0 (D-114): إعفاء واعٍ بالفواصل — كتل المثال المحلول الملفوفة بالفواصل
+#: الحارسة معفاة من الحجب عند support_level==1؛ غياب الفواصل = حجب كامل (fail-closed).
+ANSWER_REDACTION_DOCTRINE_VERSION: Final[str] = "1.1.0"
+
+#: نسخة doctrine البروتوكول السقراطي المنضبط (D-115 يَعكِس D-114 / ISS-116).
+WORKED_EXAMPLE_DOCTRINE_VERSION: Final[str] = "2.0.0"
+
+#: D-114: الفواصل الحارسة لكتلة المثال المحلول — عربي خالص (لا لاتيني/أقواس
+#: ASCII كي لا يكسرها مُطهّر اللغة أو مُفكِّك Markdown). الحجب الواعي يُعفي ما
+#: *بداخلها* عند support_level==1 ويحجب ما خارجها (راجع redact_final_answers).
+#: نفس القيمتين منسوختان في microservices response_sanitizer (حدّ معماري).
+WORKED_EXAMPLE_OPEN: Final[str] = "⟦مثال_محلول⟧"
+WORKED_EXAMPLE_CLOSE: Final[str] = "⟦/مثال_محلول⟧"
+
+#: نسخة doctrine الشرح التفصيلي.
+DETAILED_EXPLANATION_VERSION: Final[str] = "1.0.0"
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Content Invocation / Retrieval Doctrine
+# ─────────────────────────────────────────────────────────────────────────────
+# هذه القواعد تحكم *كيف يجب أن يُستدعى المحتوى التعليمي* من قاعدة المعرفة.
+# المُستفيدون: `BACExerciseSkill._retrieve`, `_stream_local_retrieval_response`,
+# `exercise_retrieval.detect_exercise_retrieval`.
+# ─────────────────────────────────────────────────────────────────────────────
+
+RETRIEVAL_DOCTRINE: Final[tuple[str, ...]] = (
+    "Indexed-first retrieval: قَبل البحث الواسع، حاول مطابقة الفهرس المحدد "
+    "(year + session + subject + exercise_number) في `knowledge_index`.",
+    "Atomic file load: عند توفر `matched_entry`، حمِّل ملفاً واحداً بالضبط. "
+    "لا scan على `knowledge_base/` يقرأ ملفات أخرى.",
+    "Display strip: المُخرَج للمستخدم يجب أن يخلو من YAML frontmatter، "
+    "أي قسم بعنوان «عناصر الإجابة» / «الإجابة النموذجية» / «الحل» / "
+    "«Solutions» / «وسوم البحث»، أو أي علامات chunk الداخلية مثل `[ex:`، "
+    "`[sol:`، `[grading:`.",
+    "Streaming chunk integrity: التقسيم يجب أن يحافظ على سلامة LaTeX. "
+    "أي `$...$` أو `$$...$$` أو `\\(...\\)` يُكشف ذرّياً.",
+    "Pre-empt orchestrator: عندما `matched_entry is not None`، يجب أن يبث الـ "
+    "monolith المحتوى المُفهرَس النظيف *قبل* محاولة orchestrator-service.",
+    "Conversation context: عند سؤال شرح/استفسار بدون مرجع صريح، يجب فحص "
+    "`history_messages` أولاً قبل اللجوء إلى retriever عام.",
+    'No leakage: لا يُرسل JSON envelope (مثل `{"المصدر":"...","التمرين":"..."}`) '
+    "للمستخدم. الـ caller المسؤول هو `_extract_human_readable_response`.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Explanation Doctrine — Socratic No-Answer (v3.0.0)
+# ─────────────────────────────────────────────────────────────────────────────
+# D-113 / ISS-115 (وَهْم الإتقان — The Illusion of Mastery):
+#   الكارثة: كشف الحلّ الكامل (P(A)=14/165، C(11,3)=165، E(X)=1.73) يُولّد عند
+#   الطالب «وهم الطلاقة» — يشعر أنه فهم بينما درّب التعرّف لا التوليد، فينهار
+#   يوم الامتحان. المعالجة (Bjork / Roediger / Kapur / Renkl): توليد مُجبَر،
+#   سُلّم تلميحات متصاعد، محاولة قبل الشرح، و«لم أفهم» = تشخيص لا إعادة اشتقاق.
+#
+#   القاعدة الذهبية فوق كل القواعد: لا تكشف أبداً نتيجةً أو خطوةً يستطيع الطالب
+#   توليدها بنفسه. ابحث عن حافّة ما يستطيعه واعمل خطوةً واحدة وراءها (ZPD).
+#
+# ⚠️ الإجابة النموذجية تُستخدم *داخلياً* كمرجع لتوجيهك الصحيح — لكنها لا تصل
+#    الطالب أبداً في الشرح العادي (تُكشف فقط في وضع التحقق المنفصل).
+#
+# ملاحظة CI: أول 5 قواعد تُبنى منها `build_exercise_explanation_prompt`، ويجب
+#   أن تحوي مجتمعةً المراسي: «الإجابة النموذجية» و«LaTeX» و«حرفياً».
+# ─────────────────────────────────────────────────────────────────────────────
+
+EXPLANATION_DOCTRINE: Final[tuple[str, ...]] = (
+    # — القاعدة الذهبية (Golden Rule) —
+    "القاعدة الذهبية: لا تكشف أبداً نتيجةً نهائية أو خطوةً يستطيع الطالب توليدها "
+    "بنفسه. قُد بالأسئلة والتلميحات لا بالحلول.",
+    # — المنع المطلق (anchors: «الإجابة النموذجية» + «حرفياً») —
+    "ممنوع منعاً باتاً كشف الإجابة النموذجية أو أي نتيجة نهائية (أرقام/كسور/صيغ "
+    "مثل P(A)=14/165 أو C(11,3)=165 أو $$\\boxed{...}$$) — لا حرفياً ولا بإعادة "
+    "صياغة. الطالب يصل للإجابة بنفسه؛ أنت تقود فقط.",
+    # — محاولة قبل الشرح (Productive Failure) —
+    "محاولة قبل الشرح: لا تشرح مسألةً لم يحاولها الطالب. ادعُه للمحاولة أولاً، "
+    "ثم وجِّه محاولته. حتى المحاولة الخاطئة تستحق التوجيه — غير المُحاوَلة لا.",
+    # — سُلّم التلميحات المتصاعد (Hint Ladder) —
+    "سُلّم التلميحات المتصاعد عند التعثّر: (1) أشِر إلى المبدأ، (2) اسأل سؤالاً "
+    "موجِّهاً، (3) اكشف الهدف الجزئي التالي، (4) اكشف الخطوة التالية فقط — لا "
+    "تقفز إلى الإجابة أبداً. معظم التعثّر ينحلّ عند 1-2.",
+    # — بروتوكول «لم أفهم» (anchor: LaTeX) —
+    "عند «لم أفهم» أو «اشرح لي»: شخّص موضع الكسر بسؤال أولاً («أين بالضبط: في "
+    "اختيار الطريقة أم في عدّ الحالات؟»)، ثم تلميح أدنى واحد عند تلك النقطة — "
+    "لا تُعِد الاشتقاق. اكتب أي رمز رياضي بـ LaTeX داخل $...$ أو $$...$$.",
+    # — الشرح الذاتي (Self-Explanation) —
+    "اطلب «لماذا؟» قبل الانتقال: بعد كل تلميح، اطلب من الطالب أن يبرّر الخطوة "
+    "أو يتوقّع التالية. لا قراءة سلبية — بناء فعّال.",
+    # — التفنيد بعد الخطأ (Refutation, not preemption) —
+    "دع الخطأ يقع ثم فنّده: لا تستبق الأخطاء المتوقّعة. إذا أخطأ الطالب، أظهِر "
+    "الخطأ ثم فنّد النموذج الخاطئ صراحةً («ظننتَ الترتيب مهمّاً، لكن... لأن...»).",
+    # — نطاق الشرح —
+    "إذا طلب الطالب جزءاً محدداً (I/II/III/أ/ب/ج)، اقتصر عليه ولا تتعدّاه.",
+    # — اللغة والشكل (D-069) —
+    "اللغة عربية فصحى نقية. لا كلمات روسية/صينية/إسبانية مُسرَّبة "
+    "(يُطبَّق `sanitize_response` نهاية كل بث للضمان).",
+    "تجنّب جدار النص: أقسام قصيرة، أسئلة موجزة، بلا جداول Markdown، تشبيه واحد يكفي.",
+    # — مكافحة تسرّب الموضوع (D-097) —
+    "التزم بمعطيات هذا التمرين حصراً. ممنوع أي موضوع خارجي (طب/أشعة/مثال بعيد). "
+    "إن غابت المعلومة عن التمرين فقل ذلك — لا تخترع سياقاً.",
+    # — منع تسرّب الواجهات (D-106) —
+    "ممنوع إخراج HTML/JSX أو أكواد واجهات (<div>، className...) في نص الطالب — "
+    "الواجهات تُبنى حصراً عبر قناة ui_component المُهيكلة.",
+    # — المبتدئ الغارق: انضباط لا كشف (D-115 يَعكِس D-114) —
+    "المبتدئ الغارق (support_level==1 أو حيرة متكرّرة) لا يتلقّى حلاً مكشوفاً، بل "
+    "بروتوكولاً سقراطياً أبسط: ثبّت فكرة واحدة («ما المعطى؟ ما المطلوب؟ هل نعدّ أم "
+    "نحسب احتمالاً شرطياً؟») ثم خطوة واحدة صغيرة جداً — لا جدران، لا مثال موازٍ، "
+    "لا قفز موضوع، لا لغة غير عربية. القاعدة الذهبية تبقى مطلقة (راجع WORKED_EXAMPLE_DOCTRINE).",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Answer Redaction Doctrine — شبكة الأمان الحتمية (D-113)
+# ─────────────────────────────────────────────────────────────────────────────
+# الدفاع الأول هو الـ doctrine السقراطي + إرسال أسئلة-فقط للـ LLM. لكن النماذج
+# المجانية قد تخترع/تخمّن نتيجة نهائية. هذه القواعد تحكم الحارس الحتمي الأخير
+# (`AnswerRedactionSkill`) الذي يحذف أي نتيجة نهائية تسرّبت قبل وصولها للطالب.
+# المُستهلكون: `sanitize_final_text` (orchestrator + local) + المسار المحلي.
+# ─────────────────────────────────────────────────────────────────────────────
+
+ANSWER_REDACTION_DOCTRINE: Final[tuple[str, ...]] = (
+    "احذف دائماً `$$\\boxed{...}$$` (و`\\boxed{...}`) — هو علامة النتيجة النهائية "
+    "في سياق البكالوريا، استبدله بدعوة سقراطية «احسبها بنفسك».",
+    "احذف أسطر النتيجة النهائية الصريحة: «الإجابة النهائية:»، «النتيجة:»، «إذن "
+    "... =» المُذيَّلة برقم/كسر كخلاصة (لا كخطوة وسيطة).",
+    "نطاق ضيّق لتجنّب false-positive: لا تلمس الصيغ التعليمية الوسيطة "
+    "(`نستخدم C(n,k)`، تعريف القانون) — فقط النتائج الختامية الصريحة.",
+    "احذف جداول توزيع الاحتمال الكاملة `P(X=...)` التي تكشف القانون جاهزاً.",
+    "fail-open مطلق: أي فشل في الحارس ⇒ النص كما هو (لا يكسر دور الطالب).",
+    "الحارس حتمي — بلا LLM، بلا عشوائية — قابل للاختبار بـ pytest.",
+    # — الإعفاء الواعي بالفواصل (D-114) —
+    "إعفاء support_level==1 الواعي بالفواصل: عند support_level==1 ووجود الفواصل "
+    "الحارسة (⟦مثال_محلول⟧ … ⟦/مثال_محلول⟧)، يُحجب كل ما *خارج* الفواصل ويمرّ ما "
+    "*بداخلها* (مثال محلول على مسألة مماثلة آمن الكشف). تُزال علامات الفواصل من "
+    "المُخرَج النهائي. غياب الفواصل أو support_level != 1 ⇒ حجب كامل (fail-closed) — "
+    "فلا تتسرّب إجابة تمرين الطالب المُقيَّم أبداً.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Disciplined Socratic Protocol Doctrine (D-115 — يَعكِس D-114)
+# ─────────────────────────────────────────────────────────────────────────────
+# D-114 (المثال المحلول المكشوف) كان كارثة حيّة: النموذج المجاني سرّب فواصل ⟦⟧
+# مشوّهة + system prompt إنجليزي + هلوسة فرنسية/ألمانية + قفز مواضيع + جدران نص.
+# التصحيح (مالك المشروع): المشكلة ليست «لا تكشف» بل التنفيذ الكارثي. القاعدة
+# الذهبية: «الرد الجيد لا يقول الجواب بل يقول ما الخطوة التالية الصحيحة».
+# (الاسم WORKED_EXAMPLE_DOCTRINE يبقى للتوافق؛ المحتوى صار بروتوكول سقراطي منضبط.)
+# المُستهلكون: search.SynthesizerNode (التوليد السقراطي)، content_integrity +
+# response_sanitizer (المُطهّر المُصفّح)، answer_redaction (الحجب fail-closed).
+# ─────────────────────────────────────────────────────────────────────────────
+
+WORKED_EXAMPLE_DOCTRINE: Final[tuple[str, ...]] = (
+    "لا مثال محلول مكشوف (D-115 يَعكِس D-114): الطالب الغارق يتلقّى بروتوكولاً "
+    "سقراطياً منضبطاً قصيراً — صياغة المطلوب + نوع المسألة + سؤال تشخيصي واحد + "
+    "أصغر خطوة تالية. لا جدران نص، لا أمثلة موازية، لا كشف للجواب.",
+    "support_level (1..5) يحكم عمق التلميح فقط: 1=ثبّت فكرة واحدة للغارق بأبسط "
+    "لغة ← 5=ادفع المتمكّن ليستنتج بنفسه. لا يفتح كشف الجواب أبداً في التدفّق العادي.",
+    "السلطة الوحيدة للتوليد التعليمي هي الـ orchestrator/LangGraph (SynthesizerNode) "
+    "— لا توليد تعليمي في المونوليث؛ النظام يتعطّل صراحةً بدون الخدمات المصغرة (D-112).",
+    "حجب نهائي fail-closed: أي نتيجة نهائية (أرقام/كسور/$$\\boxed{}$$) تُحجب قبل "
+    "وصول الطالب. الكشف لوضع مراجعة منفصل فقط (M8 — مؤجَّل، خارج هذا التدفّق).",
+    "مُطهّر مُصفّح إلزامي: أي فواصل ⟦⟧ مشوّهة/مقسّمة، أو تعليمات system prompt "
+    "مُسرَّبة (WARM-UP...)، أو لغة غير عربية معزولة — تُحذف على deltas + النهائي "
+    "في المونوليث والـ orchestrator (شبكة أمان حتمية، fail-open).",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Model Answer Reliance Rules
+# ─────────────────────────────────────────────────────────────────────────────
+# قواعد رسمية لكيفية الاعتماد على الإجابة النموذجية أثناء الشرح المفصل.
+# هذه قواعد *أحكم* من EXPLANATION_DOCTRINE — لا تُكسر بدون ADR.
+# ─────────────────────────────────────────────────────────────────────────────
+
+MODEL_ANSWER_RELIANCE_RULES: Final[tuple[str, ...]] = (
+    # — مبدأ الاحتجاج (citation principle) —
+    "الإجابة النموذجية هي *مرجع الحُجَّة*. كل نتيجة عددية يجب أن تُحتَجَّ بها.",
+    # — الأرقام (numeric invariants) —
+    "كل قيمة عددية في الشرح يجب أن تطابق نظيرتها في الإجابة النموذجية. "
+    "مثال: لو الإجابة تقول `g'(x) = (x+2)e^(-x)`، الشرح لا يكتب "
+    "`g'(x) = (x+2)·e^(-x)` بصيغة مختلفة فحسب — بل نفس الصيغة بالضبط.",
+    # — الصيغ الرمزية —
+    "الصيغ النهائية (limits, derivatives, integrals) المكتوبة في الإجابة "
+    "تُحفَظ كما هي. الشرح يضيف *لماذا* وليس *كيف نُعيد كتابتها*.",
+    # — التحقق (verification) —
+    "بعد كل خطوة كبيرة، تَحقَّق أن النتيجة المستنبطة تطابق الإجابة النموذجية. "
+    "إن لم تطابق، فالشرح غير صحيح ويجب تعديله، لا تعديل الإجابة النموذجية.",
+    # — المنطقة الحرة (interpretive freedom) —
+    "الشرح حر في:\n"
+    "  - اختيار ترتيب الأفكار التعليمية.\n"
+    "  - إضافة تذكيرات بقواعد سابقة (مثلاً تذكير بقاعدة لوبيتال قبل تطبيقها).\n"
+    "  - إضافة تفسيرات هندسية/فيزيائية للنتيجة.\n"
+    "  - استخدام أمثلة عددية مساعدة (مع تأشيرها كـ «مثال توضيحي»).",
+    # — المنطقة الممنوعة (forbidden zone) —
+    "الشرح ممنوع من:\n"
+    "  - تعديل النتائج العددية في الإجابة النموذجية.\n"
+    "  - إعادة صياغة الصيغ بشكل مختلف عن الإجابة النموذجية.\n"
+    "  - نسخ الإجابة النموذجية حرفياً (يجب وجود قيمة تعليمية مضافة).\n"
+    "  - تجاهل أجزاء من الإجابة النموذجية إذا كانت مطلوبة في السؤال.",
+    # — الحالة الخاصة: الإجابة النموذجية مفقودة —
+    "إذا لم تتوفر الإجابة النموذجية في `full_content`، الـ Skill يُرجِع "
+    "`SkillFailure(reason='no_explanation_context')` بدلاً من توليد شرح "
+    "بدون مرجع حُجَّة.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Detailed Explanation Rules — الشرح المفصل للطالب
+# ─────────────────────────────────────────────────────────────────────────────
+# هذه القواعد تُكمِّل EXPLANATION_DOCTRINE بضوابط الـ *حجم* و *التدريج*
+# لكل سؤال نوع. يُستخدمها `_classify_question_budget` في `local_graph.py`
+# لتخصيص `max_tokens` المناسب.
+# ─────────────────────────────────────────────────────────────────────────────
+
+DETAILED_EXPLANATION_RULES: Final[tuple[str, ...]] = (
+    # — السؤال المفهومي (concept) —
+    "سؤال «ماذا نقصد بـ X» / «ما هو معنى X» = شرح قصير (≤ 350 token).",
+    "السؤال المفهومي يبدأ بـ تعريف صريح، ثم مثال واحد، ثم تطبيق على التمرين.",
+    # — السؤال التبريري (justification) —
+    "سؤال «لماذا X» / «علِّل X» = شرح متوسط (≤ 450 token).",
+    "السؤال التبريري يبدأ بـ القاعدة، ثم تطبيقها على بيانات التمرين، "
+    "ثم استنتاج النتيجة المُحتجة بها.",
+    # — السؤال المنهجي (method) —
+    "سؤال «كيف نُثبت X» / «كيف نحسب X» = شرح موسَّع (≤ 600 token).",
+    "السؤال المنهجي يبدأ بـ المنهجية العامة، ثم تطبيقها على بيانات التمرين، "
+    "ثم النتيجة في `$$\\boxed{...}$$`.",
+    # — السؤال الافتراضي (default) —
+    "سؤال «اشرح ...» بدون تخصيص = شرح متوازن (≤ 700 token).",
+    # — السؤال الشامل (full) —
+    "سؤال «اشرح التمرين كاملاً» = شرح طويل (≤ 900 token).",
+    "الشرح الشامل يَطوي كل الأجزاء (I/II/III) بفقرات منفصلة، كل واحدة مُختصرة.",
+    # — قواعد تنسيقية موحَّدة —
+    "كل خطوة تبدأ بـ رقم (1) (2) (3) لتسهيل المتابعة.",
+    "العناوين الفرعية بـ `### الجزء I` / `### الخطوة 1` لتفصيل البصري.",
+    "الفرضيات بـ `**فرضية:**`، النتائج بـ `**نتيجة:**`.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Content Invocation Doctrine (D-070)
+# ─────────────────────────────────────────────────────────────────────────────
+# بروتوكول استدعاء المحتوى التعليمي — يحدد الخطوات الإلزامية التي يجب أن
+# يتبعها أي Skill أو handler عند استدعاء محتوى من `knowledge_base/`.
+# المُستفيدون: `BACExerciseSkill`, `local_graph`, `orchestrator_client`.
