+++ app/services/skills/doctrine/pedagogy.py
@@ -1,0 +1,209 @@
+"""Pedagogy — sliced verbatim from the former doctrine.py (D-173 Stage 2a).
+
+جزء من حزمة `app.services.skills.doctrine` — المصدر الوحيد للحقيقة لقواعد كل
+Skill. لا تُعرَّف doctrines خارج هذه الحزمة (بوّابة skills-doctrine تحرس ذلك).
+"""
+
+from __future__ import annotations
+
+from typing import Final
+
+# ─────────────────────────────────────────────────────────────────────────────
+# BKT Cognitive Doctrine (D-074 / Protocol V6.0) — الطبقة المعرفية الأساسية
+# ─────────────────────────────────────────────────────────────────────────────
+BKT_COGNITIVE_DOCTRINE_VERSION: Final[str] = "3.0.0"
+
+#: قوانين غير قابلة للكسر تحكم طبقة التتبّع المعرفي البايزي (BKT) — الأساس
+#: المعرفي لكل skill تربوي مستقبلي في منصة E-TAALEEM.
+BKT_COGNITIVE_DOCTRINE: Final[tuple[str, ...]] = (
+    "BKT هي الطبقة المعرفية الأساسية لكل skill تربوي مستقبلي — أي قدرة تكيّفية "
+    "(تدرّج الصعوبة، التلميحات، مسار التعلّم) تُبنى فوق student_mastery_probability.",
+    "Zero Cognitive Overload: الجمهور طلاب بكالوريا — الرموز المجرّدة (A, B|A, Ā) "
+    "ممنوعة نهائياً في عُقَد الواجهة التوليدية.",
+    "Abstraction Ban: كل مكوّن UI توليدي يستخدم نموذج الاستخراج الهجين "
+    "(deterministic-first ثم LLM عند الحاجة) لإخراج تسميات ملموسة بشرية "
+    "(كرة حمراء، سحب ناجح، قطعة معيبة) — وحتى الـ fallback ملموس، لا حرف مجرّد.",
+    "BKT سجل append-only فقط — كل تفاعل صف جديد؛ لا تحديث في المكان. الإتقان "
+    "السابق يُقرأ من آخر صف لنفس (user_id, concept_id). يحفظ السلسلة الزمنية كاملة.",
+    "المخطط الإلزامي: concept_id + cognitive_load_estimate (low/medium/high) + "
+    "student_mastery_probability ∈ [0,1] + interaction_timestamp.",
+    "student_mastery_probability يُحدَّث بمعادلة BKT بايزية حتمية "
+    "(P_L0/P_T/P_S/P_G) — نفس المدخلات تُنتج نفس المخرجات (قابل للاختبار).",
+    "BKT لا يكسر المحادثة أبداً — كل استدعاء معزول في try/except بجلسة DB مستقلة.",
+    # D-126/D-157 (الإتقان الصادق ثنائي القناة — roadmap M6):
+    "الإتقان الصادق ثنائي القناة: نفصل الأداء المدعوم (assisted — مُضخَّم بالمساعدة) "
+    "عن الإتقان الدائم (durable). durable يُحدَّث بمنحنى نسيان متّصل (Half-Life "
+    "Regression، D-157): يتراكم بالدليل غير المدعوم المؤجَّل ويضمحلّ بالزمن — حتمي، صفر LLM.",
+    "الثابت المضاد للوهم (D-157): كسب durable ∝ generation_weight(support) ⇒ ≈0 عند "
+    "الدعم الثقيل (support=1)، فلا يُضخَّم الإتقان الدائم بالمساعدة أبداً — يبقى صادقاً.",
+    "إشارة الصواب ثلاثية الحالة (D-157): CORRECT/INCORRECT/UNKNOWN. UNKNOWN (لا دليل) "
+    "⇒ لا يُحدَّث الإتقان (يُصلِح التحيّز الهابط الذي كان يفترض الخطأ افتراضياً).",
+    "الإتقان المعروض للطالب = durable الصادق، لا المدعوم المُضخَّم. المساعدة الثقيلة "
+    "تُضخّم P(صحيح بلا معرفة) (scaffold_leak) فالإجابة المدعومة غير تشخيصية وحدها.",
+    "فجوة الوهم = assisted − durable = مقياس النجاح الوحيد (§0.6) — تُبَثّ إلى "
+    "Prometheus (M9، D-157). ممنوع التحسين على «الرضا اللحظي» أو مدّة الجلسة أو عدد "
+    "الرسائل — نُحسّن على تقليص الفجوة.",
+)
+
+# ── D-127/D-128: المعمارية الإدراكية العصبية-الرمزية — الطبقات 1 (الفهم) و 4 (السقراطي) و 5 ──
+# D-131: event_meaning يشمل خصائص الجداء (B/C/D) عبر الطبقة الدلالية العامة.
+CONCEPT_DIAGNOSIS_DOCTRINE_VERSION: Final[str] = "1.2.0"
+
+#: قوانين غير قابلة للكسر تحكم تشخيص المفهوم (الطبقة 1) + النموذج العقلي (الطبقة 5) +
+#: السرد السقراطي المُولَّد (الطبقة 4، D-128).
+CONCEPT_DIAGNOSIS_DOCTRINE: Final[tuple[str, ...]] = (
+    "الطبقة 1 (LLM) تُصنّف فقط — مخرج (concept + misconception) من enum ثابت، بلا "
+    "رياضيات/أرقام/نثر/إجابة. مجرّد تشخيص لِـ «ما الذي يحاول الطالب قوله؟».",
+    "النموذج العقلي (الطبقة 5): نخزّن *سبب* الحيرة (misconception: sample_space_confusion / "
+    "fraction_meaning_confusion / order_vs_selection_confusion) لا «أخطأ». المفهوم الخاطئ "
+    "يقود اختيار التدخّل السقراطي التالي.",
+    "كل صياغات المفهوم الواحد («ما هو البسط»/«لماذا 14»/«ما الهدف من 14»/«واش راه 14») ⇒ "
+    "مفهوم واحد (numerator) ⇒ تدخّل واحد مُوجَّه. لا قاعدة جديدة لكل صياغة.",
+    "التشخيص هجين: خريطة حتمية أولاً (سريعة/آمنة)، LLM محروس فقط عند unknown — مخرج مُتحقَّق "
+    "ضد الـ enum + timeout + fallback حتمي. أي مخرج خارج الـ enum ⇒ unknown (آمن).",
+    "الحقيقة الرياضية (الأرقام) تبقى في المحرك الرمزي حصراً (الطبقة 3) — صفر LLM. 14 تبقى 14.",
+    "لا تكرار حرفي: نفس المفهوم مرّتين ⇒ تصعيد سقراطي (شرح → سؤال → تشخيص/إعادة توجيه)، "
+    "لا نصّ متطابق. الأستاذ الحقيقي لا يُعيد الشرح للمرّة الثالثة — بل يسأل.",
+    # D-128 (الـ LLM كجهاز عصبي — السرد السقراطي المُولَّد):
+    "الـ LLM = الجهاز العصبي: يُنتج **الفهم/التربية** (السؤال السقراطي عند التصعيد)، لا "
+    "**الحقيقة** (الأرقام تُحقن من المحرك الرمزي ولا يُولّدها الـ LLM). LLM = فهم؛ symbolic = حقيقة.",
+    "السرد السقراطي المُولَّد محروس إلزامياً: عربي فقط + لا كشف جواب نهائي (14/165) + لا garbage "
+    "+ timeout + fallback حتمي. أي فشل/شكّ ⇒ القالب الحتمي (لا تدهور للطالب). يُلغي التكرار.",
+    "مفهوم event_meaning (معنى الحادثة A/B/C/D) = تعريف الحدث الملموس من نص التمرين (حقيقة)، "
+    "لا الحل الكامل. كل صياغات معنى الحادثة ⇒ event_meaning.",
+    # D-131 (الطبقة الدلالية العامة):
+    "event_meaning يشمل خصائص الجداء (B فردي / C زوجي / D معدوم) عبر الطبقة الدلالية العامة "
+    "(SemanticPropertySkill) — لا special-casing لكل حادثة. «ماذا نقصد بجداء معدوم» ⇒ تعريف "
+    "حتمي (الرقم 0)، لا «نفس اللون» الخاطئ ولا حساب.",
+)
+
+
+# ── D-131: الطبقة الدلالية + Misconception Graph — الطبقة 2 (substrate التوأم المعرفي) ──
+# D-132: جاهزية الأسئلة الجديدة عبر LLM Listener-Definer المحروس + لا default مُجمَّد.
+# D-159: الحواف (MISCONCEPTION_EDGES) + تشخيص الجذر عبر الاجتياز + مرآة knowledge_nodes/edges.
+SEMANTIC_PROPERTY_DOCTRINE_VERSION: Final[str] = "1.4.0"
+
+#: قوانين غير قابلة للكسر تحكم الطبقة الدلالية + شبكة المفاهيم الخاطئة (D-131/D-132/D-136/D-137).
+SEMANTIC_PROPERTY_DOCTRINE: Final[tuple[str, ...]] = (
+    "طبقة دلالية عامة data-driven لا special-casing: أي خاصية/اعتقاد خاطئ جديد = مدخل في "
+    "PROPERTY_REGISTRY / MISCONCEPTION_GRAPH، لا فرع if/elif جديد. مفسّر واحد لكل العبارات الرياضية.",
+    "المعمارية العصبية-الرمزية: الـ LLM = Listener فقط (لغة حرّة → property/misconception canonical "
+    "عبر مخرج enum مُقيَّد محروس + fallback حتمي) — ليس حَكماً خفياً ولا مصدراً للحقيقة. الأرقام من "
+    "المحرك الرمزي حصراً (الطبقة 3). دمج لا اختيار.",
+    "Misconception Graph «شخّص ثم تدخّل»: نفس المفهوم بـ misconceptions مختلفة ⇒ تدخّلات مختلفة. "
+    "عند الغموض يُصدَر probe تشخيصي قبل التدخّل. ممنوع تدخّل واحد موحَّد لكل من يسأل عن نفس المفهوم.",
+    "تصنيف الاعتقاد الخاطئ مبدئي صغير: rule_property | symbol_meaning | example_linking — لا اسم "
+    "جديد لكل خطأ (يمنع تضخّم عقد متشابهة). كل عقدة لها bkt_concept صريح (لا عقدة بلا أثر مقاس).",
+    "التعريف حقيقة حتمية (الطبقة 3) يُسلَّم للسياسة لا نهايةُ المسار: التعريف ثم السياسة التربوية "
+    "(D-129: تعريف→سؤال→اعتراف→تقدّم) — فتُجاب «لماذا/كيف» لا «ما معنى» فقط.",
+    "القياس قبل التوسّع: أي توسّع (مفاهيم منهج أخرى / twin دائم) مؤجَّل حتى تُثبت المقاييس السلوكية "
+    "(cogniforge_tutor_*) تحسّناً حقيقياً: تقليل التكرار، تعريف فوري، اختلاف التدخّلات، تقدّم بعد كل تفاعل.",
+    # D-132 (جاهزية الأسئلة الجديدة):
+    "جاهزية الأسئلة الجديدة دائماً: المفاهيم المعروفة تعريفها حتمي من السجلّ؛ المفهوم الجديد يُعرَّف "
+    "عبر LLM Listener-Definer محروس (تعريف مفاهيمي عام قصير، عربي، لا حساب، لا كشف نتيجة نهائية). "
+    "ممنوع default مُجمَّد لمفهوم واحد (unknown ≠ event_meaning، interpret None ≠ same_color) — المفهوم "
+    "غير المعروف يُعرَّف أو يُترك للسياسة، لا يُقحَم في الحادثة A. القياس عبر source=llm.",
+    # D-136 (مثال واعٍ بالمفهوم — قتل المثال الأعمى المتكرّر):
+    "المثال واعٍ بالمفهوم النشط لا أعمى: «اعطني مثال» ⇒ مثال **المفهوم الذي يجري الحوار عنه** "
+    "(detect_active_concept من السياق: PropertySpec.example الحتمي لكل مفهوم)، لا مثال الحادثة A "
+    "الافتراضي. ممنوع مثال/تعريف مكرّر حرفياً — المعروض already ⇒ زاوية مختلفة (LLM محروس) أو تقدّم. "
+    "ومحرّك حالة الفهم (D-135) لا يختطف أسئلة خارج مسار نفس-اللون (decide ⇒ None) فتُسلَّم للمعالج الواعي.",
+    # D-137 (فصل التوجيه — قتل تصارع النظامين):
+    "فصل صارم بين طبقتي المفهوم والحساب: أسئلة **المفهوم** (تعريف/معنى/مثال لـ A/B/C/D/X/E(X)/الشرطي/"
+    "الحالات الملائمة) ⇒ semantic_property حصراً؛ أسئلة **خطوة الحساب** (كيف/لماذا رقمٌ في حلّ P(A)) ⇒ "
+    "understanding_state (D-135) حصراً. لا تداخل، لا اختطاف.",
+    "«ما هو X» تعريف موثوق: أي سؤال نيّته definition (StudentState) أو «ما هو/ما هي/عرّف/ماذا نقصد» ⇒ "
+    "مسار التعريف **قبل** D-135. إشارات D-135 يجب أن تكون خاصة بخطوة الحساب لا كلمات مفهوم عامة — "
+    "ممنوع «الاحتمال»/«نفس اللون» المجرّدة في gap_signals (تختطف). الحيرة تُعيد إشراك المفهوم النشط "
+    "(detect_active_concept) لا تسقط للحادثة A الافتراضية؛ و_current_focus_kc لا يقفل على مثال/تعريف مفهوم.",
+    # D-159 (WP-D — الـ graph الحقيقي بالحواف):
+    "الشبكة graph حقيقي قابل للاجتياز لا dict مسطّحاً: الحواف (MISCONCEPTION_EDGES — prerequisite_of/"
+    "confused_with) تربط المفاهيم، و`diagnose_root` يجتازها ليجد **أعمق شرطٍ مسبق ضعيف** بحسب الحالة "
+    "الدائمة (tutor_state.kc_progress) ⇒ التدخّل يستهدف الجذر لا العرَض (أساس فجوة الوهم). مفهوم بلا "
+    "سجل دائم ليس «ضعيفاً» — لا تدخّل أعمى. أي حافة جديدة = مدخل data لا فرع كود.",
+    "جدولا knowledge_nodes/knowledge_edges مرآة حيّة للـ graph (scripts/sync_knowledge_graph.py عبر "
+    "جسر HTTPS + بوّابة تكافؤ code↔DB) — المسار الحي يبقى in-code حتمياً (لا I/O لقاعدة البيانات في "
+    "مسار الدور). المزامنة idempotent (UUIDv5 + ON CONFLICT DO UPDATE).",
+)
+
+
+# ── D-129: محرّك السياسة التربوية — الطبقة 4 (أقل تدخّل مفيد الآن) ──────────────────
+# D-130 (تحسينا المالك): كشف الإجابة بالفعل الكلامي لا بالطول + ميزانية تكيّفية.
+PEDAGOGICAL_POLICY_DOCTRINE_VERSION: Final[str] = "1.1.0"
+
+#: قوانين غير قابلة للكسر تحكم اختيار التدخّل التربوي (الطبقة 4، D-129/D-130).
+PEDAGOGICAL_POLICY_DOCTRINE: Final[tuple[str, ...]] = (
+    "العقد التربوي: تعريف → سؤال سقراطي محدود → اعتراف بالإجابة + تقدّم → حلّ رمزي عند الحاجة. "
+    "كل خطوة = أصغر تدخّل مفيد الآن، لا «رد» منفصل بل إدارة لرحلة الإتقان.",
+    "السقراطية محدودة بميزانية: بعد استنفادها بلا تقدّم ⇒ حلّ رمزي متدرّج (إنقاذ تربوي)، "
+    "لا سؤال إضافي. ممنوع الاستجواب اللانهائي.",
+    "الاعتراف بإجابة الطالب إلزامي: إجابة بعد سؤال ⇒ اعتراف «أحسنت» + تقدّم خطوة، لا سؤال "
+    "أصمّ آخر. الإجابة لا تُقابل بالصمت بل بالتقدير.",
+    "القرار حتمي تماماً — لا LLM في اختيار التدخّل. الـ LLM للسرد السقراطي/التقييم فقط (D-128/D-130، "
+    "محروس)؛ الحلّ الرمزي من المحرك الرمزي (D-113-safe).",
+    "القاعدة الذهبية: لا سؤال إضافي بلا معنى · لا تكرار بلا تقدّم · لا إجابة بلا سياق · "
+    "لا حلّ رمزي قبل أن تُستنفد محاولة الفهم.",
+    # D-130 تحسين A (Concern 1): كشف الإجابة بالفعل الكلامي لا بالطول.
+    "كشف «هل أجاب الطالب» بالفعل الكلامي (Speech-Act) من التاريخ لا بطول الرسالة: إجابة = "
+    "«سؤال معلّق + ليست أمراً صريحاً ولا تبديل موضوع». «نفس اللون» القصيرة والإجابة العبقرية "
+    "الطويلة كلتاهما إجابة. ممنوع ربط الإجابة بعدد الكلمات.",
+    # D-130 تحسين B (Concern 2): ميزانية تكيّفية حسب المفهوم وحالة الطالب.
+    "ميزانية الأسئلة تكيّفية لا ثابتة: socratic_budget(concept, support_level, confusion_count) "
+    "مرتبطة بالمفهوم (مفاهيمي أعمق، بصيرة سريعة أقصر) وبحالة الطالب (BKT/الإتقان: المتعثّر يُسرَّع "
+    "للسقالة، المتمكّن يحتمل عمقاً)، clamp [1,3]. ممنوع سقف تدخّل واحد ثابت لكل المفاهيم.",
+)
+
+
+# ── D-130: مُقيّم الإجابات السقراطي — الطبقة 1 العصبي-الرمزي (الإصغاء النشط) ──────────
+SOCRATIC_EVALUATOR_DOCTRINE_VERSION: Final[str] = "1.0.0"
+
+#: قوانين غير قابلة للكسر تحكم الإصغاء النشط وتقييم الرد الحرّ (الطبقة 1، D-130).
+SOCRATIC_EVALUATOR_DOCTRINE: Final[tuple[str, ...]] = (
+    "إصغاء نشط إلزامي: إذا كانت أحدث رسالة مساعد سؤالاً سقراطياً طرحناه، فرسالة الطالب التالية "
+    "= إجابة تُقيَّم قبل أي إعادة استرجاع. ممنوع منعاً باتاً إعادة طباعة التمرين ردّاً على إجابة.",
+    "اللُّبّ العصبي-الرمزي: الـ LLM يُقيّم *فهم* الطالب (يقارن رده الحرّ بالهدف التربوي المستتر) — "
+    "الطبقة 1؛ المحرك الرمزي يكتب *الخطوة* الناتجة (الأرقام 4/10/14/165) — الطبقة 3. LLM = فهم؛ symbolic = حقيقة.",
+    "التشجيع (encouragement) لا يحوي نتيجة نهائية أبداً (14/165/56/165)؛ محروس بـ "
+    "redact_final_answers + _strip_garbage_markers + is_probably_non_arabic + timeout.",
+    "التسليم الرمزي متدرّج: عند الفهم يُكشف خطوة المفهوم الحالي (مثلاً البسط: C(4,3)+C(5,3)=14) + "
+    "سؤال المتابعة، لا الحل كاملاً — يُبقي السُّلّم السقراطي حياً.",
+    "fail-open مطلق: أي تعذّر في المُقيّم ⇒ اعتراف لطيف + تلميح (لا عقاب، لا إعادة طباعة). الطالب "
+    "الذي شغّل عقله لا يُصفَع برمجياً أبداً.",
+    "حارس تبديل الموضوع (D-101) يبقى: إجابة سقراطية لا تُلتقَط إن كانت أمر تمرين جديد صريح أو موضوعاً آخر.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Adaptive Pedagogy Doctrine (D-104) — إغلاق حلقة BKT: القراءة تقود السلوك
+# ─────────────────────────────────────────────────────────────────────────────
+ADAPTIVE_PEDAGOGY_DOCTRINE_VERSION: Final[str] = "1.2.0"
+
+#: «المعيار الأعلى ليس الانبهار اللحظي، بل الاستقلال المعرفي» — E-TAALEEM.
+#: قوانين طبقة البيداغوجيا التكيفية: قراءة الإتقان تُحدِّد عمق التدريس.
+ADAPTIVE_PEDAGOGY_DOCTRINE: Final[tuple[str, ...]] = (
+    "عمق التدريس يتبع الإتقان: socratic عند mastery >= 0.7 (قُد بأسئلة، لا تكشف "
+    "الحل مباشرة)، guided في المنتصف (تلميح ثم خطوات بفجوات)، scaffolded تحت 0.35 "
+    "أو عند غياب البيانات (شرح كامل بسقالات).",
+    "يُبنى حصراً فوق student_mastery_probability من BKT (D-074) — ممنوع إعادة "
+    "اختراع تتبّع إتقان موازٍ.",
+    "fail-open مطلق: أي تعذّر (DB/تصنيف/مهلة) ينتج توجيهاً فارغاً — التوجيه "
+    "التربوي لا يكسر دور الطالب أبداً.",
+    "السياسة حتمية بعتبات ثابتة — نفس (mastery, load) تُنتج نفس المستوى دائماً؛ "
+    "تغيير العتبات يتطلب ترقية إصدار هذا الـ doctrine.",
+    "directive_text أقل من 400 حرف، عربي بسيط، بلا LaTeX وبلا رموز box-drawing "
+    "(قواعد D-067) — يُسبق في السؤال كتوجيه لا كـ system prompt.",
+    "الحِمل المعرفي المرتفع يُخفِّض درجة واحدة — «دع العقل يعمل، لكن لا تحرقه بالحمل الزائد».",
+    "التقاط المفهوم الخاطئ قبل أن يتجذر: markers حتمية لكل concept_id — التصحيح "
+    "«على مستوى الفكر لا على مستوى الحكم»، بلطف ودون لوم.",
+    "الغاية العليا الاستقلال المعرفي: المتمكن يُقاد لحلّه بنفسه — أن يصبح الطالب "
+    "أكبر من حاجته إلى الجواب.",
+    # D-113 (v1.1.0): سُلّم الدعم الخماسي — السقالة تتلاشى مع الإتقان.
+    "سُلّم الدعم 5 درجات (support_level 1..5): 1=مثال محلول كامل ← 2=إكمال "
+    "← 3=تلاشٍ عكسي ← 4=تلميحات ← 5=بلا دعم (unaided). السقالة تتلاشى مع نموّ "
+    "الإتقان (worked-example fading) — السقالة التي لا تُزال تصير قفصاً.",
+    # D-114 (v1.2.0): support_level يعبر السلسلة كاملةً ويحكم الكشف.
+    "support_level (int 1..5) يُمرَّر صراحةً عبر السلسلة كاملةً (customer_chat → "
+    "context → orchestrator_client → routes → AgentState → SynthesizerNode) "
+    "ويحكم عمق الكشف وبوّابة المثال المحلول (D-114). الافتراض الآمن عند أي "
+    "تعذّر = 5 (محجوب كلياً) — لا 1 — كي لا يكشف فشلٌ واحد كل الإجابات.",
+)
