+++ app/services/skills/doctrine/probability.py
@@ -1,0 +1,111 @@
+"""Probability — sliced verbatim from the former doctrine.py (D-173 Stage 2a).
+
+جزء من حزمة `app.services.skills.doctrine` — المصدر الوحيد للحقيقة لقواعد كل
+Skill. لا تُعرَّف doctrines خارج هذه الحزمة (بوّابة skills-doctrine تحرس ذلك).
+"""
+
+from __future__ import annotations
+
+from typing import Final
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Probability Calculation Doctrine (D-075 — Protocol V14.0 / Generative UI)
+# ─────────────────────────────────────────────────────────────────────────────
+
+PROBABILITY_CALCULATION_DOCTRINE_VERSION: Final[str] = "1.5.0"
+
+#: قوانين غير قابلة للكسر تحكم محرّك حساب الاحتمالات الحتمي الذي يغذّي شجرة
+#: الاحتمالات التوليدية. الخلفية لا تُخرج HTML ولا قيماً وهمية — تحسب كسوراً
+#: حقيقية من تركيبة المسألة العربية (Protocol V14.0 §3).
+PROBABILITY_CALCULATION_DOCTRINE: Final[tuple[str, ...]] = (
+    "حساب حقيقي لا قيم وهمية: P(الحدث) = العدد / المجموع كـ كسر مختزَل دقيق "
+    "(مثل 4/11) — ممنوع منعاً باتاً إخراج 0.5 افتراضية حين تتوفّر التركيبة.",
+    "حتمي تماماً: نفس نص المسألة يُنتج نفس الكسور دائماً — لا LLM، لا عشوائية، "
+    "قابل للاختبار بـ pytest عادي.",
+    "استخراج التركيبة من العربية: أعداد رقمية وكلمات (كرتان=2، أربع=4، خمس=5) "
+    "مقرونة بالكيانات الملموسة (حمراء، بيضاء، خضراء) — لا رموز مجرّدة.",
+    "السحب بدون إرجاع يُحدِّث المقام والبسط شرطياً ((العدد-1)/(المجموع-1))؛ "
+    "السحب مع الإرجاع يُبقي الاحتمال ثابتاً — يُكتشف الوضع من نص المسألة.",
+    "كل عقدة شجرة تحمل الكسر الدقيق (p_num/p_den) إضافةً للقيمة العشرية، "
+    "فتُصيّر الواجهة 4/11 تماماً دون إعادة بناء تقريبية من العشري.",
+    "Abstraction Ban: تسميات العُقَد ملموسة بشرية مستخرَجة من السياق — وحتى "
+    "fallback المجموع المجهول لا يُخرج حرفاً مجرّداً.",
+    "العقد فصل صارم: المحرّك يُرجع Pydantic منظَّماً فقط؛ التصيير مسؤولية "
+    "الواجهة (RSC) — الخلفية لا تعرف شيئاً عن SVG أو HTML.",
+    "الموجِّه التربوي (V19.0): «دفعة واحدة» (سحب آني) = تأليفي C(n,k) → مكوّن "
+    "combinations_visualizer؛ «على التوالي» (تتابعي) → probability_tree. ممنوع "
+    "فرض شجرة تتابعية على مسألة آنية (كارثة تربوية).",
+    "جذر الشجرة بلا احتمال (لا 1/1 وهمي)، ولا كسر منحلّ (1/0) يصل للطالب أبداً.",
+    "استخراج الكيانات الملموسة (ISS-083): يقبل اسم الرقم الصريح المستقل فقط "
+    "(«بطاقة رقم 1»، «تحمل الرقم 2») ويرفض صفة «مرقمة/مرقمتان/مرقم» (تعني "
+    "«مُعلَّمة بـ» وتصف ترقيم العناصر لا نوعها). مطابقة «رقم» كسلسلة فرعية داخل "
+    "«مرقمة» كانت تُنتج كياناً زائفاً («كرة رقم 0») يُضخّم المجموع (17 بدل 11) "
+    "ويُخرج كسوراً خاطئة وشجرة تتابعية مضلِّلة لمسألة سحب آني — كارثة تربوية.",
+    "القصة التربوية الشاملة (V31.5 — Full Exercise OS): حين يعبّر الطالب عن "
+    "حيرة كاملة في مسألة سحب آني، تُولَّد سلسلة خطوات بصرية مستقلّة (Carousel) "
+    "تغطّي التمرين كاملاً — معطيات → فضاء عيّنة → حدث مركّب → متغيّر عشوائي "
+    "(توزيع فوق-هندسي بـ math.comb) — بدل مكوّن تأليفات واحد. كل خطوة تفصل "
+    "visual_directives عن numerical_state عن pedagogical_message فصلاً صارماً، "
+    "والمجموعة المستحيلة (count<k) تحمل is_possible=False ولا يتسرّب منها C_n^k=0 "
+    "ولا كسر صفري للطالب. الخطوات معمّمة من التركيبة لا مفصّلة لمسألة بعينها.",
+    "نية السؤال الحالي إلزامية (ISS-110): سياق الاحتمالات القادم من history "
+    "المحادثة وحده لا يكفي لتفعيل الأداة البصرية — السؤال الحالي نفسه يجب أن "
+    "يحمل نية احتمالية (كلمة سياق صريحة، أو إشارة حيرة، أو متابعة خطوة مثل "
+    "E(X)/جداء/نفس اللون). طلب موضوع آخر صريح («اعطني تمرين الدوال العددية») "
+    "يحجب الأداة تماماً، والاسترجاع المُفهرَس يسبق دائماً الواجهة المحسوبة في "
+    "chat_with_agent — مكوّن احتمالات مبني من كيس تمرين سابق رداً على طلب "
+    "موضوع جديد = كارثة توجيه حية.",
+    "كلمات فضاء العيّنة (نرد/عملة) تُؤخذ من سؤال الطالب ورسائله فقط — نص "
+    "المساعد في الـ history ليس دليلاً (ISS-114 · مرآة D-102): مثال توضيحي "
+    "بالنرد في إجابة سابقة كان يطغى على تركيبة الكيس الحقيقية. والتركيبة "
+    "المستخرَجة من نص التمرين تسبق فضاء العيّنة العام في خط الاستراتيجيات.",
+)
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Impossible Case UX Doctrine (D-082)
+# ─────────────────────────────────────────────────────────────────────────────
+# قواعد تجربة الحالة المستحيلة — تُطبَّق عند requested > bag_count.
+# المُستفيدون: content-retrieval-skill /impossible-case + frontend renderer.
+# ─────────────────────────────────────────────────────────────────────────────
+
+#: نسخة doctrine الحالة المستحيلة.
+IMPOSSIBLE_CASE_DOCTRINE_VERSION: Final[str] = "1.0.0"
+
+IMPOSSIBLE_CASE_DOCTRINE: Final[tuple[str, ...]] = (
+    "لا تبدأ بالرياضيات: المرحلة 1 و2 و3 خالية تماماً من المعادلات والصيغ.",
+    "مشهد واحد في كل مرحلة: لا تعرض أكثر من فكرة واحدة في أي شاشة.",
+    "لا شجرة احتمالات في الحالة المستحيلة: الشجرة تظهر فقط للحالات الممكنة.",
+    "السؤال الحدسي (المرحلة 2) يجب أن يكون مباشراً وبسيطاً — لا معادلة، لا صفر.",
+    "الحركة البصرية (المرحلة 3) قصيرة وهادئة — shake ≤ 600ms، لا تكرار.",
+    "الرسالة التربوية (المرحلة 4) واحدة فقط — واضحة ولطيفة، تذكر الأعداد الفعلية.",
+    "deep_dive اختياري: يُعرض فقط عند ضغط المستخدم على 'اشرح لي أكثر'.",
+    "backend يُرسل contract بصري فقط (حالة + بيانات + رسالة) — لا HTML، لا SVG.",
+    "frontend مسؤول عن تحويل الـ contract إلى قصة بصرية — لا خلط منطق بحركة.",
+    "الشرح الرياضي العميق (deep_dive) يحتوي: formula + explanation + why_zero.",
+)
+
+
+def get_impossible_case_summary() -> str:
+    """يُرجِع ملخصاً موجزاً لـ doctrine الحالة المستحيلة."""
+    return (
+        f"[v{IMPOSSIBLE_CASE_DOCTRINE_VERSION}] "
+        f"{len(IMPOSSIBLE_CASE_DOCTRINE)} قاعدة — "
+        "4 مراحل بصرية، لا رياضيات قبل الفهم الحدسي، deep_dive اختياري."
+    )
+
+
+# ─────────────────────────────────────────────────────────────────────────────
+# Realtime Protocol Doctrine — D-WS-FLAP-002 / ISS-WS-FLAP-002 (2026-05-26)
+# ─────────────────────────────────────────────────────────────────────────────
+# هذه القواعد تحكم سلوك الـ WebSocket message loop عبر كل WS endpoints في النظام.
+# الكارثة المُشخَّصة: الواجهة ترسل `{type:"ping"}` كل 25 ثانية كـ app-level
+# heartbeat، لكن الـ WS routers (customer_chat, admin, conversation_service)
+# كانت تعالج كل رسالة واردة كـ «سؤال»، تُرجع خطأ «Question is required»،
+# ولا تُرسل `{type:"pong"}` → timeout بعد 10s → close(1001) → reconnect → flap.
+#
+# المُستفيدون:
+# - `app.services.skills.ws_heartbeat_skill.handle_control_message`
+# - `app.api.routers.customer_chat.chat_stream_ws` (WS loop)
+# - `app.api.routers.admin.admin_chat_stream_ws` (WS loop)
+# - `microservices.conversation_service.main` (WS endpoints)
